### 2025-01-17 #7684 @editorialbot
[PRE REVIEW]: Sunny.jl: A Julia Package for Spin Dynamics
**Submitting author:** @ddahlbom (<a href="http://orcid.org/0000-0002-0221-5086">David Dahlbom</a>)
**Repository:** https://github.com/SunnySuite/Sunny.jl
**Branch with paper.md** (empty if default branch): joss-manuscript
**Version:** v0.7.4
