### 2025-01-17 bbb3bac
Fix color parsing
