+++ Project.toml
@@ -37,15 +37,15 @@ WGLMakiePrecompilesExt = ["Makie", "WGLMakie"]
 
 [compat]
 Brillouin = "0.5.19"
-CrystalInfoFramework = "0.7.0"
+CrystalInfoFramework = "0.8"
 DynamicPolynomials = "0.6.0"
 ElasticArrays = "1.2.12"
 FFTW = "1.4.5"
-GLMakie = "0.9, 0.10"
+GLMakie = "0.9, 0.10, 0.11"
 HCubature = "1.7.0"
 JLD2 = "0.5.4"
 LinearAlgebra = "1.10"
-Makie = "0.20, 0.21.1"
+Makie = "0.20, 0.21.1, 0.22"
 OffsetArrays = "1.10.8"
 Optim = "1.7.6"
 PrecompileTools = "1.2.0"
@@ -56,6 +56,6 @@ SpecialFunctions = "2.1.0"
 Spglib = "0.9.4"
 StaticArrays = "1.9.1"
 Statistics = "1.10"
-WGLMakie = "0.9, 0.10"
+WGLMakie = "0.9, 0.10, 0.11"
 WriteVTK = "1.18.0"
 julia = "1.10"
