### 2025-01-20 908d106
Add CompatHelper action
