### 2025-01-21 b4dd8e7
Lanczos implementation
