+++ src/KPM/Lanczos.jl
@@ -1,13 +1,40 @@
-# Reference implementation, without consideration of speed.
+# Reference implementation of the generalized Lanczos method, without
+# consideration of speed. The input A can be any Hermitian matrix. The optional
+# argument S is an inner-product that must be both Hermitian and positive
+# definite. Intuitively, Lanczos finds a low-rank approximant A S ≈ Q T Q† S,
+# where T is tridiagonal and Q is orthogonal in the sense that Q† S Q = I. The
+# first column of Q is the initial vector v, which must be normalized. The
+# Lanczos method is useful in two ways. First, eigenvalues of T are
+# representative of extremal eigenvalues of A (in an appropriate S-measure
+# sense). Second, one obtains a very good approximation to the matrix-vector
+# product, f(A S) v ≈ Q f(T) e₁, valid for any function f [1].
+# 
+# The generalization of Lanczos to an arbitrary measure S is implemented as
+# follows: Each matrix-vector product A v is replaced by A S v, and each vector
+# inner product w† v is replaced by w† S v. Similar generalizations of Lanczos
+# have been considered in [2] and [3].
+#
+# 1. N. Amsel, T. Chen, A. Greenbaum, C. Musco, C. Musco, Near-optimal
+#    approximation of matrix functions by the Lanczos method, (2013)
+#    https://arxiv.org/abs/2303.03358v2.
+# 2. H. A. van der Vorst, Math. Comp. 39, 559-561 (1982),
+#    https://doi.org/10.1090/s0025-5718-1982-0669648-0
+# 3. M. Grüning, A. Marini, X. Gonze, Comput. Mater. Sci. 50, 2148-2156 (2011),
+#    https://doi.org/10.1016/j.commatsci.2011.02.021.
 function lanczos_ref(A, S, v; niters)
     αs = Float64[]
     βs = Float64[]
+    vs = typeof(v)[]
 
-    v /= sqrt(real(dot(v, S, v)))
+    c = sqrt(real(dot(v, S, v)))
+    @assert c ≈ 1 "Initial vector not normalized"
+    v /= c
+
     w = A * S * v
     α = real(dot(w, S, v))
     w = w - α * v
     push!(αs, α)
+    push!(vs, v)
 
     for _ in 2:niters
         β = sqrt(real(dot(w, S, w)))
@@ -19,35 +46,54 @@ function lanczos_ref(A, S, v; niters)
         v = vp
         push!(αs, α)
         push!(βs, β)
+        push!(vs, v)
     end
 
-    return SymTridiagonal(αs, βs)
+    Q = reduce(hcat, vs)
+    T = SymTridiagonal(αs, βs)
+    return (; Q, T)
 end
 
 
-
-# Use Lanczos iterations to find a truncated tridiagonal representation of A S,
-# up to similarity transformation. Here, A is any Hermitian matrix, while S is
-# both Hermitian and positive definite. Traditional Lanczos uses the identity
-# matrix, S = I. The extension to non-identity matrices S is as follows: Each
-# matrix-vector product A v becomes A S v, and each vector inner product w† v
-# becomes w† S v. The implementation below follows the notation of Wikipedia,
-# and is the most stable of the four variants considered by Paige [1]. Each
-# Lanczos iteration requires only one matrix-vector multiplication for A and S,
-# respectively.
-
-# Similar generalizations of Lanczos have been considered in [2] and [3].
+# An optimized implementation of the generalized Lanczos method. See
+# `lanczos_ref` for explanation, and a reference implementation. This optimized
+# implementation follows "the most stable" of the four variants considered by
[…]
