+++ dev/assets/scripts/SW11_La2CuO4.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.5"
+@assert pkgversion(Sunny) >= v"0.7.6"
 
 units = Units(:meV, :angstrom)
 latvecs = lattice_vectors(1, 1, 10, 90, 90, 90)
