+++ dev/assets/scripts/SW15_Ba3NbFe3Si2O14.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.5"
+@assert pkgversion(Sunny) >= v"0.7.6"
 
 units = Units(:meV, :angstrom)
 a = b = 8.539 # (Å)
