+++ dev/assets/scripts/SW04_Frustrated_square.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.5"
+@assert pkgversion(Sunny) >= v"0.7.6"
 
 units = Units(:meV, :angstrom)
 latvecs = lattice_vectors(3.0, 3.0, 6.0, 90, 90, 90)
