### 2025-01-25 a4b2b62
Fix `suggest_magnetic_supercell` on multiple k
