### 2025-01-27 #7684 @ddahlbom
> [@ddahlbom](https://github.com/ddahlbom) - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
>
> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

Ok. Thanks for the update!
