+++ README.md
@@ -7,7 +7,7 @@
     </a>
     <br>
     <a href="https://sunnysuite.github.io/Sunny.jl/stable"><img src="https://img.shields.io/badge/docs-stable-blue.svg" alt="Docs-stable"></a>
-    <a href="https://github.com/SunnySuite/Sunny.jl/actions/workflows/CI.yml?query=branch%3Amain"><img src="https://github.com/SunnySuite/Sunny.jl/actions/workflows/CI.yml/badge.svg?branch=main" alt="CI"></a>
+    <a href="https://github.com/SunnySuite/Sunny.jl/actions/workflows/CI.yml?query=branch%3Amain"><img src="https://github.com/SunnySuite/Sunny.jl/actions/workflows/CI.yml/badge.svg?branch=main" alt="CI"></a><a href="https://arxiv.org/abs/2501.13095"><img src="https://img.shields.io/badge/arXiv-2501.1309v2-b31b1b.svg" alt="arXiv"></a>
 </div>
 
 ## Overview
