### 2025-02-03 71ad19c
Attempt simplification
