### 2025-02-03 ed48877
Fix `suggest_magnetic_supercell` with multiple wavevectors (#352)
Related fixes and simplifications are included as well.
