+++ docs/src/versions.md
@@ -5,6 +5,8 @@
 
 * Vacancies defined by [`set_vacancy_at!`](@ref) are supported in linear spin
   wave theory. Empty sites are modeled using bosons that do not excite.
+* Fix correctness of [`suggest_magnetic_supercell`](@ref) when multiple
+  wavevectors are provided.
 
 ## v0.7.5
 (Jan 20, 2025)
