+++ README.md
@@ -1,5 +1,3 @@
-<div align="center">
-
 <a href="https://github.com/SunnySuite/Sunny.jl/">
     <picture>
         <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/SunnySuite/Sunny.jl/main/assets/sunny_logo-dark.svg">
@@ -8,7 +6,7 @@
 </a>
 
 | **Documentation**         | **Build Status**      |  **Citation**           |
-| :-----------------------: | :-------------------: | :---------------------: |
+| ------------------------- | --------------------- | ----------------------- |
 | [![][docs-img]][docs-url] | [![][ci-img]][ci-url] | [![][doi-img]][doi-url] |
 
 [docs-img]: https://img.shields.io/badge/docs-stable-blue.svg
@@ -17,8 +15,6 @@
 [ci-url]: https://github.com/SunnySuite/Sunny.jl/actions/workflows/CI.yml?query=branch%3Amain
 [doi-img]: https://img.shields.io/badge/DOI-10.48550-blue
 [doi-url]: https://doi.org/10.48550/arXiv.2501.13095
-
-</div>
 
 ## Overview
 
