+++ joss/paper.md
@@ -306,7 +306,7 @@ to integrate Sunny into the Calvera platform for neutron data analysis
 analysis techniques, building on its mature model specification and data
 retrieval features. Current efforts are directed at supporting: the
 self-consistent Gaussian approximation for diffuse scattering, enabling
-functionality inspired by [@paddison:2024]; the modeling of local entanglement
+functionality inspired by [@paddison:2023]; the modeling of local entanglement
 effects generated by spin-orbit coupling or strongly coupled clusters of spins;
 non-perturbative corrections to LSWT for the modeling of continua and bound
 states, which can be probed in INS and terahertz spectroscopy experiments
