### 2025-02-06 48fb2d3
Lanczos implementation
