### 2025-02-06 c43a29e
Explicit left-justification in table
