### 2025-02-06 a9ccd70
Update README with Sunny paper reference (#353)
Co-authored-by: ddahlbom <david.dahlbom@gmail.com>
