+++ src/KPM/SpinWaveTheoryKPM.jl
@@ -1,30 +1,76 @@
 """
-    SpinWaveTheoryKPM(sys::System; measure, regularization=1e-8, tol)
+    SpinWaveTheoryKPM(sys::System; measure, regularization=1e-8, tol=nothing,
+                      niters=nothing, method=:lanczos)
 
-A variant of [`SpinWaveTheory`](@ref) that uses the kernel polynomial method
-(KPM) to calculate [`intensities`](@ref) [1]. This method avoids direct matrix
-diagonalization, which scales cubically in the system size ``N``. Instead, using
-``M`` iterative matrix-vector multiplications, the cost becomes ``𝒪(N M +
-M^2)``. This provides a very significant acceleration when ``N`` is large.
+A variant of [`SpinWaveTheory`](@ref) that uses matrix-vector products to
+estimate [`intensities`](@ref) [1]. This approach avoids direct matrix
+diagonalization and therefore yields significant acceleration when the system
+size ``N`` is large. Large magnetic supercells may arise when studying models
+with quenched disorder, or for approximation of incommensurate ordering
+wavevectors.
 
-The number of iterations is typically not too large: `M ≈ -2 log10(tol) Δϵ /
+!!! warning "Accuracy considerations"  
+    The available energy-space resolution scales inversely with the dimension
+    ``M`` of the Krylov subspace. In practice, energy broadening artifacts can
+    be well controlled through appropriate choice of ``M``. A more serious
+    problem is intensity loss at small excitation energies, e.g., in the
+    viscinity of Goldstone modes. This missing intensity may be unrecoverable,
+    even when ``M`` is large.
+
+Computational cost scales like ``𝒪(N M + M^2)``. The number of iterations ``M``
+can be specified in two possible ways: Directly with the `niters` parameter, or
+indirectly through the `tol` parameter. For the latter, `M ≈ -2 log10(tol) Δϵ /
 fwhm` where `Δϵ` is the estimated spectral bandwidth of excitations, `fwhm` is
-the full width at half maximum of the broadening kernel, and `tol` is a
-dimensionless tolerance parameter. Good choices are `0.05` (more speed) or
-`0.01` (more accuracy).
+the full width at half maximum of the user-supplied broadening `kernel`. Good
+choices for the dimensionless `tol` parameter may be `0.05` (more speed) or
+`0.01` (more accuracy). Exactly one of `tol` or `niters` must be provided.
+
+The current default implementation uses the Lanczos method [1], which achieves
+near-optimal accuracy for a given Krylov subspace [2]. Sunny's original
+implementation used Kernel Polynomial Method [3] which is less accurate than
+Lanczos. Select `method=:kpm` to test this historical method.
 
 ## References
 
-1. [H. Lane et al., _Kernel Polynomial Method for Linear Spin Wave Theory_
+1. [T. Chen, _The Lanczos algorithm for matrix functions: a handbook for
+   scientists_ (2024) [arXiv:2410.11090]](https://arxiv.org/abs/2410.11090).
+2. [N. Amsel, T. Chen, A. Greenbaum, C. Musco, C. Musco, _Near-Optimal
+   Approximation of Matrix Functions by the Lanczos Method_ (2023)
+   [arXiv:2303.03358]](https://arxiv.org/abs/2303.03358).
+3. [H. Lane et al., _Kernel Polynomial Method for Linear Spin Wave Theory_
    (2023) [arXiv:2312.08349]](https://arxiv.org/abs/2312.08349).
 """
[…]
