### 2025-02-07 3182b8a
More robust niters_bounds setting.
