+++ src/KPM/SpinWaveTheoryKPM.jl
@@ -1,34 +1,35 @@
 """
     SpinWaveTheoryKPM(sys::System; measure, regularization=1e-8, tol=nothing,
-                      niters=nothing, method=:lanczos)
+                      niters=nothing, niters_bounds=10, method=:lanczos)
 
-A variant of [`SpinWaveTheory`](@ref) that uses matrix-vector products to
-estimate [`intensities`](@ref) [1]. This approach avoids direct matrix
-diagonalization and therefore yields significant acceleration when the system
-size ``N`` is large. Large magnetic supercells may arise when studying models
-with quenched disorder, or for approximation of incommensurate ordering
-wavevectors.
+A variant of [`SpinWaveTheory`](@ref) that estimates [`intensities`](@ref) using
+only iterated matrix-vector products. By avoiding direct matrix diagonalization,
+this method reduces computational cost from cubic to linear-scaling in the
+system size ``N``. Iterative methods can be especially useful for models of
+quenched disorder or models with nearly incommensurate ordering wavevectors.
 
 !!! warning "Accuracy considerations"  
-    The available energy-space resolution scales inversely with the dimension
-    ``M`` of the Krylov subspace. In practice, energy broadening artifacts can
-    be well controlled through appropriate choice of ``M``. A more serious
-    problem is intensity loss at small excitation energies, e.g., in the
-    viscinity of Goldstone modes. This missing intensity may be unrecoverable,
-    even when ``M`` is large.
+    Energy-space resolution scales inversely with the number ``M`` of
+    iterations. In practice, numerical broadening effects can be well controlled
+    via the tolerance parameter `tol`. A more serious problem is intensity loss
+    at bands with small excitation energy, e.g., in the viscinity of Goldstone
+    modes. This missing intensity may be due to floating point round-off errors,
+    whcih are not fixable by increasing ``M``.
 
 Computational cost scales like ``𝒪(N M + M^2)``. The number of iterations ``M``
-can be specified in two possible ways: Directly with the `niters` parameter, or
-indirectly through the `tol` parameter. For the latter, `M ≈ -2 log10(tol) Δϵ /
-fwhm` where `Δϵ` is the estimated spectral bandwidth of excitations, `fwhm` is
-the full width at half maximum of the user-supplied broadening `kernel`. Good
+can be specified directly with the `niters` parameter. More commonly an error
+tolerance `tol` will be specified; in this case, `M ≈ -2 log10(tol) Δϵ / fwhm`
+where `Δϵ` is the estimated spectral bandwidth of excitations and `fwhm` is the
+full width at half maximum of the user-supplied broadening `kernel`. Good
 choices for the dimensionless `tol` parameter may be `0.05` (more speed) or
-`0.01` (more accuracy). Exactly one of `tol` or `niters` must be provided.
+`0.01` (more accuracy). Exactly one of `tol` or `niters` must be provided. The
+parameter `niters_bounds` selects the Krylov subspace dimension to be used for
+estimating spectral bounds.
 
-The current default implementation uses the Lanczos method [1], which achieves
-near-optimal accuracy for a given Krylov subspace [2]. Sunny's original
-implementation used Kernel Polynomial Method [3] which is less accurate than
-Lanczos. Select `method=:kpm` to test this historical method.
+The default Krylov sub-space method is Lanczos [1], which achieves near-optimal
[…]
