+++ src/KPM/SpinWaveTheoryKPM.jl
@@ -3,20 +3,19 @@
                       niters=nothing, niters_bounds=10, method=:lanczos)
 
 A variant of [`SpinWaveTheory`](@ref) that estimates [`intensities`](@ref) using
-only iterated matrix-vector products. By avoiding direct matrix diagonalization,
-this method reduces computational cost from cubic to linear-scaling in the
-system size ``N``. Iterative methods can be especially useful for models of
-quenched disorder or models with nearly incommensurate ordering wavevectors.
+iterated matrix-vector products. By avoiding direct matrix diagonalization, this
+method reduces computational cost from cubic to linear-scaling in the system
+size ``N``. Large system sizes can arise, e.g., for models of quenched disorder
+or models with nearly incommensurate ordering wavevectors.
 
 Computational cost scales like ``𝒪(N M + M^2)``. The number of iterations ``M``
-can be specified directly with the `niters` parameter. More commonly an error
-tolerance `tol` will be specified; in this case, `M ≈ -2 log10(tol) Δϵ / fwhm`
-where `Δϵ` is the estimated spectral bandwidth of excitations and `fwhm` is the
-full width at half maximum of the user-supplied broadening `kernel`. Common
-choices for the dimensionless `tol` parameter are `0.05` (more speed) or `0.01`
-(more accuracy). Exactly one of `tol` or `niters` must be provided. The
-parameter `niters_bounds` selects the Krylov subspace dimension to be used for
-estimating spectral bounds.
+may be specified directly with the `niters` parameter. Alternatively, one may
+specify a dimensionless error tolerance `tol` such that `M ≈ -2 log10(tol) Δϵ /
+fwhm` where `Δϵ` is the estimated spectral bandwidth of excitations and `fwhm`
+is the full width at half maximum of the user-supplied broadening `kernel`.
+Common choices for `tol` are `0.05` (more speed) or `0.01` (more accuracy).
+Either `tol` or `niters` is required. The parameter `niters_bounds` selects the
+Krylov subspace dimension to be used for estimating spectral bounds.
 
 !!! warning "Accuracy considerations"  
     Energy-space resolution scales inversely with the number ``M`` of
@@ -24,12 +23,12 @@ estimating spectral bounds.
     tolerance parameter `tol`. A more serious problem is intensity loss at bands
     with small excitation energy, e.g., in the viscinity of Goldstone modes. In
     certain cases this missing intensity will be due to numerical roundoff error
-    and cannot be fixed by increasing the number of iterations ``M``.
+    that cannot be fixed by increasing the number of iterations ``M``.
 
-The default Krylov subspace method is Lanczos [1], which achieves near-optimal
-accuracy for a given number of iterations [2]. Alternatively, `method=:kpm` will
-select the historical Kernel Polynomial Method [3], which is expected to be less
-accurate.
+Two `method` options are available, `:lanczos` and `:kpm`.  Lanczos is generally
+preferred, as it achieves near-optimal accuracy for a given number of iterations
+[1, 2]. The Lanczos implementation builds on earlier research using the Kernel
+Polynomial Method [3], which may be of historical interest.
 
 ## References
 
