+++ docs/src/structure-factor.md
@@ -374,12 +374,12 @@ calculated to leading order in inverse powers of the quantum spin-$s$. For
 systems constructed with `mode = :SUN`, Sunny automatically switches to a
 multi-flavor boson variant of spin wave theory, which captures more single-ion
 physics. Use [`SpinWaveTheorySpiral`](@ref) to study generalized spiral phases,
-which allow for an incommensurate propagation wavevector. The experimental
-module [`SpinWaveTheoryKPM`](@ref) implements [spin wave calculations using the
-kernel polynomial method](https://arxiv.org/abs/2312.0834). In the KPM approach,
-the computational cost scales linearly in the magnetic cell size. It can be
-useful for studying systems with large magnetic cells include systems with
-long-wavelength structures, or systems with quenched chemical disorder.
+which allow for an incommensurate propagation wavevector. The module
+[`SpinWaveTheoryKPM`](@ref) uses [iterated matrix-vector
+multiplication](https://arxiv.org/abs/2312.0834) to achieve linear-scaling cost
+in the magnetic cell size. It can be useful for studying systems with large
+magnetic cells include systems with long-wavelength structures, or systems with
+quenched chemical disorder.
 
 
 ## Calculations with classical spin dynamics
