+++ examples/09_Disorder_KPM.jl
@@ -1,14 +1,14 @@
-# # 9. Disordered system with KPM
+# # 9. Disordered systems
 #
-# This example uses the [kernel polynomial method](@ref SpinWaveTheoryKPM) to
-# efficiently calculate the neutron scattering spectrum of a disordered
-# triangular antiferromagnet. The model is inspired by YbMgGaO4, as studied in
-# [Paddison et al, Nature Phys., **13**, 117–122
-# (2017)](https://doi.org/10.1038/nphys3971) and [Zhu et al, Phys. Rev. Lett.
-# **119**, 157201 (2017)](https://doi.org/10.1103/PhysRevLett.119.157201).
-# Disordered occupancy of non-magnetic Mg/Ga sites can be modeled as a
-# stochastic distribution of exchange constants and ``g``-factors. Including
-# this disorder introduces broadening of the spin wave spectrum.
+# This example uses [`SpinWaveTheoryKPM`](@ref) to efficiently calculate the
+# neutron scattering spectrum of a disordered triangular antiferromagnet. The
+# model is inspired by YbMgGaO4, as studied in [Paddison et al, Nature Phys.,
+# **13**, 117–122 (2017)](https://doi.org/10.1038/nphys3971) and [Zhu et al,
+# Phys. Rev. Lett. **119**, 157201
+# (2017)](https://doi.org/10.1103/PhysRevLett.119.157201). Disordered occupancy
+# of non-magnetic Mg/Ga sites can be modeled as a stochastic distribution of
+# exchange constants and ``g``-factors. Including this disorder introduces
+# broadening of the spin wave spectrum.
 
 using Sunny, GLMakie
 
@@ -47,12 +47,13 @@ plot_intensities(res)
 
 sys_inhom = to_inhomogeneous(repeat_periodically(sys, (10, 10, 1)))
 
-# Use [`symmetry_equivalent_bonds`](@ref) to iterate over all nearest neighbor
-# bonds of the inhomogeneous system. Modify each AFM exchange with a noise term
-# that has variance of 1/3. The newly minimized energy configuration allows for
-# long wavelength modulations on top of the original 120° order.
+# Iterate over all [`symmetry_equivalent_bonds`](@ref) for nearest neighbor
+# sites in the inhomogeneous system. Set the AFM exchange to include a noise
+# term that has a standard deviation of 1/3. The newly minimized energy
+# configuration allows for long wavelength modulations on top of the original
+# 120° order.
 
-for (site1, site2, offset) in symmetry_equivalent_bonds(sys_inhom, Bond(1,1,[1,0,0]))
+for (site1, site2, offset) in symmetry_equivalent_bonds(sys_inhom, Bond(1, 1, [1, 0, 0]))
     noise = randn()/3
     set_exchange_at!(sys_inhom, 1.0 + noise, site1, site2; offset)
 end
@@ -60,38 +61,31 @@ end
 minimize_energy!(sys_inhom, maxiters=5_000)
 plot_spins(sys_inhom; color=[S[3] for S in sys_inhom.dipoles], ndims=2)
 
-# Traditional spin wave theory calculations become impractical for large system
-# sizes. Significant acceleration is possible with the [kernel polynomial
-# method](https://arxiv.org/abs/2312.08349). Enable it by selecting
-# [`SpinWaveTheoryKPM`](@ref) in place of the traditional
-# [`SpinWaveTheory`](@ref). Using KPM, the cost of an [`intensities`](@ref)
-# calculation becomes linear in system size and scales inversely with the width
-# of the line broadening `kernel`. Error tolerance is controlled through the
[…]
