### 2025-02-17 00a1e31
Replace KPM with Lanczos (#339)
In all cases tested, Lanczos appears to be more accurate at a fixed number of iterations, with significant reductions in certain types of numerical artifacts. The cost per iteration is currently slower, but this is outweighed by the reduced number of iterations required.
