+++ src/KPM/SpinWaveTheoryKPM.jl
@@ -1,38 +1,83 @@
 """
-    SpinWaveTheoryKPM(sys::System; measure, regularization=1e-8, tol)
+    SpinWaveTheoryKPM(sys::System; measure, regularization=1e-8, tol=nothing,
+                      niters=nothing, niters_bounds=10, method=:lanczos)
 
-A variant of [`SpinWaveTheory`](@ref) that uses the kernel polynomial method
-(KPM) to perform [`intensities`](@ref) calculations [1]. Instead of explicitly
-diagonalizing the dynamical matrix, KPM approximates intensities using
-polynomial expansion truncated at order ``M``. The reduces the computational
-cost from ``𝒪(N^3)`` to ``𝒪(N M)``, which is favorable for large system sizes
-``N``.
+A variant of [`SpinWaveTheory`](@ref) that estimates [`intensities`](@ref) using
+iterated matrix-vector products. By avoiding direct matrix diagonalization, this
+method reduces computational cost from cubic to linear-scaling in the system
+size ``N``. Large system sizes can arise, e.g., for models of quenched disorder
+or models with nearly incommensurate ordering wavevectors.
 
-The polynomial order ``M`` will be determined from the line broadening kernel
-and the specified error tolerance `tol`. Specifically, for each wavevector,
-``M`` scales like the spectral bandwidth of excitations, divided by the energy
-resolution of the broadening kernel, times the negative logarithm of `tol`.
+Computational cost scales like ``𝒪(N M + M^2)``. The number of iterations ``M``
+may be specified directly with the `niters` parameter. Alternatively, one may
+specify a dimensionless error tolerance `tol` such that `M ≈ -2 log10(tol) Δϵ /
+fwhm` where `Δϵ` is the estimated spectral bandwidth of excitations and `fwhm`
+is the full width at half maximum of the user-supplied broadening `kernel`.
+Common choices for `tol` are `0.05` (more speed) or `0.01` (more accuracy).
+Either `tol` or `niters` is required. The parameter `niters_bounds` selects the
+Krylov subspace dimension to be used for estimating spectral bounds.
 
-The error tolerance `tol` should be tuned empirically for each calculation.
-Reasonable starting points are `1e-1` (more speed) or `1e-2` (more accuracy).
+!!! warning "Accuracy considerations"  
+    Energy-space resolution scales inversely with the number ``M`` of
+    iterations. Such broadening errors can usually be well controlled via the
+    tolerance parameter `tol`. A more serious problem is intensity loss at bands
+    with small excitation energy, e.g., in the viscinity of Goldstone modes. In
+    certain cases this missing intensity will be due to numerical roundoff error
+    that cannot be fixed by increasing the number of iterations ``M``.
 
-!!! warning "Missing intensity at small quasi-particle energy"  
-    The KPM-calculated intensities are unreliable at small energies ``ω``. In
-    particular, KPM may mask intensities that arise near the Goldstone modes of
-    an ordered state with continuous symmetry.
+Two `method` options are available, `:lanczos` and `:kpm`.  Lanczos is generally
+preferred, as it achieves near-optimal accuracy for a given number of iterations
+[1, 2]. The Lanczos implementation builds on earlier research using the Kernel
+Polynomial Method [3], which may be of historical interest.
 
 ## References
 
[…]
