+++ src/SpinWaveTheory/HamiltonianDipole.jl
@@ -4,7 +4,7 @@ function swt_hamiltonian_dipole!(H::Matrix{ComplexF64}, swt::SpinWaveTheory, q_r
     (; local_rotations, stevens_coefs, sqrtS) = data
     (; extfield, gs) = sys
 
-    L = nbands(swt) 
+    L = nbands(swt)
     @assert size(H) == (2L, 2L)
 
     # Initialize Hamiltonian buffer 
