+++ src/Spiral/SpinWaveTheorySpiral2.jl
@@ -1,0 +1,353 @@
+"""
+    SpinWaveTheorySpiral(sys::System; k, axis, measure, regularization=1e-8)
+
+Analogous to [`SpinWaveTheory`](@ref), but interprets the provided system as
+having a generalized spiral order. This order is described by a single
+propagation wavevector `k`, which may be incommensurate. The `axis` vector
+defines the polarization plane via its surface normal. Typically the spin
+configuration in `sys` and the propagation wavevector `k` will be optimized
+using [`minimize_energy_spiral!`](@ref). In contrast, `axis` will typically be
+determined from symmetry considerations.
+
+The resulting object can be used to calculate the spin wave
+[`dispersion`](@ref), or the structure factor via [`intensities_bands`](@ref)
+and [`intensities`](@ref).
+
+The algorithm for this calculation was developed in [Toth and Lake, J. Phys.:
+Condens. Matter **27**, 166002 (2015)](https://arxiv.org/abs/1402.6069) and
+implemented in the [SpinW code](https://spinw.org/).
+"""
+struct SpinWaveTheorySpiral <: Sunny.AbstractSpinWaveTheory
+    swt :: SpinWaveTheory
+    k :: Sunny.Vec3
+    axis :: Sunny.Vec3
+
+    function SpinWaveTheorySpiral(sys::System; k::AbstractVector, axis::AbstractVector, measure::Union{Nothing, Sunny.MeasureSpec}, regularization=1e-8)
+        return new(SpinWaveTheory(sys; measure, regularization), k, axis)
+    end
+end
+
+function construct_uniaxial_anisotropy(; axis, c20=0., c40=0., c60=0., S)
+    # Anisotropy operator in local frame
+    O = stevens_matrices(S)
+    op = c20*O[2, 0] + c40*O[4, 0] + c60*O[6, 0]
+    # Rotate operator into global frame, defined by axis
+    R = rotation_between_vectors(axis, [0, 0, 1])
+    return rotate_operator(op, R)
+end
+
+function fourier_bilinear_interaction(swt::SpinWaveTheory,q)
+    (;sys, data) = swt
+    (; local_rotations, stevens_coefs, sqrtS) = data
+    (; extfield, gs) = sys
+    L = Sunny.nbands(swt)
+    @assert sys.dims == (1, 1, 1) "System must have only a single cell"
+    Rs = local_rotations
+    Na = Sunny.natoms(sys.crystal)
+    J_k = zeros(ComplexF64, 3, Na, 3, Na)
+
+    for i in 1:Sunny.natoms(sys.crystal)
+        for coupling in sys.interactions_union[i].pair
+            (; isculled, bond, bilin) = coupling
+            isculled && break
+
+            (; j, n) = bond
+            J_undo = Rs[i] * Sunny.Mat3(bilin*I) * Rs[j]' # Undo the transformation for exchange Matrix
+            J = exp(-2π * im * dot(q, n)) * J_undo
+            J_k[:, i, :, j] += J / 2
+            J_k[:, j, :, i] += J' / 2
+        end
+    end
+
+    if !isnothing(sys.ewald)
+        A = Sunny.precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), -q) * sys.ewald.μ0_μB²
[…]
