### 2025-02-17 8f90e59
Fix k=0 case
