+++ test/test_chirality.jl
@@ -54,7 +54,7 @@ end
     swt = SpinWaveTheorySpiral(sys; measure=ssf_trace(sys; apply_g=false), k=[0,0,0], axis=[0,0,1])
     res = intensities_bands(swt, qs)
     @test res.disp[1, :] ≈ res.disp[2, :] ≈ res.disp[3, :] ≈ [B + 2D*sin(2π*q[3]) for q in qs]
-    @test_broken res.data ≈ [1 1; 0 0; 0 0]
+    @test res.data ≈ [1 1; 1 1; 1 1] / 3
 
     # Below the saturation field, the ground state is a canted spiral
 
