### 2025-02-17 2be3ee8
Remove unused g extraction
