+++ src/Spiral/SpinWaveTheorySpiral2.jl
@@ -17,12 +17,12 @@ The algorithm for this calculation was developed in [Toth and Lake, J. Phys.:
 Condens. Matter **27**, 166002 (2015)](https://arxiv.org/abs/1402.6069) and
 implemented in the [SpinW code](https://spinw.org/).
 """
-struct SpinWaveTheorySpiral <: Sunny.AbstractSpinWaveTheory
+struct SpinWaveTheorySpiral <: AbstractSpinWaveTheory
     swt :: SpinWaveTheory
-    k :: Sunny.Vec3
-    axis :: Sunny.Vec3
+    k :: Vec3
+    axis :: Vec3
 
-    function SpinWaveTheorySpiral(sys::System; k::AbstractVector, axis::AbstractVector, measure::Union{Nothing, Sunny.MeasureSpec}, regularization=1e-8)
+    function SpinWaveTheorySpiral(sys::System; k::AbstractVector, axis::AbstractVector, measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
         return new(SpinWaveTheory(sys; measure, regularization), k, axis)
     end
 end
@@ -36,23 +36,23 @@ function construct_uniaxial_anisotropy(; axis, c20=0., c40=0., c60=0., S)
     return rotate_operator(op, R)
 end
 
-function fourier_bilinear_interaction(swt::SpinWaveTheory,q)
+function fourier_bilinear_interaction(swt::SpinWaveTheory, q)
     (;sys, data) = swt
     (; local_rotations, stevens_coefs, sqrtS) = data
     (; extfield, gs) = sys
-    L = Sunny.nbands(swt)
+    L = nbands(swt)
     @assert sys.dims == (1, 1, 1) "System must have only a single cell"
     Rs = local_rotations
-    Na = Sunny.natoms(sys.crystal)
+    Na = natoms(sys.crystal)
     J_k = zeros(ComplexF64, 3, Na, 3, Na)
 
-    for i in 1:Sunny.natoms(sys.crystal)
+    for i in 1:natoms(sys.crystal)
         for coupling in sys.interactions_union[i].pair
             (; isculled, bond, bilin) = coupling
             isculled && break
 
             (; j, n) = bond
-            J_undo = Rs[i] * Sunny.Mat3(bilin*I) * Rs[j]' # Undo the transformation for exchange Matrix
+            J_undo = Rs[i] * Mat3(bilin*I) * Rs[j]' # Undo the transformation for exchange Matrix
             J = exp(-2π * im * dot(q, n)) * J_undo
             J_k[:, i, :, j] += J / 2
             J_k[:, j, :, i] += J' / 2
@@ -60,7 +60,7 @@ function fourier_bilinear_interaction(swt::SpinWaveTheory,q)
     end
 
     if !isnothing(sys.ewald)
-        A = Sunny.precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), -q) * sys.ewald.μ0_μB²
+        A = precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), -q) * sys.ewald.μ0_μB²
         A = reshape(A, Na, Na)
         for i in 1:Na, j in 1:Na
             J_k[:, i, :, j] += gs[i]' * A[i, j] * gs[j] / 2
@@ -69,24 +69,14 @@ function fourier_bilinear_interaction(swt::SpinWaveTheory,q)
     return J_k            
 end
 
-function invariance_J(Jq,n)
-    nx = [0 -n[3] n[2];n[3] 0 -n[1];-n[2] n[1] 0]
[…]
