### 2025-02-17 583534e
Document cases for k
