+++ src/Spiral/SpinWaveTheorySpiral2.jl
@@ -36,15 +36,38 @@ function construct_uniaxial_anisotropy(; axis, c20=0., c40=0., c60=0., S)
     return rotate_operator(op, R)
 end
 
+const CMat3 = SMatrix{3, 3, ComplexF64, 9}
+
+# Identify the "special cases" for the propagation wavevector k. Case 1 is all
+# integer k components (i.e., k=[0,0,0] up to periodicity), and Case 2 is all
+# half integer k components, apart from Case 1. The fallback, Case 3, is any
+# other k. For local interactions in real-space, there is no singularity in k,
+# and the fallback case can always be used. But if a Fourier-space interaction
+# J(q) is specified, e.g. long-range dipole-dipole, there may be a true
+# mathematical discontinuity between, say, k = [1/2, 0, 0] and [1/2+ϵ, 0, 0] in
+# the limit ϵ → 0.
+function spiral_propagation_case(k)
+    # The choice of ϵ is a bit ambiguous. Would a user consider [1/2+ϵ, 0, 0]
+    # morally the same as [1/2, 0, 0]? Optimization of k usually gives at least
+    # 8 digits.
+    ϵ = 1e-8
+    if norm(k - round.(k)) < ϵ
+        return 1
+    elseif norm(2k - round.(2k)) < 2ϵ
+        return 2
+    else
+        return 3
+    end
+end
+
 function fourier_bilinear_interaction(swt::SpinWaveTheory, q)
     (;sys, data) = swt
-    (; local_rotations, stevens_coefs, sqrtS) = data
-    (; extfield, gs) = sys
-    L = nbands(swt)
+    (; local_rotations) = data
+    (; gs) = sys
     @assert sys.dims == (1, 1, 1) "System must have only a single cell"
     Rs = local_rotations
     Na = natoms(sys.crystal)
-    J_k = zeros(ComplexF64, 3, Na, 3, Na)
+    J_k = zeros(CMat3, Na, Na)
 
     for i in 1:natoms(sys.crystal)
         for coupling in sys.interactions_union[i].pair
@@ -54,8 +77,8 @@ function fourier_bilinear_interaction(swt::SpinWaveTheory, q)
             (; j, n) = bond
             J_undo = Rs[i] * Mat3(bilin*I) * Rs[j]' # Undo the transformation for exchange Matrix
             J = exp(-2π * im * dot(q, n)) * J_undo
-            J_k[:, i, :, j] += J / 2
-            J_k[:, j, :, i] += J' / 2
+            J_k[i, j] += J / 2
+            J_k[j, i] += J' / 2
         end
     end
 
@@ -63,10 +86,10 @@ function fourier_bilinear_interaction(swt::SpinWaveTheory, q)
         A = precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), -q) * sys.ewald.μ0_μB²
         A = reshape(A, Na, Na)
         for i in 1:Na, j in 1:Na
-            J_k[:, i, :, j] += gs[i]' * A[i, j] * gs[j] / 2
-        end    
-    end       
-    return J_k            
+            J_k[i, j] += gs[i]' * A[i, j] * gs[j] / 2
+        end
+    end
+    return J_k
 end
 
 
@@ -85,10 +108,9 @@ function swt_hamiltonian_dipole_spiral!(H::Matrix{ComplexF64}, sswt::SpinWaveThe
     H12 = view(H, 1:L, L+1:2L)
     H21 = view(H, L+1:2L, 1:L)
     H22 = view(H, L+1:2L, L+1:2L)
[…]
