### 2025-02-17 b71e430
Optimize and fix g
