+++ src/Spiral/SpinWaveTheorySpiral2.jl
@@ -1,348 +1,0 @@
-"""
-    SpinWaveTheorySpiral(sys::System; k, axis, measure, regularization=1e-8)
-
-Analogous to [`SpinWaveTheory`](@ref), but interprets the provided system as
-having a generalized spiral order. This order is described by a single
-propagation wavevector `k`, which may be incommensurate. The `axis` vector
-defines the polarization plane via its surface normal. Typically the spin
-configuration in `sys` and the propagation wavevector `k` will be optimized
-using [`minimize_spiral_energy!`](@ref). In contrast, `axis` will typically be
-determined from symmetry considerations.
-
-The resulting object can be used to calculate the spin wave
-[`dispersion`](@ref), or the structure factor via [`intensities_bands`](@ref)
-and [`intensities`](@ref).
-
-The algorithm for this calculation was developed in [Toth and Lake, J. Phys.:
-Condens. Matter **27**, 166002 (2015)](https://arxiv.org/abs/1402.6069) and
-implemented in the [SpinW code](https://spinw.org/).
-"""
-struct SpinWaveTheorySpiral <: AbstractSpinWaveTheory
-    swt :: SpinWaveTheory
-    k :: Vec3
-    axis :: Vec3
-    buffers :: Vector{Array{CMat3, 2}}
-
-    function SpinWaveTheorySpiral(sys::System; k::AbstractVector, axis::AbstractVector, measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
-        L = length(eachsite(sys))
-        buffers = [zeros(CMat3, L, L) for _ in 1:6]
-        return new(SpinWaveTheory(sys; measure, regularization), k, axis, buffers)
-    end
-end
-
-function construct_uniaxial_anisotropy(; axis, c20=0., c40=0., c60=0., S)
-    # Anisotropy operator in local frame
-    O = stevens_matrices(S)
-    op = c20*O[2, 0] + c40*O[4, 0] + c60*O[6, 0]
-    # Rotate operator into global frame, defined by axis
-    R = rotation_between_vectors(axis, [0, 0, 1])
-    return rotate_operator(op, R)
-end
-
-# Sunny's original implementation followed Toth and Lake, which employed
-# pairwise exchange interactions J in real-space. Fourier-space exchange J(q) is
-# required, for example, to support long-range dipole-dipole interactions. Alin
-# Niraula and Xiaojian Bai developed and implemented the approprate
-# generalization in https://github.com/SunnySuite/Sunny.jl/pull/304.
-
-function fourier_bilinear_interaction!(J_k, swt::SpinWaveTheory, q)
-    (; sys, data) = swt
-    (; local_rotations) = data
-    (; gs) = sys
-    @assert sys.dims == (1, 1, 1) "System must have only a single cell"
-    Rs = local_rotations
-    Na = natoms(sys.crystal)
-    fill!(J_k, zero(CMat3))
-
-    for i in 1:natoms(sys.crystal)
-        for coupling in sys.interactions_union[i].pair
-            (; isculled, bond, bilin) = coupling
-            isculled && break
-
-            (; j, n) = bond
[…]
