### 2025-02-21 62c430f
Removing SW08 timing.
