+++ test/test_spiral.jl
@@ -49,42 +49,49 @@ end
 
 
 @testitem "Canted AFM" begin
-    function test_canted_afm(; s)
-        J, D, h = 1.0, 0.54, 0.76
-        rcs = Sunny.rcs_factors(s)[2]
-        a = 1
-        latvecs = lattice_vectors(a, a, 10a, 90, 90, 90)
-        positions = [[0, 0, 0]]
-        cryst = Crystal(latvecs, positions)
+    s = 3/2
+    J, D, h = 1.0, 0.54, 0.76
+    a = 1
+    latvecs = lattice_vectors(a, a, 10a, 90, 90, 90)
+    positions = [[0, 0, 0]]
+    cryst = Crystal(latvecs, positions)
 
-        sys = System(cryst, [1 => Moment(; s, g=-1)], :dipole)
-        set_exchange!(sys, J, Bond(1, 1, [1, 0, 0]))
-        set_onsite_coupling!(sys, S -> (D/rcs)*S[3]^2, 1)
-        set_field!(sys, [0, 0, h])
+    sys = System(cryst, [1 => Moment(; s, g=-1)], :dipole_uncorrected)
+    set_exchange!(sys, J, Bond(1, 1, [1, 0, 0]))
+    set_onsite_coupling!(sys, S -> D*S[3]^2, 1)
+    set_field!(sys, [0, 0, h])
 
-        k = Sunny.minimize_luttinger_tisza_exchange(sys; k_guess=randn(3))
-        @test k[1:2] ≈ [0.5, 0.5]
+    k = Sunny.minimize_luttinger_tisza_exchange(sys; k_guess=randn(3))
+    @test k[1:2] ≈ [0.5, 0.5]
 
-        axis = [0, 0, 1]
-        randomize_spins!(sys)
-        k = minimize_spiral_energy!(sys, axis; k_guess=randn(3))
-        @test k[1:2] ≈ [0.5, 0.5]
-        @test isapprox(only(sys.dipoles)[3], h / (8J + 2D); atol=1e-6)
+    axis = [0, 0, 1]
+    randomize_spins!(sys)
+    k = minimize_spiral_energy!(sys, axis; k_guess=randn(3))
+    @test k[1:2] ≈ [0.5, 0.5]
+    @test isapprox(only(sys.dipoles)[3], h / (8J + 2D); atol=1e-6)
 
-        q = [0.12, 0.23, 0.34]
-        swt = SpinWaveTheorySpiral(sys; measure=nothing, k, axis)
-        ϵq_num = dispersion(swt, [q])
+    q = [0.12, 0.23, 0.34]
+    swt = SpinWaveTheorySpiral(sys; measure=ssf_trace(sys), k, axis)
+    res = intensities_bands(swt, [q])
 
-        # Analytical
-        θ = acos(h / (2s*(4J+D)))
-        Jq = 2J*(cos(2π*q[1])+cos(2π*q[2]))
-        ϵq_ana = real(√Complex(4J*s*(4J*s+2D*s*sin(θ)^2) + cos(2θ)*(Jq*s)^2 + 2s*Jq*(4J*s*cos(θ)^2 + D*s*sin(θ)^2)))
+    # Analytical check on dispersion
 
-        @test ϵq_num[begin, 1] ≈ ϵq_ana
-    end
+    θ = acos(h / (2s*(4J+D)))
+    Jq = 2J*(cos(2π*q[1])+cos(2π*q[2]))
+    disp_ref = real(√Complex(4J*s*(4J*s+2D*s*sin(θ)^2) + cos(2θ)*(Jq*s)^2 + 2s*Jq*(4J*s*cos(θ)^2 + D*s*sin(θ)^2)))
[…]
