### 2025-02-21 01cacf5
Rewrite of Spiral SWT (#304)
Spiral LSWT calculations now employ exchange interactions expressed in Fourier space, J(q). This allows to support long-range dipole-dipole interactions. This commit also fixes correctness when the ordering wavevector k has components that are all integer or all half-integer.

Co-authored-by: Xiaojian Bai <39226486+XiaojianBai@users.noreply.github.com>
Co-authored-by: AlinNiraula <156613755+AlinNiraula@users.noreply.github.com>
