+++ src/Spiral/SpinWaveTheorySpiral.jl
@@ -21,9 +21,12 @@ struct SpinWaveTheorySpiral <: AbstractSpinWaveTheory
     swt :: SpinWaveTheory
     k :: Vec3
     axis :: Vec3
+    buffers :: Vector{Array{CMat3, 2}}
 
     function SpinWaveTheorySpiral(sys::System; k::AbstractVector, axis::AbstractVector, measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
-        return new(SpinWaveTheory(sys; measure, regularization), k, axis)
+        L = length(eachsite(sys))
+        buffers = [zeros(CMat3, L, L) for _ in 1:6]
+        return new(SpinWaveTheory(sys; measure, regularization), k, axis, buffers)
     end
 end
 
@@ -34,6 +37,43 @@ function construct_uniaxial_anisotropy(; axis, c20=0., c40=0., c60=0., s)
     # Rotate operator into global frame, defined by axis
     R = rotation_between_vectors(axis, [0, 0, 1])
     return rotate_operator(op, R)
+end
+
+# Sunny's original implementation followed Toth and Lake, which employed
+# pairwise exchange interactions J in real-space. Fourier-space exchange J(q) is
+# required, for example, to support long-range dipole-dipole interactions. Alin
+# Niraula and Xiaojian Bai developed and implemented the approprate
+# generalization in https://github.com/SunnySuite/Sunny.jl/pull/304.
+
+function fourier_bilinear_interaction!(J_k, swt::SpinWaveTheory, q)
+    (; sys, data) = swt
+    (; local_rotations) = data
+    (; gs) = sys
+    @assert sys.dims == (1, 1, 1) "System must have only a single cell"
+    Rs = local_rotations
+    Na = natoms(sys.crystal)
+    fill!(J_k, zero(CMat3))
+
+    for i in 1:natoms(sys.crystal)
+        for coupling in sys.interactions_union[i].pair
+            (; isculled, bond, bilin) = coupling
+            isculled && break
+
+            (; j, n) = bond
+            J_lab = Rs[i] * Mat3(bilin*I) * Rs[j]' # Undo transformation in `swt_data`
+            J = exp(-2π * im * dot(q, n)) * J_lab
+            J_k[i, j] += J
+            J_k[j, i] += J'
+        end
+    end
+
+    if !isnothing(sys.ewald)
+        A = precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), -q) * sys.ewald.μ0_μB²
+        A = reshape(A, Na, Na)
+        for i in 1:Na, j in 1:Na
+            J_k[i, j] += gs[i]' * A[i, j] * gs[j]
+        end
+    end
 end
 
 
@@ -47,42 +87,73 @@ function swt_hamiltonian_dipole_spiral!(H::Matrix{ComplexF64}, sswt::SpinWaveThe
     @assert size(H) == (2L, 2L)
     H .= 0.0
 
+    # preallocation
+    H11 = view(H, 1:L, 1:L)
+    H12 = view(H, 1:L, L+1:2L)
+    H21 = view(H, L+1:2L, 1:L)
+    H22 = view(H, L+1:2L, L+1:2L)
+    nx = SA[0 -axis[3] axis[2]; axis[3] 0 -axis[1]; -axis[2] axis[1] 0]
+    R2 = axis * axis'
+    R1 = (I - im * nx - R2) / 2
+    Rs = local_rotations
+
+    q_reshaped = q_reshaped + (branch - 2) .* k
[…]
