### 2025-02-22 ffe2400
Fix anisotropy scale and cleanup
