+++ test/test_scga.jl
@@ -5,9 +5,7 @@
     a = 8.5031 # (Å)
     latvecs = lattice_vectors(a, a, a, 90, 90, 90)
     cryst = Crystal(latvecs, [[0,0,0]], 227; choice="1")
-    dims = (1, 1, 1)
-    spininfos = [1 => Moment(; s=1, g=1)]
-    sys = System(cryst, spininfos, :dipole;dims, seed=0)
+    sys = System(cryst, [1 => Moment(; s=1, g=1)], :dipole; seed=0)
     set_exchange!(sys, -1, Bond(1, 3, [0,0,0]))
     set_exchange!(sys, 0.25, Bond(1, 2, [0,0,0]))
     measure = ssf_perp(sys;)
@@ -24,19 +22,17 @@ end
 @testitem "square_lattice" begin
     # test against JuliaSCGA (S. Gao)
     tol = 1e-5
-    sq=[0.3578208506624103, 0.3850668587212228, 0.4211633125595451, 0.3930558742921638, 0.3586715762816389, 0.3775147329089877, 0.4188431376589759, 0.4009835864744404, 0.36119385315852837, 0.3850668587212228, 0.4063051066815142, 0.4182331099724885, 
-    0.36751592492720886, 0.32829927921801794, 0.3490454303292574, 0.4062747596231383, 0.41626772462487194, 0.38789161791308147, 0.4211633125595451, 0.4182331099724885, 0.3710074695764304, 0.2924827325221696, 0.253584799014828, 0.2732638980307635, 
-    0.34406882609120737, 0.409719043169102, 0.4213848627804445, 0.3930558742921638, 0.36751592492720886, 0.2924827325221696, 0.21946942640699554, 0.18901809131848027, 0.2041521166894061, 0.26469888441402106, 0.3466451094806684, 0.3903985131607334, 
-    0.3586715762816389, 0.32829927921801794, 0.25358479901482806, 0.18901809131848027, 0.16311271576100106, 0.17594081183630647, 0.2284557904965832, 0.3060310966855855, 0.35526653426084637, 0.3775147329089877, 0.3490454303292574, 0.2732638980307635, 
-    0.20415211668940608, 0.17594081183630647, 0.18993269166531307, 0.2466353934125701, 0.32714470142342356, 0.37442377142308597, 0.4188431376589759, 0.4062747596231383, 0.34406882609120737, 0.26469888441402106, 0.2284557904965832, 0.2466353934125701, 
-    0.3154328042516787, 0.3918720231667212, 0.4179005987958521, 0.4009835864744404, 0.41626772462487194, 0.409719043169102, 0.3466451094806684, 0.3060310966855855, 0.32714470142342356, 0.3918720231667212, 0.42107414035570556, 0.40321330759757756, 
-    0.36119385315852837, 0.38789161791308147, 0.4213848627804445, 0.3903985131607334, 0.35526653426084637, 0.37442377142308597, 0.4179005987958521, 0.40321330759757756, 0.3645203324273335];
+    sq = [0.3578208506624103, 0.3850668587212228, 0.4211633125595451, 0.3930558742921638, 0.3586715762816389, 0.3775147329089877, 0.4188431376589759, 0.4009835864744404, 0.36119385315852837, 0.3850668587212228, 0.4063051066815142, 0.4182331099724885,
[…]
