### 2025-02-23 e27a2b3
Streamline interface
