+++ src/SCGA/SCGA.jl
@@ -1,22 +1,44 @@
-"""
-    SCGA(sys::System; measure, quantum_sum_rule=true)
+# The default value `sublattice_resolved=true` is generally preferred, and imposes
+# one global sum-rule constraint per spin-sublattice. Selecting instead
+# `sublattice_resolved=false` will employ only a _single_ constraint that is
+# averaged over sublattices. This will be faster, but risks producing incorrect
+# results when the sublattices are symmetry-distinct.
 
-Constructs an object to perform the self consistent gaussian approximation.
-Enables the use of `intensities_static`](@ref) with an SCGA type to calculate
-the static structure factor in the paramagnetic phase. If the temperature is
-below the ordering temperature, the intensities will be negative.
+"""
+    SCGA(sys::System; measure, Nq, quantum_sum_rule=false)
+
+Constructs an object to calculate [`intensities_static`](@ref) within the self
+consistent gaussian approximation (SCGA). This approximation assumes a classical
+Boltzmann distribution, and is expected to be meangingful above the ordering
+temperature, where fluctuations are approximately Gaussian. If the temperature
+is not sufficiently high, then `intensities_static` may report negative
+energies, which would indicate an instability to magnetic ordering.
+
+Currently only `:dipole` and `:dipole_uncorrected` system modes are supported.
+
+The theory of SCGA approximates local spin magnitude constraints with a _weaker_
+global constraint condition. This global constraint is implemented as a sum
+rule, expressed as an integral over Fourier modes. This integral is approximated
+as a discrete sum over `Nq^3` wavevectors for the provided integer `Nq`.
+
+By default, each classical spin dipole is assumed to have magnitude ``s`` that
+matches the [`Moment`](@ref) specification. Selecting `quantum_sum_rule=true`
+will modify this magnitude to ``√s(s+1)``.
 """
 struct SCGA
     sys :: System
     measure :: MeasureSpec
+    Nq :: Int
     quantum_sum_rule :: Bool
+    sublattice_resolved :: Bool
 
-    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, quantum_sum_rule)
+    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, Nq::Int, quantum_sum_rule=false, sublattice_resolved=true)
         measure = @something measure empty_measurespec(sys)
+        # FIXME: Copy logic from Spiral SWT
         if length(eachsite(sys)) != prod(size(measure.observables)[2:5])
             error("Size mismatch. Check that measure is built using consistent system.")
         end
-        return new(sys, measure, quantum_sum_rule)
+        return new(sys, measure, Nq, quantum_sum_rule, sublattice_resolved)
     end
 end
 
@@ -53,6 +75,7 @@ function fourier_transform_interaction_matrix(sys::System; k)
     for i in 1:Na
         onsite_coupling = sys.interactions_union[i].onsite
         (; c2, c4, c6) = onsite_coupling
+        iszero(c4) && iszero(c6) || error("Single-ion anisotropy beyond quadratic order not supported")
         anisotropy = [c2[1]-c2[3]        c2[5] 0.5c2[2];
[…]
