### 2025-03-22 78549d2
Fix spin_label and related indexing.
