+++ src/System/Interactions.jl
@@ -285,7 +285,7 @@ function energy(sys::System{N}) where N
         for i in 1:natoms(sys.crystal)
             # Interactions for sublattice i (same for every cell)
             interactions = sys.interactions_union[i]
-            E += energy_aux(interactions, sys, eachsite(sys, i))
+            E += energy_aux(interactions, sys, eachsite_sublattice(sys, i))
         end
     else
         for site in eachsite(sys)
@@ -396,7 +396,7 @@ function set_energy_grad_dipoles!(∇E, dipoles::Array{Vec3, 4}, sys::System{N})
     if is_homogeneous(sys)
         for i in 1:natoms(sys.crystal)
             interactions = sys.interactions_union[i]
-            set_energy_grad_dipoles_aux!(∇E, dipoles, interactions, sys, eachsite(sys, i))
+            set_energy_grad_dipoles_aux!(∇E, dipoles, interactions, sys, eachsite_sublattice(sys, i))
         end
     else
         for site in eachsite(sys)
@@ -481,7 +481,7 @@ function set_energy_grad_coherents!(HZ, Z::Array{CVec{N}, 4}, sys::System{N}) wh
     if is_homogeneous(sys)
         for i in 1:natoms(sys.crystal)
             interactions = sys.interactions_union[i]
-            set_energy_grad_coherents_aux!(HZ, Z, dE_dS, interactions, sys, eachsite(sys, i))
+            set_energy_grad_coherents_aux!(HZ, Z, dE_dS, interactions, sys, eachsite_sublattice(sys, i))
         end
     else
         for site in eachsite(sys)
