### 2025-03-22 63b2bc1
Fix spin_label and related indexing.
