+++ src/Spiral/LuttingerTisza.jl
@@ -1,7 +1,14 @@
-# Returns the minimum eigenvalue of the Luttinger-Tisza exchange matrix. It is a
-# lower-bound on the exchange energy for magnetic order that has propagation
-# wavevector k between Bravais lattice cells. This analysis ignores local
-# normalization constraints for the spins within the cell.
+# Returns the Luttinger-Tisza predicted exchange energy associated with the
+# propagation wavevector k between chemical cells. The LT analysis minimizes
+# energy E = (1/2) Sₖ† Jₖ Sₖ, where Sₖ is some length-3Nₐ vector and Jₖ is the
+# 3Nₐ×3Nₐ exchange matrix. Given a minimum energy eigenpair (λₖ, Sₖ) the
+# LT-predicted energy is |Sₖ|^2 λₖ/2 with normalization |Sₖ|^2 = ∑ᵢ|Sᵢ|^2, where
+# i denotes spin sublattice of the chemical cell. If the components of Sₖ
+# satisfy local spin normalization constraints, then the LT energy minimized
+# over k is physically correct for the spiral ground state. In practice, the
+# eigenvector Sₖ may violate local spin normalization and the LT-predicted
+# energy is only a lower-bound on the exchange energy. Conditions for
+# correctness are given by Xiong and Wen (2013), arXiv:1208.1512.
 function luttinger_tisza_exchange(sys::System; k, ϵ=0)
     @assert sys.mode in (:dipole, :dipole_uncorrected) "SU(N) mode not supported"
     @assert sys.dims == (1, 1, 1) "System must have only a single cell"
@@ -16,8 +23,8 @@ function luttinger_tisza_exchange(sys::System; k, ϵ=0)
 
             (; j, n) = bond
             J = exp(2π * im * dot(k, n)) * Mat3(bilin*I)
-            J_k[:, i, :, j] += J / 2
-            J_k[:, j, :, i] += J' / 2
+            J_k[:, i, :, j] += J
+            J_k[:, j, :, i] += J'
         end
     end
 
@@ -25,7 +32,7 @@ function luttinger_tisza_exchange(sys::System; k, ϵ=0)
         A = Sunny.precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), k) * sys.ewald.μ0_μB²
         A = reshape(A, Na, Na)
         for i in 1:Na, j in 1:Na
-            J_k[:, i, :, j] += sys.gs[i]' * A[i, j] * sys.gs[j] / 2
+            J_k[:, i, :, j] += sys.gs[i]' * A[i, j] * sys.gs[j]
         end
     end
 
@@ -34,14 +41,14 @@ function luttinger_tisza_exchange(sys::System; k, ϵ=0)
     J_k = hermitianpart(J_k)
 
     E = if iszero(ϵ)
-        eigmin(J_k)
+        eigmin(J_k) / 2
     else
         # Estimate the minimum eigenvalue E as a weighted average of all small
         # eigenvalues, E = tr [exp(-β J) J] / tr exp(-β J), where β = 1/ϵ.
         # Finite but small ϵ will regularize the potential energy surface in the
         # viscinity of degenerate eigenvalues. Imposing this smoothness may aid
         # optimization of k with gradient-based methods.
-        λs = eigvals(J_k)
+        λs = eigvals(J_k) / 2
         λmin = minimum(λs)
         # Scale all weights exp(-λ/ϵ) by the factor exp(λmin/ϵ). Scaling of
         # weights has no mathematical effect, but avoids numerical overflow.
@@ -49,8 +56,8 @@ function luttinger_tisza_exchange(sys::System; k, ϵ=0)
         sum(ws .* λs) / sum(ws)
     end
 
[…]
