### 2025-03-24 78567cc
Fix anisotropy scale and cleanup
