+++ src/SCGA/SCGA.jl
@@ -11,7 +11,7 @@ struct SCGA
     measure :: MeasureSpec
     quantum_sum_rule :: Bool
 
-    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, quantum_sum_rule=true)
+    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, quantum_sum_rule)
         measure = @something measure empty_measurespec(sys)
         if length(eachsite(sys)) != prod(size(measure.observables)[2:5])
             error("Size mismatch. Check that measure is built using consistent system.")
@@ -25,7 +25,7 @@ end
 # q=k. This function is adapted from `luttinger_tisza_exchange`](@ref) to
 # include the correct prefactors and local anisotropies.
 function fourier_transform_interaction_matrix(sys::System; k)
-    @assert sys.mode in (:dipole, :dipole_large_S) "SU(N) mode not supported"
+    @assert sys.mode in (:dipole, :dipole_uncorrected) "SU(N) mode not supported"
     @assert sys.dims == (1, 1, 1) "System must have only a single cell"
     Na = natoms(sys.crystal)
     J_k = zeros(ComplexF64, 3, Na, 3, Na)
@@ -56,7 +56,7 @@ function fourier_transform_interaction_matrix(sys::System; k)
         anisotropy = [c2[1]-c2[3]        c2[5] 0.5c2[2];
                             c2[5] -c2[1]-c2[3] 0.5c2[4];
                          0.5c2[2]     0.5c2[4]   2c2[3]]
-        J_k[:, i, :, i] += 2anisotropy # FIXME: Understand this factor of 2
+        J_k[:, i, :, i] += anisotropy
     end
 
     J_k = reshape(J_k, 3*Na, 3*Na)
@@ -365,128 +365,6 @@ function find_lagrange_multiplier_opt_sublattice(scga, λs, kT; maxiters=500, to
     return real.(min)
 end
 
-#=
-function find_lagrange_multiplier_opt_WORKING(scga, λs, kT; method="ConjugateGradient", maxiters=500, tol=1e-10)
-    (; sys, quantum_sum_rule) = scga
-    if quantum_sum_rule
-        S_sq = vec(sys.κs .* (sys.κs .+ 1))
-    else
-        S_sq = vec(sys.κs.^2)
-    end
-    Na = natoms(sys.crystal)
-    Nq = 8
-    dq = 1/Nq;
-    qarray = -0.5: dq : 0.5-dq
-    q = [[qx, qy, qz] for qx in qarray, qy in qarray, qz in qarray]
-    N = length(q)
-    if method == "ConjugateGradient"
-        function f(λs)
-            Λ =  diagm(repeat(λs, inner=3))
-            A_array = [(1/kT)*fourier_transform_interaction_matrix(sys; k=q_in) .+  (1/kT)*Λ for q_in in q]
-            eig_vals = zeros(3Na,length(A_array))
-            Us = zeros(ComplexF64,3Na,3Na,length(A_array))
-            for j in 1:length(A_array)
-                T = eigen(A_array[j])
-                eig_vals[:,j] .= T.values
-                Us[:,:,j] .= T.vectors
-            end
-            if minimum(eig_vals) < 0
-                F =  -Inf
-            else
-                F =  0.5*kT*sum(log.(eig_vals))
-            end
-            G = F - 0.5*N*sum(λs.*S_sq)
-            return -G
-        end
-        function g!(storage,λs)
-        # gradient
-            Λ =  diagm(repeat(λs, inner=3))
[…]
