### 2025-03-24 2483a60
Fix anisotropy strength
