+++ src/SCGA/SCGA.jl
@@ -42,17 +42,12 @@ struct SCGA
 end
 
 
-# Computes the Lagrange multiplier for the standard SCGA approach with a common
-# Lagrange multiplier for all sublattices.
-function find_lagrange_multiplier(scga::SCGA, β)
-    starting_offset = 0.2
-    maxiters = 500
-    tol = 1e-10
+function make_q_grid(scga)
+    (; sys, Nq) = scga
 
-    (; sys, quantum_sum_rule, Nq) = scga
-    dq = 1/Nq;
     Na = natoms(sys.crystal)
-    bond_counter = zeros(Float64,3)
+    dq = 1/Nq;
+    bond_counter = zeros(Float64, 3)
     for i in 1:Na
         for coupling in sys.interactions_union[i].pair
             (; isculled, bond) = coupling
@@ -64,12 +59,27 @@ function find_lagrange_multiplier(scga::SCGA, β)
     qarrays = []
     for i in 1:3
         if bond_counter[i] == 0
-            push!(qarrays,[0])
+            push!(qarrays, [0])
         else
-            push!(qarrays,qarray)
+            push!(qarrays, qarray)
         end
     end
-    q = [[qx, qy, qz] for qx in qarrays[1], qy in qarrays[2], qz in qarrays[3]]
+    return [[qx, qy, qz] for qx in qarrays[1], qy in qarrays[2], qz in qarrays[3]]
+end
+
+# Computes the Lagrange multiplier for the standard SCGA approach with a common
+# Lagrange multiplier for all sublattices.
+function find_lagrange_multiplier(scga::SCGA, β)
+    starting_offset = 0.2
+    maxiters = 500
+    tol = 1e-10
+
+    (; sys, quantum_sum_rule) = scga
+    Na = natoms(sys.crystal)
+
+    q = make_q_grid(scga)
+    N = length(q)
+
     Jq_array = [fourier_exchange_matrix(sys; k=q_in) for q_in in q]
     #TODO throw error if spins different
     if quantum_sum_rule
@@ -78,18 +88,18 @@ function find_lagrange_multiplier(scga::SCGA, β)
         S_sq = sum(sys.κs.^2) / Na
     end
 
-    eig_vals = zeros(3Na, length(Jq_array))
+    eig_vals = zeros(3Na, N)
     for j in eachindex(Jq_array)
          eig_vals[:, j] .= eigvals(Jq_array[j])
     end
 
     function f(λ)
         sum_term = sum(1 ./ (λ .+ β * eig_vals ))
-        return (1/(Na*length(q))) * sum_term
+        return (1/(Na*N)) * sum_term
     end
     function J(λ)
         sum_term = sum((1 ./ (λ .+ β * eig_vals)).^2)
-        return -(1/(Na*length(q))) * sum_term
+        return -(1/(Na*N)) * sum_term
     end
 
     lower = - β * minimum(eig_vals)
@@ -201,13 +211,14 @@ function intensities_static(scga::SCGA, qpts; kT, λs_init=nothing)
     end
 end
 
+
 # Computes the Lagrange multiplier for the sublattice resolved SCGA method.
 
 function find_lagrange_multiplier_opt_sublattice(scga, λs, β)
     tol = 1e-6
     maxiters = 500
 
-    (; sys, quantum_sum_rule, Nq) = scga
+    (; sys, quantum_sum_rule) = scga
 
     if quantum_sum_rule
         S_sq =vec(sys.κs .* (sys.κs .+ 1))
@@ -215,51 +226,12 @@ function find_lagrange_multiplier_opt_sublattice(scga, λs, β)
         S_sq = vec(sys.κs.^2)
     end
     Na = natoms(sys.crystal)
-    dq = 1/Nq;
 
-    bond_counter = zeros(Float64,3)
[…]
