+++ src/SCGA/SCGA.jl
@@ -108,15 +108,11 @@ end
 
 const CMat3 = SMatrix{3, 3, ComplexF64, 9}
 
-# Computes the static structure factor in the standard SCGA approach, with a
-# single Lagrange multiplier, over the qpts specified.
-
-function intensities_static_single(scga::SCGA, qpts; β)
+function intensities_static_aux(scga::SCGA, qpts; β, Λ)
     qpts = convert(AbstractQPoints, qpts)
     (; sys, measure) = scga
     Na = natoms(sys.crystal)
     Nobs = num_observables(measure)
-    λ = find_lagrange_multiplier(scga, β)
     intensity = zeros(eltype(measure),length(qpts.qs))
     Ncorr = length(measure.corr_pairs)
     r = sys.crystal.positions
@@ -132,9 +128,7 @@ function intensities_static_single(scga::SCGA, qpts; β)
             pref[μ, i] = exp(-2π * im * dot(q_reshaped, r[i])) * compute_form_factor(ff, norm2(q_global))
         end
         Jq = fourier_exchange_matrix(sys; k=q_reshaped)         # FIXME: q not q_reshaped
-        # Jq = reshape(Jq, 3, Na, 3, Na)
-
-        inverted_matrix = inv(I(3Na)*λ + β*Jq) # this is [(Iλ+J(q))^-1]^αβ_μν
+        inverted_matrix = inv(β*Λ + β*Jq) # this is [(Iλ+J(q))^-1]^αβ_μν
         inverted_matrix = reshape(inverted_matrix, 3, Na, 3, Na)
 
         for i in 1:Na, j in 1:Na
@@ -146,63 +140,26 @@ function intensities_static_single(scga::SCGA, qpts; β)
         intensity[iq] = measure.combiner(q_global, corrbuf)
     end
     if extrema(intensity)[1] < -1e-2
-        @warn "Warning: negative intensities! kT is probably below the ordering temperature."
-        # TODO Throw an error. This is for diagnostic purposes.
+        error("Negative intensity indicates that kT is below ordering")
     end
     return StaticIntensities(sys.crystal, qpts, reshape(intensity,size(qpts.qs)))
 end
 
-# Computes the static structure factor in the sublattice resolved SCGA method,
-# over the qpts specified.
-
-function intensities_static_sublattice(scga::SCGA, qpts; β, λs_init)
-    qpts = convert(AbstractQPoints, qpts)
-    (; sys, measure) = scga
-    Na = natoms(sys.crystal)
-    Nobs = num_observables(measure)
-    λs = find_lagrange_multiplier_opt_sublattice(scga, λs_init, β)
-    intensity = zeros(eltype(measure),length(qpts.qs))
-    Ncorr = length(measure.corr_pairs)
-    r = sys.crystal.positions
-
-    for (iq, q) in enumerate(qpts.qs)
-        pref = zeros(ComplexF64, Nobs, Na)
-        corrbuf = zeros(ComplexF64, Ncorr)
-        intensitybuf = zeros(eltype(measure),3,3)
-        q_reshaped = to_reshaped_rlu(sys, q)
-        q_global = sys.crystal.recipvecs * q
-        for i in 1:Na, μ in 1:3
-            ff = get_swt_formfactor(measure, μ, i)
-            pref[μ, i] = exp(-2π * im * dot(q_reshaped, r[i])) * compute_form_factor(ff, norm2(q_global))
-        end
-        J_mat = fourier_exchange_matrix(sys; k=q_reshaped)
-        Λ =  diagm(repeat(λs, inner=3))
-        inverted_matrix = (inv(β*Λ + β*J_mat)) # this is [(Iλ+J(q))^-1]^αβ_μν
-        for i in 1:Na, j in 1:Na
[…]
