### 2025-03-24 20b370a
Some refactors
