+++ src/SCGA/SCGA.jl
@@ -4,6 +4,7 @@
 # averaged over sublattices. This will be faster, but risks producing incorrect
 # results when the sublattices are symmetry-distinct.
 
+
 """
     SCGA(sys::System; measure, Nq, quantum_sum_rule=false)
 
@@ -28,23 +29,30 @@ will modify this magnitude to ``√s(s+1)``.
 struct SCGA
     sys :: System
     measure :: MeasureSpec
-    Nq :: Int
     quantum_sum_rule :: Bool
     sublattice_resolved :: Bool
+
+    qs :: Vector{Vec3}
+    Js :: Vector{HermitianC64}
+    eigvals :: Vector{Vector{Float64}}
+    eigvecs :: Vector{Matrix{ComplexF64}}
 
     function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, Nq::Int, quantum_sum_rule=false, sublattice_resolved=true)
         measure = @something measure empty_measurespec(sys)
         if size(eachsite(sys)) != size(measure.observables)[2:5]
             error("Size mismatch. Check that measure is built using consistent system.")
         end
-        return new(sys, measure, Nq, quantum_sum_rule, sublattice_resolved)
+
+        qs = make_q_grid(sys, Nq)
+        Js = [fourier_exchange_matrix(sys; k) for k in qs]
+        eigs = map(eigen, Js)
+        eigvals = [e.values for e in eigs]
+        eigvecs = [e.vectors for e in eigs]
+        return new(sys, measure, quantum_sum_rule, sublattice_resolved, qs, Js, eigvals, eigvecs)
     end
 end
 
-
-function make_q_grid(scga)
-    (; sys, Nq) = scga
-
+function make_q_grid(sys, Nq)
     Na = natoms(sys.crystal)
     dq = 1/Nq;
     wraps = [false, false, false]
@@ -56,7 +64,7 @@ function make_q_grid(scga)
         end
     end
     qarrays = [w ? (-1/2 : dq : 1/2-dq) : [0] for w in wraps]
-    return collect(Iterators.product(qarrays...))
+    return vec(Vec3.(Iterators.product(qarrays...)))
 end
 
 # Computes the Lagrange multiplier for the standard SCGA approach with a common
@@ -66,35 +74,25 @@ function find_lagrange_multiplier(scga::SCGA, β)
     maxiters = 500
     tol = 1e-10
 
-    (; sys, quantum_sum_rule) = scga
+    (; sys, quantum_sum_rule, qs, eigvals) = scga
+    eigvals = Iterators.flatten(eigvals)
     Na = natoms(sys.crystal)
+    Nq = length(qs)
 
-    q = make_q_grid(scga)
-    Nq = length(q)
-
-    Jq_array = [fourier_exchange_matrix(sys; k=q_in) for q_in in q]
-    #TODO throw error if spins different
     if quantum_sum_rule
-        S_sq = sum(sys.κs .* (sys.κs.+1)) / Na
+        S_sq = sum(κ * (κ + 1) for κ in sys.κs) / Na
     else
-        S_sq = sum(sys.κs.^2) / Na
-    end
-
-    eig_vals = zeros(3Na, Nq)
-    for j in eachindex(Jq_array)
-         eig_vals[:, j] .= eigvals(Jq_array[j])
+        S_sq = norm2(sys.κs) / Na
     end
 
     function f(λ)
-        sum_term = sum(1 ./ (λ .+ β * eig_vals ))
-        return (1/(Na*Nq)) * sum_term
+        return sum(1 / (λ + β * ev) for ev in eigvals) / (Na * Nq)
     end
     function J(λ)
-        sum_term = sum((1 ./ (λ .+ β * eig_vals)).^2)
[…]
