+++ test/test_scga.jl
@@ -61,7 +61,7 @@ end
     # values from Conlon + Chalker
     MgCrO = [ 5.7767400e-02   7.4075451e-02   1.5971443e-01   5.5634087e-01   1.3136038e+00   5.5634087e-01   1.5971443e-01   7.4075451e-02   5.7767400e-02   7.4075451e-02   1.5971443e-01   5.5634087e-01   1.3136038e+00   5.5634087e-01   1.5971443e-01   7.4075451e-02   5.7767400e-02   7.4075451e-02   1.6242284e-01   6.1505984e-01   1.2377330e+00   1.8741195e+00   1.2377330e+00   6.1505984e-01   1.6242284e-01   7.4075451e-02   1.6242284e-01   6.1505984e-01   1.2377330e+00   1.8741195e+00   1.2377330e+00   6.1505984e-01   1.6242284e-01   7.4075451e-02   1.5971443e-01   6.1505984e-01   1.7143609e+00   2.8136620e+00   3.2690074e+00   2.8136620e+00   1.7143609e+00   6.1505984e-01   1.5971443e-01   6.1505984e-01   1.7143609e+00   2.8136620e+00   3.2690074e+00   2.8136620e+00   1.7143609e+00   6.1505984e-01   1.5971443e-01   5.5634087e-01   1.2377330e+00   2.8136620e+00   2.4168819e+00   1.8741195e+00   2.4168819e+00   2.8136620e+00   1.2377330e+00   5.5634087e-01   1.2377330e+00   2.8136620e+00   2.4168819e+00   1.8741195e+00   2.4168819e+00   2.8136620e+00   1.2377330e+00   5.5634087e-01   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.3136038e+00   5.5634087e-01   1.2377330e+00   2.8136620e+00   2.4168819e+00   1.8741195e+00   2.4168819e+00   2.8136620e+00   1.2377330e+00   5.5634087e-01   1.2377330e+00   2.8136620e+00   2.4168819e+00   1.8741195e+00   2.4168819e+00   2.8136620e+00   1.2377330e+00   5.5634087e-01   1.5971443e-01   6.1505984e-01   1.7143609e+00   2.8136620e+00   3.2690074e+00   2.8136620e+00   1.7143609e+00   6.1505984e-01   1.5971443e-01   6.1505984e-01   1.7143609e+00   2.8136620e+00   3.2690074e+00   2.8136620e+00   1.7143609e+00   6.1505984e-01   1.5971443e-01   7.4075451e-02   1.6242284e-01   6.1505984e-01   1.2377330e+00   1.8741195e+00   1.2377330e+00   6.1505984e-01   1.6242284e-01   7.4075451e-02   1.6242284e-01   6.1505984e-01   1.2377330e+00   1.8741195e+00   1.2377330e+00   6.1505984e-01   1.6242284e-01   7.4075451e-02   5.7767400e-02   7.4075451e-02   1.5971443e-01   5.5634087e-01   1.3136038e+00   5.5634087e-01   1.5971443e-01   7.4075451e-02   5.7767400e-02   7.4075451e-02   1.5971443e-01   5.5634087e-01   1.3136038e+00   5.5634087e-01   1.5971443e-01   7.4075451e-02   5.7767400e-02   7.4075451e-02   1.6242284e-01   6.1505984e-01   1.2377330e+00   1.
[…]
