+++ test/test_scga.jl
@@ -12,11 +12,11 @@
     set_exchange!(sys, -1/s^2, Bond(1, 3, [0, 0, 0]))
     set_exchange!(sys, 1/(2s)^2, Bond(1, 2, [0, 0, 0]))
     measure = ssf_perp(sys)
-    scga = Sunny.SCGA(sys; measure, Nq=8, sublattice_resolved=false)
     kT = 15*meV_per_K
+    scga = Sunny.SCGA(sys; measure, kT, Nq=8)
     γ = s^2 * Sunny.natoms(cryst)
     grid = q_space_grid(cryst, [1, 0, 0], 0:0.9:4, [0, 1, 0], 0:0.9:4; orthogonalize=true)
-    res = intensities_static(scga, grid; kT)
+    res = intensities_static(scga, grid)
     @test isapprox(vec(res.data)/γ, dia; rtol=1e-8)
 end
 
@@ -38,10 +38,10 @@ end
     set_exchange!(sys, 0.5, Bond(1, 1, [1,1,0]))
     set_exchange!(sys, 0.25, Bond(1, 1, [2,0,0]))
     measure = ssf_perp(sys)
-    scga = Sunny.SCGA(sys; measure, Nq=8, sublattice_resolved=false)
     kT = 27.5*meV_per_K
+    scga = Sunny.SCGA(sys; measure, kT, Nq=8)
     grid = q_space_grid(cryst, [1, 0, 0], -1:0.12:0, [0, 1, 0], -1:0.12:0; orthogonalize=true)
-    res = Sunny.intensities_static(scga, grid; kT)
+    res = Sunny.intensities_static(scga, grid)
     @test isapprox(vec(res.data)/2, sq; rtol=1e-5)
 end
 
@@ -61,10 +61,10 @@ end
     # values from Conlon + Chalker
     MgCrO = [ 5.7767400e-02   7.4075451e-02   1.5971443e-01   5.5634087e-01   1.3136038e+00   5.5634087e-01   1.5971443e-01   7.4075451e-02   5.7767400e-02   7.4075451e-02   1.5971443e-01   5.5634087e-01   1.3136038e+00   5.5634087e-01   1.5971443e-01   7.4075451e-02   5.7767400e-02   7.4075451e-02   1.6242284e-01   6.1505984e-01   1.2377330e+00   1.8741195e+00   1.2377330e+00   6.1505984e-01   1.6242284e-01   7.4075451e-02   1.6242284e-01   6.1505984e-01   1.2377330e+00   1.8741195e+00   1.2377330e+00   6.1505984e-01   1.6242284e-01   7.4075451e-02   1.5971443e-01   6.1505984e-01   1.7143609e+00   2.8136620e+00   3.2690074e+00   2.8136620e+00   1.7143609e+00   6.1505984e-01   1.5971443e-01   6.1505984e-01   1.7143609e+00   2.8136620e+00   3.2690074e+00   2.8136620e+00   1.7143609e+00   6.1505984e-01   1.5971443e-01   5.5634087e-01   1.2377330e+00   2.8136620e+00   2.4168819e+00   1.8741195e+00   2.4168819e+00   2.8136620e+00   1.2377330e+00   5.5634087e-01   1.2377330e+00   2.8136620e+00   2.4168819e+00   1.8741195e+00   2.4168819e+00   2.8136620e+00   1.2377330e+00   5.5634087e-01   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.3136038e+00   1.8741195e+00   3.2690074e+00   1.8741195e+00   1.
[…]
