+++ src/SCGA/SCGA.jl
@@ -1,21 +1,15 @@
-# The default value `sublattice_resolved=true` is generally preferred, and imposes
-# one global sum-rule constraint per spin-sublattice. Selecting instead
-# `sublattice_resolved=false` will employ only a _single_ constraint that is
-# averaged over sublattices. This will be faster, but risks producing incorrect
-# results when the sublattices are symmetry-distinct.
-
-
 """
-    SCGA(sys::System; measure, Nq, quantum_sum_rule=false)
+    SCGA(sys::System; measure, kT, Nq, quantum_sum_rule=false)
 
 Constructs an object to calculate [`intensities_static`](@ref) within the self
 consistent gaussian approximation (SCGA). This approximation assumes a classical
-Boltzmann distribution, and is expected to be meangingful above the ordering
-temperature, where fluctuations are approximately Gaussian. If the temperature
-is not sufficiently high, then `intensities_static` may report negative
-energies, which would indicate an instability to magnetic ordering.
+Boltzmann distribution with temperature `kT`. It is expected to be meangingful
+above the ordering temperature, where fluctuations are approximately Gaussian.
+If the temperature is not sufficiently high, then `intensities_static` may
+report negative energies, which would indicate an instability to magnetic
+ordering.
 
-Currently only `:dipole` and `:dipole_uncorrected` system modes are supported.
+Only `:dipole` and `:dipole_uncorrected` system modes are supported.
 
 The theory of SCGA approximates local spin magnitude constraints with a _weaker_
 global constraint condition. This global constraint is implemented as a sum
@@ -29,26 +23,30 @@ will modify this magnitude to ``√s(s+1)``.
 struct SCGA
     sys :: System
     measure :: MeasureSpec
-    quantum_sum_rule :: Bool
-    sublattice_resolved :: Bool
+    β :: Float64
+    λs :: Vector{Float64}
 
-    qs :: Vector{Vec3}
-    Js :: Vector{HermitianC64}
-    eigvals :: Vector{Vector{Float64}}
-    eigvecs :: Vector{Matrix{ComplexF64}}
-
-    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, Nq::Int, quantum_sum_rule=false, sublattice_resolved=true)
+    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, kT::Float64, Nq::Int, quantum_sum_rule=false)
         measure = @something measure empty_measurespec(sys)
         if size(eachsite(sys)) != size(measure.observables)[2:5]
             error("Size mismatch. Check that measure is built using consistent system.")
         end
 
+        kT > 0 || error("Temperature kT must be positive")
+        β = 1 / kT
+
         qs = make_q_grid(sys, Nq)
         Js = [fourier_exchange_matrix(sys; k) for k in qs]
-        eigs = map(eigen, Js)
-        eigvals = [e.values for e in eigs]
-        eigvecs = [e.vectors for e in eigs]
-        return new(sys, measure, quantum_sum_rule, sublattice_resolved, qs, Js, eigvals, eigvecs)
+
+        sublattice_resolved = !allequal(sys.crystal.classes)
+        if sublattice_resolved
[…]
