### 2025-03-24 9f121a2
Fix reshaping of q
