### 2025-03-24 9ac3d7c
Simplifications
