### 2025-03-24 e26be7b
Miscellaneous cleanups
