### 2025-03-25 9d67365
Assertion test
