### 2025-03-25 b9f98e2
Adapt to 1/dq notation
