+++ test/test_scga.jl
@@ -11,7 +11,7 @@
     set_exchange!(sys, 1/(2s)^2, Bond(1, 2, [0, 0, 0]))
     measure = ssf_perp(sys)
     kT = 15*meV_per_K
-    scga = Sunny.SCGA(sys; measure, kT, Nq=8)
+    scga = Sunny.SCGA(sys; measure, kT, dq=1/8)
     γ = s^2 * Sunny.natoms(cryst)
     grid = q_space_grid(cryst, [1, 0, 0], 0:0.9:4, [0, 1, 0], 0:0.9:4; orthogonalize=true)
     res = intensities_static(scga, grid)
@@ -36,37 +36,37 @@ end
     set_exchange!(sys, 0.25, Bond(1, 1, [2,0,0]))
     measure = ssf_perp(sys)
     kT = 27.5*meV_per_K
-    scga = Sunny.SCGA(sys; measure, kT, Nq=8)
+    scga = Sunny.SCGA(sys; measure, kT, dq=1/8)
     grid = q_space_grid(cryst, [1, 0, 0], -1:0.12:0, [0, 1, 0], -1:0.12:0; orthogonalize=true)
     res = Sunny.intensities_static(scga, grid)
     @test isapprox(vec(res.data)/2, sq; rtol=1e-5)
 end
 
 @testitem "MgCr2O4" begin
+    using LinearAlgebra
     # Reproduce calculation in Conlon and Chalker, PRL 102, 237206 (2009)
     latvecs    = lattice_vectors(8.3342, 8.3342, 8.3342, 90, 90, 90)
     positions  = [[1/2, 1/2, 1/2]]
     cryst = Crystal(latvecs, positions, 227)
-    sys = System(cryst, [1 => Moment(; s=3/2, g=1)], :dipole)
+    s = 3/2
+    sys = System(cryst, [1 => Moment(; s, g=1)], :dipole)
     sys = reshape_supercell(sys, primitive_cell(cryst))
+    Sunny.enable_quantum_dipole_amplification!(sys)
     J1 = 3.27                                            # in meV, from Bai's PRL
-    J_mgcro = [1.0, 0.0815, 0.1050, 0.0085] * J1;        # further neighbor pyrochlore relevant for MgCr2O4
+    J_mgcro = [1.0, 0.0815, 0.1050, 0.0085] * J1         # further neighbor pyrochlore relevant for MgCr2O4
     set_exchange!(sys, J_mgcro[1], Bond(1, 2, [0,0,0]))  # J1
     set_exchange!(sys, J_mgcro[2], Bond(1, 7, [0,0,0]))  # J2
     set_exchange!(sys, J_mgcro[3], Bond(1, 3, [1,0,0]))  # J3a -- Careful here!
-    set_exchange!(sys, J_mgcro[4], Bond(1, 3, [0,0,0])); # J3b -- And here!
+    set_exchange!(sys, J_mgcro[4], Bond(1, 3, [0,0,0]))  # J3b -- And here!
     # values from Conlon + Chalker
[…]
