### 2025-03-25 85204ae
Fix assertion test
