+++ versions.js
@@ -1,0 +1,9 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v0.7",
+  "v0.6",
+  "v0.5",
+  "dev",
+];
+var DOCUMENTER_NEWEST = "v0.7.5";
+var DOCUMENTER_STABLE = "stable";
