+++ previews/PR355/assets/documenter.js
@@ -1,0 +1,1178 @@
+// Generated by Documenter.jl
+requirejs.config({
+  paths: {
+    'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/julia.min',
+    'headroom': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/headroom.min',
+    'jqueryui': 'https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min',
+    'jquery': 'https://cdnjs.cloudflare.com/ajax/libs/jquery/3.7.0/jquery.min',
+    'headroom-jquery': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/jQuery.headroom.min',
+    'highlight': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/highlight.min',
+    'highlight-julia-repl': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/julia-repl.min',
+  },
+  shim: {
+  "highlight-julia": {
+    "deps": [
+      "highlight"
+    ]
+  },
+  "headroom-jquery": {
+    "deps": [
+      "jquery",
+      "headroom"
+    ]
+  },
+  "highlight-julia-repl": {
+    "deps": [
+      "highlight"
+    ]
+  }
+}
+});
+////////////////////////////////////////////////////////////////////////////////
+require([], function() {
+window.MathJax = {
+  "tex": {
+    "inlineMath": [
+      [
+        "$",
+        "$"
+      ]
+    ],
+    "tags": "ams"
+  },
+  "options": {
+    "ignoreHtmlClass": "tex2jax_ignore",
+    "processHtmlClass": "tex2jax_process"
+  }
+}
+;
+
+(function () {
+    var script = document.createElement('script');
+    script.src = 'https://cdnjs.cloudflare.com/ajax/libs/mathjax/3.2.2/es5/tex-svg-full.js';
+    script.async = true;
+    document.head.appendChild(script);
+})();
+
+})
+////////////////////////////////////////////////////////////////////////////////
+require(['jquery', 'highlight', 'highlight-julia', 'highlight-julia-repl'], function($) {
+$(document).ready(function() {
+    hljs.highlightAll();
+})
+
+})
+////////////////////////////////////////////////////////////////////////////////
+require(['jquery'], function($) {
+
+let timer = 0;
+var isExpanded = true;
+
+$(document).on(
+  "click",
+  ".docstring .docstring-article-toggle-button",
+  function () {
+    let articleToggleTitle = "Expand docstring";
+    const parent = $(this).parent();
+
+    debounce(() => {
+      if (parent.siblings("section").is(":visible")) {
+        parent
+          .find("a.docstring-article-toggle-button")
+          .removeClass("fa-chevron-down")
+          .addClass("fa-chevron-right");
+      } else {
+        parent
+          .find("a.docstring-article-toggle-button")
+          .removeClass("fa-chevron-right")
+          .addClass("fa-chevron-down");
+
+        articleToggleTitle = "Collapse docstring";
+      }
+
+      parent
+        .children(".docstring-article-toggle-button")
+        .prop("title", articleToggleTitle);
+      parent.siblings("section").slideToggle();
+    });
+  }
+);
+
+$(document).on("click", ".docs-article-toggle-button", function (event) {
+  let articleToggleTitle = "Expand docstring";
+  let navArticleToggleTitle = "Expand all docstrings";
+  let animationSpeed = event.noToggleAnimation ? 0 : 400;
+
+  debounce(() => {
[…]
