+++ dev/assets/documenter.js
@@ -230,6 +230,20 @@ $(document).ready(function () {
       type: "click",
       noToggleAnimation: true,
     });
+
+    setTimeout(function () {
+      if (window.location.hash) {
+        const targetId = window.location.hash.substring(1);
+        const targetElement = document.getElementById(targetId);
+
+        if (targetElement) {
+          targetElement.scrollIntoView({
+            behavior: "smooth",
+            block: "center",
+          });
+        }
+      }
+    }, 100);
   }
 });
 
@@ -638,6 +652,42 @@ function runSearchMainCode() {
   // Which filter is currently selected
   var selected_filter = "";
 
+  document.addEventListener("reset-filter", function () {
+    selected_filter = "";
+    update_search();
+  });
+
+  //update the url with search query
+  function updateSearchURL(query) {
+    const url = new URL(window.location);
+
+    if (query && query.trim() !== "") {
+      url.searchParams.set("q", query);
+    } else {
+      // remove the 'q' param if it exists
+      if (url.searchParams.has("q")) {
+        url.searchParams.delete("q");
+      }
+    }
+
+    // Add or remove the filter parameter based on selected_filter
+    if (selected_filter && selected_filter.trim() !== "") {
+      url.searchParams.set("filter", selected_filter);
+    } else {
+      // remove the 'filter' param if it exists
+      if (url.searchParams.has("filter")) {
+        url.searchParams.delete("filter");
+      }
+    }
+
+    // Only update history if there are parameters, otherwise use the base URL
+    if (url.search) {
+      window.history.replaceState({}, "", url);
+    } else {
+      window.history.replaceState({}, "", url.pathname + url.hash);
+    }
+  }
+
   $(document).on("input", ".documenter-search-input", function (event) {
     if (!worker_is_running) {
       launch_search();
@@ -647,6 +697,7 @@ function runSearchMainCode() {
   function launch_search() {
     worker_is_running = true;
     last_search_text = $(".documenter-search-input").val();
+    updateSearchURL(last_search_text);
     worker.postMessage(last_search_text);
   }
 
@@ -677,6 +728,7 @@ function runSearchMainCode() {
    */
   function update_search() {
     let querystring = $(".documenter-search-input").val();
+    updateSearchURL(querystring);
 
     if (querystring.trim()) {
       if (selected_filter == "") {
@@ -751,6 +803,24 @@ function runSearchMainCode() {
     }
   }
 
+  //url param checking
+  function checkURLForSearch() {
+    const urlParams = new URLSearchParams(window.location.search);
+    const searchQuery = urlParams.get("q");
+    const filterParam = urlParams.get("filter");
+
+    // Set the selected filter if present in URL
+    if (filterParam) {
+      selected_filter = filterParam.toLowerCase();
+    }
+
+    // Trigger input event if there's a search query to perform the search
+    if (searchQuery) {
+      $(".documenter-search-input").val(searchQuery).trigger("input");
+    }
+  }
+  setTimeout(checkURLForSearch, 100);
+
   /**
    * Make the modal filter html
    *
@@ -821,8 +891,9 @@ $(document).ready(function () {
       <div class="field mb-0 w-100">
         <p class="control has-icons-right">
           <input class="input documenter-search-input" type="text" placeholder="Search" />
-          <span class="icon is-small is-right has-text-primary-dark">
-            <i class="fas fa-magnifying-glass"></i>
+          <span class="icon is-small is-right has-text-primary-dark gap-2">
+            <i class="fas fa-link link-icon is-clickable"></i>
+            <i class="fas fa-magnifying-glass mr-4"></i>
           </span>
         </p>
       </div>
[…]
