### 2025-03-28 db08a10
Update conventions, compacity tests
