+++ test/test_scga.jl
@@ -1,6 +1,7 @@
 @testitem "diamond_lattice" begin
     # test against JuliaSCGA (S. Gao)
-    dia = [0.7046602277469309, 0.8230846832863896, 0.23309034250417973, 0.40975668535137943, 0.8474163786642979, 0.8230846832694241, 0.723491683211756, 0.5939752161027589, 0.6506966347286152, 0.8012263819500781, 0.23309034265153963, 0.5939752161512792, 0.7779185770415442, 0.9619923476121188, 0.28363795492206234, 0.4097566857133061, 0.6506966347914551, 0.9619923474164132, 0.8576708385848646, 0.45534457001475764, 0.8474163779976567, 0.8012263817424181, 0.2836379554342603, 0.4553445702621035, 0.9352400940660723]
+    res_jscga = [0.7046602277469309, 0.8230846832863896, 0.23309034250417973, 0.40975668535137943, 0.8474163786642979,
+                 0.8230846832694241, 0.723491683211756, 0.5939752161027589, 0.6506966347286152, 0.8012263819500781]
     a = 8.5031 # (Å)
     latvecs = lattice_vectors(a, a, a, 90, 90, 90)
     cryst = Crystal(latvecs, [[0, 0, 0]], 227; choice="1")
@@ -11,70 +12,65 @@
     set_exchange!(sys, 1/(2s)^2, Bond(1, 2, [0, 0, 0]))
     measure = ssf_perp(sys)
     kT = 15*meV_per_K
-    scga = Sunny.SCGA(sys; measure, kT, dq=1/8)
+    scga = StaticCorrelationsSCGA(sys; measure, kT, dq=1/8)
     γ = s^2 * Sunny.natoms(cryst)
-    grid = q_space_grid(cryst, [1, 0, 0], 0:0.9:4, [0, 1, 0], 0:0.9:4; orthogonalize=true)
+    grid = q_space_grid(cryst, [1, 0, 0], range(0, 3.6, 5), [0, 1, 0], range(0, 0.9, 2))
     res = intensities_static(scga, grid)
-    @test isapprox(vec(res.data)/γ, dia; rtol=1e-8)
+    @test isapprox(vec(res.data)/γ, res_jscga; rtol=1e-8)
 end
 
 @testitem "square_lattice" begin
     # test against JuliaSCGA (S. Gao)
-    sq = [0.3578208506624103, 0.3850668587212228, 0.4211633125595451, 0.3930558742921638, 0.3586715762816389, 0.3775147329089877, 0.4188431376589759, 0.4009835864744404, 0.36119385315852837, 0.3850668587212228, 0.4063051066815142, 0.4182331099724885,
-          0.36751592492720886, 0.32829927921801794, 0.3490454303292574, 0.4062747596231383, 0.41626772462487194, 0.38789161791308147, 0.4211633125595451, 0.4182331099724885, 0.3710074695764304, 0.2924827325221696, 0.253584799014828, 0.2732638980307635,
-          0.34406882609120737, 0.409719043169102, 0.4213848627804445, 0.3930558742921638, 0.36751592492720886, 0.2924827325221696, 0.21946942640699554, 0.18901809131848027, 0.2041521166894061, 0.26469888441402106, 0.3466451094806684, 0.3903985131607334,
[…]
