+++ joss/paper.bib
@@ -12,6 +12,7 @@
   volume={3},
   number={24},
   year={2018},
+  doi={10.21105/joss.00615},
   publisher={Open Journals}
 }
 @article{wang:2001,
@@ -22,6 +23,7 @@
   number={10},
   pages={2050},
   year={2001},
+  doi={10.1103/PhysRevLett.86.2050},
   publisher={APS}
 }
 
@@ -33,6 +35,7 @@
   number={21},
   pages={2607},
   year={1986},
+  doi={10.1103/PhysRevLett.57.2607},
   publisher={APS}
 }
 
@@ -43,6 +46,7 @@
   volume={6},
   number={65},
   pages={3349},
+  doi={10.21105/joss.03349},
   year={2021}
 }
 
@@ -57,6 +61,7 @@
   number={10},
   pages={104409},
   year={2021},
+  doi={10.1103/PhysRevB.104.104409},
   publisher={APS}
 }
 
@@ -68,6 +73,7 @@
   number={5},
   pages={054423},
   year={2022},
+  doi={10.1103/PhysRevB.106.054423},
   publisher={APS}
 }
 
@@ -79,6 +85,7 @@
   number={23},
   pages={235154},
   year={2022},
+  doi={10.1103/PhysRevB.106.235154},
   publisher={APS}
 }
 
@@ -86,6 +93,7 @@
   title={Renormalized classical theory of quantum magnets},
   author={{Dahlbom}, D.~A. and {Zhang}, H. and {Laraib}, Z. and {Pajerowski}, D.~M. and {Barros}, K. and {Batista}, C.~D.},
   journal={arXiv preprint arXiv:2304.03874},
+  doi={10.48550/arXiv.2304.03874},
   year={2023}
 }
 
@@ -97,6 +105,7 @@
   number={1},
   pages={014427},
   year={2024},
+  doi={10.1103/PhysRevB.109.014427},
   publisher={APS}
 }
 
@@ -108,6 +117,7 @@
   number={10},
   pages={104403},
   year={2024},
+  doi={10.1103/PhysRevB.110.104403},
   publisher={APS}
 }
 
@@ -118,6 +128,7 @@
   volume={17},
   number={5},
   pages={145},
+  doi={10.21468/SciPostPhys.17.5.145},
   year={2024}
 }
 
@@ -130,6 +141,7 @@
   number={8},
   pages={083I01},
   year={2014},
+  doi={10.1093/ptep/ptu109},
   publisher={Oxford University Press}
 }
 
@@ -141,6 +153,7 @@
   number={26},
   pages={267201},
   year={2021},
+  doi={10.1103/PhysRevLett.127.267201},
   publisher={APS}
 }
 
@@ -155,6 +168,7 @@
   number={4},
   pages={467--472},
   year={2021},
+  doi={10.1038/s41567-020-01110-1},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -166,6 +180,7 @@
   number={14},
   pages={144427},
   year={2023},
+  doi={10.1103/PhysRevB.107.144427},
   publisher={APS}
 }
 
@@ -177,6 +192,7 @@
   number={1},
   pages={3626},
   year={2023},
+  doi={10.1038/s41467-023-39232-8},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -188,6 +204,7 @@
   number={1},
   pages={5},
   year={2023},
+  doi={10.1038/s41535-022-00526-7},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -199,6 +216,7 @@
   number={1},
   pages={4199},
   year={2023},
+  doi={10.1038/s41467-023-39940-1},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -210,6 +228,7 @@
[…]
