### 2025-04-01 #7684 @ddahlbom
@phibeck Thanks for taking this on! I updated the references to include DOIs. (The only entries without DOIs are repo references.)
