### 2025-04-02 #7684 @phibeck
> Sunny.jl is amazing software package! Unfortunately, I do not feel qualified to serve as a sufficiently impartial reviewer of this work owing to my current and ongoing close working relationship with many of the authors.

Okay, thanks for letting me know!
