### 2025-04-02 #7684 @phibeck
> [@phibeck](https://github.com/phibeck) Thanks for taking this on! I updated the references to include DOIs. (The only entries without DOIs are repo references.)

Looks good, thank you.
