### 2025-04-07 #7684 @PhilipFackler
@phibeck it seems I have a COI since I work at the same institution and on the same project as @ddahlbom (though not on Sunny.jl myself).
