### 2025-04-08 #7684 @jabirali
@phibeck I'd be happy to review this project, it looks quite interesting. However, I might not be able to start the process until ~1 month from now due to the upcoming Easter vacation followed by final exams. Does that timeline sound OK?
