### 2025-04-11 61c05f4
Explicitly normalize axis in SpinWaveTheorySpiral
