+++ docs/src/versions.md
@@ -11,6 +11,7 @@
   higher accuracy.
 * Fix atom indexing when setting interactions for a reshaped system ([PR
   #359](https://github.com/SunnySuite/Sunny.jl/pull/359)).
+* Normalize `axis` argument to `SpinWaveTheorySpiral` for correctness.
 
 ## v0.7.5
 (Jan 20, 2025)
