### 2025-04-15 712c1ea
Replace @test_warn with @test_logs
