### 2025-04-15 770d6a1
Fix thermal prefactors in SpinWaveTheory
