### 2025-04-15 52bd9c1
Fix thermal prefactors in SpinWaveTheory
