+++ test/test_spiral.jl
@@ -206,7 +206,6 @@ end
     cryst = Crystal(latvecs, [[0, 0, 0], [0, 0.1, 0]]; types=["a", "b"])
     cryst = subcrystal(cryst, "a")
     sys = System(cryst, [1 => Moment(s=1, g=-1)], :dipole)
-    print_bond(cryst, Bond(1, 1, [1, 0, 0]))
 
     D = 1
     axis = [1, 0, 0]
@@ -233,9 +232,9 @@ end
 
     qs = [[0, 0, -1/2], [0, 0, 1/3]]
     swt = SpinWaveTheorySpiral(sys; measure=ssf_trace(sys; apply_g=false), k, axis)
-    res = intensities_bands(swt, qs)
+    res = intensities_bands(swt, qs, kT=D/10)
     swt2 = SpinWaveTheory(sys2; measure=ssf_trace(sys2; apply_g=false))
-    res2 = intensities_bands(swt2, qs)
+    res2 = intensities_bands(swt2, qs, kT=D/10)
 
     @test res.disp[1:3, 1] ≈ res2.disp[1:3, 1]
     @test res.disp[1:3, 2] ≈ res2.disp[[1, 3, 4], 2]
