### 2025-04-18 8882a2b
Slightly more accurate thermal_prefactor
