### 2025-04-18 00e52a6
Include check on intensities with kT
