+++ test/test_lswt.jl
@@ -45,11 +45,11 @@
     qs = [[0.24331089495721447, 0.2818361515716459, 0.21954858411037714]]
     swt = SpinWaveTheory(sys; measure=ssf_perp(sys; apply_g=false))
 
-    res = intensities_bands(swt, qs)
+    res = intensities_bands(swt, qs; kT=100.0)
     # println(round.(vec(res.disp); digits=12))
     # println(round.(vec(res.data); digits=12))
     disps_golden = [1394.440092579925, 1393.728009951747, 1393.008551251224, 1392.919524974098, 1279.239919068637, 1279.094568472202, 1278.224518515366, 1277.69176148292, 1194.366336255056, 1193.750083625346, 1191.583519659758, 1189.794451340792, 1131.422439587906, 1131.202770074575, 1065.242927850642, 1065.09589244662, 1026.649340922954, 1024.028348558092, 1022.830406299696, 1020.767349639404, 945.202397530689, 944.795817851584, 835.545028394171, 832.001588695236, 827.939501409146, 827.307586947863, 821.216582166429, 820.430993567084, 820.29454877608, 818.594570998006, 810.207001090189, 808.553158273681, 766.524411070996, 766.51610275022, 766.513825852464, 766.508655558188, 758.579854167682, 754.683765885699, 750.572578891218, 750.471006252543, 665.954573008179, 662.421047653194, 651.465562550036, 651.417940124412, 581.258189152574, 568.105209800088, 559.053702296449, 558.493005822973, 552.043762736839, 550.131096070953, 539.733572947827, 530.698033192904, 499.661483510112, 494.928560823174, 435.233706061902, 427.70227706432, 408.128705853668, 399.856401749667, 370.0693430633, 369.845327686247, 365.04951424025, 363.639416669404, 354.648012591371, 346.609483926993, 341.989165167298, 339.373361067981, 318.363717384318, 276.219249203163, 263.161053831818, 257.409506246762, 230.539454193868, 229.778324172696, 203.971681278995, 197.504237153905, 193.879371534689, 189.866421874996, 189.815806967662, 167.944134431612, 154.923566498395, 146.219538847758] 
-    data_golden = [0.00038665026, 0.0, 0.007231537047, 0.0, 0.008665001888, 0.0, 0.015340530772, 0.0, 0.0, 0.054200267869, 0.073309637341, 0.0, 0.005276014091, 0.0, 0.0, 0.026708463804, 0.0, 0.062332773978, 0.112027719874, 0.0, 0.031129659091, 0.0, 0.115569104228, 0.0, 0.0, 0.038994974901, 0.0, 0.031152494381, 0.029113142876, 0.0, 0.0, 0.00467095768, 0.000762927464, 0.000940556177, 0.0, 0.0, 0.00877327198, 0.0, 0.033263081608, 0.0, 0.066082588247, 0.0, 0.0, 0.059956690592, 0.334527718881, 0.007608761425, 0.0, 0.0, 0.0, 0.068446181755, 0.029366565904, 0.0, 0.0, 0.605679689903, 0.378667970539, 0.0, 0.858558512424, 0.0, 0.0, 0.290056738763, 0.349117318828, 0.0, 0.0, 0.955658357373, 0.0, 1.606184739758, 0.0, 0.201711065755, 0.181786511323, 0.0, 0.0, 0.313711125446, 0.769401980923, 0.0, 0.010377681107, 0.0, 0.151200728664, 0.327686235744, 0.0, 0.0] 
[…]
