+++ test/test_pkg_analysis.jl
@@ -19,7 +19,7 @@ end
 
 @testitem "ExplicitImports" begin
     import ExplicitImports, LinearAlgebra
-    try        
+    try
         ExplicitImports.check_no_implicit_imports(Sunny; skip=(mod, Base, Core, LinearAlgebra))
     catch _
         @test false
