### 2025-04-18 0ee11f6
Miscellaneous cleanups (#372)
