### 2025-04-18 c7aee62
Adapt test tolerances to true Float64 precision (#360)
This fixes tests on Intel machines and on Mac with Julia 1.12.
