+++ test/test_entangled_units.jl
@@ -73,9 +73,9 @@ end
 
     set_field!(esys, [0, 0, 0])
     randomize_spins!(esys)
-    minimize_energy!(esys; g_abstol=1e-14)
-    @test norm(esys.sys_origin.dipoles[1]) < 1e-14
-    @test norm(esys.sys_origin.dipoles[2]) < 1e-14
+    minimize_energy!(esys)
+    @test norm(esys.sys_origin.dipoles[1]) < 1e-10
+    @test norm(esys.sys_origin.dipoles[2]) < 1e-10
 
     # Test inter-bond exchange
     pc = Sunny.as_general_pair_coupling(interactions.pair[1], esys.sys)
