### 2025-04-18 709e005
Update conventions, compacity tests
