### 2025-04-18 13ce1d9
Streamline interface
