+++ src/SCGA/SCGA.jl
@@ -1,38 +1,30 @@
-abstract type AbstractSCGA end
-
-struct SCGA <: AbstractSCGA
-    sys            :: System
-    measure        :: MeasureSpec
-    regularization :: Float64
-end
-
 """
     SCGA(sys::System; measure, regularization=1e-8)
 
-Constructs an object to perform the self consistent gaussian approximation. Enables 
-the use of `intensities_static`](@ref) with an SCGA type to calculate the static 
-structure factor in the paramagnetic phase. If the temperature is below the ordering
-temperature, the intensities will be negative.
+Constructs an object to perform the self consistent gaussian approximation.
+Enables the use of `intensities_static`](@ref) with an SCGA type to calculate
+the static structure factor in the paramagnetic phase. If the temperature is
+below the ordering temperature, the intensities will be negative.
 """
+struct SCGA
+    sys :: System
+    measure :: MeasureSpec
+    regularization :: Float64
 
-function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
-    measure = @something measure empty_measurespec(sys)
-    if length(eachsite(sys)) != prod(size(measure.observables)[2:5])
-        error("Size mismatch. Check that measure is built using consistent system.")
+    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
+        measure = @something measure empty_measurespec(sys)
+        if length(eachsite(sys)) != prod(size(measure.observables)[2:5])
+            error("Size mismatch. Check that measure is built using consistent system.")
+        end
+        return new(sys, measure, regularization)
     end
-    return SCGA(sys,measure, regularization)
 end
 
-"""
-    fourier_transform_interaction_matrix(sys::System; k, ϵ=0)
 
-Fourier transforms the interaaction matrix for a System and evaluatues it at q=k. A 
-small regularization, ϵ, can be added. This function is adapted from 
-`luttinger_tisza_exchange`](@ref) to include the correct prefactors and local anisotropies.
-
-"""
-
-function fourier_transform_interaction_matrix(sys::System; k, ϵ=0)
+# Fourier transforms the interaaction matrix for a System and evaluatues it at
+# q=k. This function is adapted from `luttinger_tisza_exchange`](@ref) to
+# include the correct prefactors and local anisotropies.
+function fourier_transform_interaction_matrix(sys::System; k)
     @assert sys.mode in (:dipole, :dipole_large_S) "SU(N) mode not supported"
     @assert sys.dims == (1, 1, 1) "System must have only a single cell"
     Na = natoms(sys.crystal)
@@ -45,45 +37,40 @@ function fourier_transform_interaction_matrix(sys::System; k, ϵ=0)
 
             (; j, n) = bond
             J = exp(2π * im * dot(k, n+sys.crystal.positions[j]-sys.crystal.positions[i])) * Mat3(bilin*I)
-            J_k[:, i, :, j] += J / 2
-            J_k[:, j, :, i] += J' / 2
+            J_k[:, i, :, j] += J
+            J_k[:, j, :, i] += J'
         end
     end
 
     if !isnothing(sys.ewald)
[…]
