### 2025-04-18 8f09fcd
Fix anisotropy strength
