+++ test/test_scga.jl
@@ -49,29 +49,24 @@ end
 @testitem "MgCr2O4" begin
     # test against Conlon and Chalker
     latvecs    = lattice_vectors(8.3342, 8.3342, 8.3342, 90, 90, 90)
-    positions  = [[0.1250, 0.1250, 0.1250],
-                [0.5000, 0.5000, 0.5000],
-                [0.2607, 0.2607, 0.2607]]
-    types      = ["Mg","Cr","O"]
-    spacegroup = 227 # Space Group Number
-    xtal_mgcro = Crystal(latvecs, positions, spacegroup; types)
-    cryst = subcrystal(xtal_mgcro,"Cr")
-    sys = System(cryst, [1 => Moment(; s=3/2, g=1)], :dipole) # Same on MgCr2O4 crystal
-    J1      = 3.27  # value of J1 in meV from Bai's PRL paper
-    J_mgcro = [1.00,0.0815,0.1050,0.0085]*J1; # further neighbor pyrochlore relevant for MgCr2O4
+    positions  = [[1/2, 1/2, 1/2]]
+    cryst = Crystal(latvecs, positions, 227)
+    sys = System(cryst, [1 => Moment(; s=3/2, g=1)], :dipole)
+    J1 = 3.27                                            # in meV, from Bai's PRL
+    J_mgcro = [1.0, 0.0815, 0.1050, 0.0085] * J1;        # further neighbor pyrochlore relevant for MgCr2O4
     set_exchange!(sys, J_mgcro[1], Bond(1, 2, [0,0,0]))  # J1
     set_exchange!(sys, J_mgcro[2], Bond(1, 7, [0,0,0]))  # J2
     set_exchange!(sys, J_mgcro[3], Bond(1, 3, [1,0,0]))  # J3a -- Careful here!
     set_exchange!(sys, J_mgcro[4], Bond(1, 3, [0,0,0])); # J3b -- And here!
     # values from Conlon + Chalker
[…]
