### 2025-04-18 df495e9
Fix reshaping of q
