### 2025-04-18 d9f073a
Negative intensities seem mathematically impossible
