### 2025-04-18 d4e6b21
Newton method working
