+++ src/SCGA/SCGA.jl
@@ -42,51 +42,6 @@ struct SCGA
 end
 
 
-# Calculate the Fourier transformed interaction matrix for wavevector k. Adapted
-# from `luttinger_tisza_exchange`.
-function fourier_transform_interaction_matrix(sys::System; k)
-    @assert sys.mode in (:dipole, :dipole_uncorrected) "SU(N) mode not supported"
-    @assert sys.dims == (1, 1, 1) "System must have only a single cell"
-    Na = natoms(sys.crystal)
-    J_k = zeros(ComplexF64, 3, Na, 3, Na)
-    r = sys.crystal.positions
-
-    for i in 1:Na
-        for coupling in sys.interactions_union[i].pair
-            (; isculled, bond, bilin) = coupling
-            isculled && break
-
-            (; j, n) = bond
-            J = exp(2π * im * dot(k, n + r[j] - r[i])) * Mat3(bilin*I)
-            J_k[:, i, :, j] += J
-            J_k[:, j, :, i] += J'
-        end
-    end
-
-    if !isnothing(sys.ewald)
-        A = precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), k) * sys.ewald.μ0_μB² # FIXME: check this is in the right coordinate system
-        A = reshape(A, Na, Na)
-        for i in 1:Na, j in 1:Na
-            J_k[:, i, :, j] += sys.gs[i]' * A[i, j] * sys.gs[j] # FIXME: Need a test
-        end
-    end
-
-    for i in 1:Na
-        onsite_coupling = sys.interactions_union[i].onsite
-        (; c2, c4, c6) = onsite_coupling
-        iszero(c4) && iszero(c6) || error("Single-ion anisotropy beyond quadratic order not supported")
-        anisotropy = SA[c2[1]-c2[3]        c2[5] 0.5c2[2];
-                              c2[5] -c2[1]-c2[3] 0.5c2[4];
-                           0.5c2[2]     0.5c2[4]   2c2[3]]
-        J_k[:, i, :, i] += 2 * anisotropy
-    end
-
-    J_k = reshape(J_k, 3*Na, 3*Na)
-    @assert diffnorm2(J_k, J_k') < 1e-15
-    J_k = hermitianpart(J_k)
-    return J_k
-end
-
 # Computes the Lagrange multiplier for the standard SCGA approach with a common
 # Lagrange multiplier for all sublattices.
 function find_lagrange_multiplier(scga::SCGA, β)
@@ -115,7 +70,7 @@ function find_lagrange_multiplier(scga::SCGA, β)
         end
     end
     q = [[qx, qy, qz] for qx in qarrays[1], qy in qarrays[2], qz in qarrays[3]]
-    Jq_array = [fourier_transform_interaction_matrix(sys; k=q_in) for q_in in q]
+    Jq_array = [fourier_exchange_matrix(sys; k=q_in) for q_in in q]
     #TODO throw error if spins different
     if quantum_sum_rule
         S_sq = sum(sys.κs .* (sys.κs.+1)) / Na
@@ -162,26 +117,24 @@ function intensities_static_single(scga::SCGA, qpts; β)
     λ = find_lagrange_multiplier(scga, β)
     intensity = zeros(eltype(measure),length(qpts.qs))
     Ncorr = length(measure.corr_pairs)
+    r = sys.crystal.positions
+
     for (iq, q) in enumerate(qpts.qs)
         pref = zeros(ComplexF64, Nobs, Na)
         corrbuf = zeros(ComplexF64, Ncorr)
         intensitybuf = zeros(eltype(measure), 3, 3)
         q_reshaped = to_reshaped_rlu(sys, q)
[…]
