### 2025-04-18 728db6f
Adapt to 1/dq notation
