+++ src/SCGA/SCGA.jl
@@ -1,5 +1,5 @@
 """
-    SCGA(sys::System; measure, regularization=1e-8)
+    SCGA(sys::System; measure, quantum_sum_rule=true)
 
 Constructs an object to perform the self consistent gaussian approximation.
 Enables the use of `intensities_static`](@ref) with an SCGA type to calculate
@@ -9,14 +9,14 @@ below the ordering temperature, the intensities will be negative.
 struct SCGA
     sys :: System
     measure :: MeasureSpec
-    regularization :: Float64
+    quantum_sum_rule :: Bool
 
-    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
+    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, quantum_sum_rule=true)
         measure = @something measure empty_measurespec(sys)
         if length(eachsite(sys)) != prod(size(measure.observables)[2:5])
             error("Size mismatch. Check that measure is built using consistent system.")
         end
-        return new(sys, measure, regularization)
+        return new(sys, measure, quantum_sum_rule)
     end
 end
 
@@ -71,7 +71,8 @@ end
 # and "Newton's Method" which is a fast converging root finding algorithm.
 # SumRule specifices the sum rule that the Lagrange multipliers are chosen to
 # satisfy, either the Classical or Quantum sum rules.
-function find_lagrange_multiplier(sys::System,kT; ϵ=0, SumRule = "Classical",starting_offset = 0.2, maxiters=1_000,tol = 1e-10,method = "Nelder Mead",Nq = 20)
+function find_lagrange_multiplier(scga::SCGA, kT; starting_offset=0.2, maxiters=1_000, tol=1e-10, Nq=20)
+    (; sys, quantum_sum_rule) = scga
     dq = 1/Nq;
     Na = natoms(sys.crystal)
     bond_counter = zeros(Float64,3)
@@ -95,67 +96,50 @@ function find_lagrange_multiplier(sys::System,kT; ϵ=0, SumRule = "Classical",st
     q = [[qx, qy, qz] for qx in qarrays[1], qy in qarrays[2], qz in qarrays[3]]
     Jq_array = [fourier_transform_interaction_matrix(sys; k=q_in) for q_in in q]
     #TODO throw error if spins different
-    if SumRule == "Classical"
-        S_sq = sum(sys.κs.^2)/Na
-    elseif SumRule == "Quantum"
+    if quantum_sum_rule
         S_sq = sum(sys.κs .*(sys.κs.+1))/Na
     else
-        error("Unsupported SumRule: $SumRule. Expected 'Classical' or 'Quantum'.")
+        S_sq = sum(sys.κs.^2)/Na
     end
-    println("Sum rule: $SumRule")
+
     eig_vals = zeros(3Na, length(Jq_array))
     for j in 1:length(Jq_array)
          eig_vals[:,j] .= eigvals(Jq_array[j])
     end
-    if method == "Nelder Mead"
-        function loss(λ)
-            sum_term = sum(1 ./ (λ .+ (1/(kT)).*eig_vals ))
-            return ((1/(Na*length(q))) * sum_term - S_sq)^2
-        end
-        lower = -minimum(eig_vals)/kT
-        upper = Inf
-        p = [lower+starting_offset] # some offset from the lower bound.
-        options = Optim.Options(; iterations=maxiters, show_trace=false,g_tol=tol)
-        result = optimize(loss, lower, upper, p,NelderMead(), options)
-        min = Optim.minimizer(result)[1]
[…]
