+++ test/test_scga.jl
@@ -4,19 +4,19 @@
     dia = [0.7046602277469309, 0.8230846832863896, 0.23309034250417973, 0.40975668535137943, 0.8474163786642979, 0.8230846832694241, 0.723491683211756, 0.5939752161027589, 0.6506966347286152, 0.8012263819500781, 0.23309034265153963, 0.5939752161512792, 0.7779185770415442, 0.9619923476121188, 0.28363795492206234, 0.4097566857133061, 0.6506966347914551, 0.9619923474164132, 0.8576708385848646, 0.45534457001475764, 0.8474163779976567, 0.8012263817424181, 0.2836379554342603, 0.4553445702621035, 0.9352400940660723]
     a = 8.5031 # (Å)
     latvecs = lattice_vectors(a, a, a, 90, 90, 90)
-    cryst = Crystal(latvecs, [[0,0,0]], 227, setting="1")
+    cryst = Crystal(latvecs, [[0,0,0]], 227; choice="1")
     dims = (1, 1, 1)
     spininfos = [1 => Moment(; s=1, g=1)]
     sys = System(cryst, spininfos, :dipole;dims, seed=0)
     set_exchange!(sys, -1, Bond(1, 3, [0,0,0]))
     set_exchange!(sys, 0.25, Bond(1, 2, [0,0,0]))
     measure = ssf_perp(sys;)
-    scga = Sunny.SCGA(sys;measure)
+    scga = Sunny.SCGA(sys; measure, quantum_sum_rule=false)
     kT = 15*meV_per_K
     S=1
     γ=S*(S+1)*length(cryst.positions)
     grid = q_space_grid(cryst, [1, 0, 0], 0:0.9:4, [0, 1, 0], 0:0.9:4; orthogonalize=true)
-    res = Sunny.intensities_static(scga, grid; kT, SumRule="Classical",sublattice_resolved=false,Nq = 8)
+    res = Sunny.intensities_static(scga, grid; kT, sublattice_resolved=false, Nq=8)
     # Gao had 2 basis sites
     @test abs(sum(reshape(res.data,size(dia))/γ -dia/2))/length(grid.qs)< tol
 end
@@ -41,12 +41,12 @@ end
     set_exchange!(sys, 0.5, Bond(1, 1, [1,1,0]))
     set_exchange!(sys, 0.25, Bond(1, 1, [2,0,0]))
     measure = ssf_perp(sys)
-    scga = Sunny.SCGA(sys;measure, regularization=1e-8)
+    scga = Sunny.SCGA(sys; measure, quantum_sum_rule=false)
     kT = 27.5*meV_per_K
     S=1
     γ=S*(S+1)
     grid = q_space_grid(cryst, [1, 0, 0], -1:0.12:0, [0, 1, 0], -1:0.12:0; orthogonalize=true)
-    res = Sunny.intensities_static(scga, grid; kT, SumRule="Classical",Nq = 8,sublattice_resolved = false)
+    res = Sunny.intensities_static(scga, grid; kT, Nq=8, sublattice_resolved=false)
     @test abs(sum(reshape(res.data,size(sq))/γ-sq))/length(grid.qs) <tol
 end
 
@@ -61,8 +61,7 @@ end
                 [0.2607, 0.2607, 0.2607]]
     types      = ["Mg","Cr","O"]
     spacegroup = 227 # Space Group Number
-    setting    = "2" # Space Group setting
-    xtal_mgcro = Crystal(latvecs, positions, spacegroup; types, setting)
+    xtal_mgcro = Crystal(latvecs, positions, spacegroup; types)
     cryst = subcrystal(xtal_mgcro,"Cr")
[…]
