+++ src/SCGA/SCGA.jl
@@ -1,0 +1,575 @@
+abstract type AbstractSCGA end
+
+struct SCGA <: AbstractSCGA
+    sys            :: System
+    measure        :: MeasureSpec
+    regularization :: Float64
+end
+
+"""
+    SCGA(sys::System; measure, regularization=1e-8)
+
+Constructs an object to perform the self consistent gaussian approximation. Enables 
+the use of `intensities_static`](@ref) with an SCGA type to calculate the static 
+structure factor in the paramagnetic phase. If the temperature is below the ordering
+temperature, the intensities will be negative.
+"""
+
+function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
+    measure = @something measure empty_measurespec(sys)
+    if length(eachsite(sys)) != prod(size(measure.observables)[2:5])
+        error("Size mismatch. Check that measure is built using consistent system.")
+    end
+    return SCGA(sys,measure, regularization)
+end
+
+"""
+    fourier_transform_interaction_matrix(sys::System; k, ϵ=0)
+
+Fourier transforms the interaaction matrix for a System and evaluatues it at q=k. A 
+small regularization, ϵ, can be added. This function is adapted from 
+`luttinger_tisza_exchange`](@ref) to include the correct prefactors and local anisotropies.
+
+"""
+
+function fourier_transform_interaction_matrix(sys::System; k, ϵ=0)
+    @assert sys.mode in (:dipole, :dipole_large_S) "SU(N) mode not supported"
+    @assert sys.dims == (1, 1, 1) "System must have only a single cell"
+    Na = natoms(sys.crystal)
+    J_k = zeros(ComplexF64, 3, Na, 3, Na)
+
+    for i in 1:Na
+        for coupling in sys.interactions_union[i].pair
+            (; isculled, bond, bilin) = coupling
+            isculled && break
+
+            (; j, n) = bond
+            J = exp(2π * im * dot(k, n+sys.crystal.positions[j]-sys.crystal.positions[i])) * Mat3(bilin*I)
+            J_k[:, i, :, j] += J / 2
+            J_k[:, j, :, i] += J' / 2
+        end
+    end
+
+    if !isnothing(sys.ewald)
+        A = precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), k) * sys.ewald.μ0_μB² # check this is in the right coordinate system
+        A = reshape(A, Na, Na)
+        for i in 1:Na, j in 1:Na
+            J_k[:, i, :, j] += sys.gs[i]' * A[i, j] * sys.gs[j] / 2
+        end
+    end
+    J_k = 2J_k # I'm not so sure why we need this
+    for i in 1:Na
+        onsite_coupling = sys.interactions_union[i].onsite
+        (; c2, c4, c6) = onsite_coupling
+        anisotropy = [c2[1]-c2[3]   c2[5]   0.5c2[2];
+                        c2[5]   -c2[1]-c2[3]  0.5c2[4];
+                        0.5c2[2]    0.5c2[4]    2c2[3]]
+        J_k[:, i, :, i] += 2*anisotropy
+    end
+
+    J_k = reshape(J_k, 3*Na, 3*Na)
+    @assert diffnorm2(J_k, J_k') < 1e-15
+    J_k = hermitianpart(J_k)
+    return J_k 
+end
+
+"""
+    find_lagrange_multiplier(sys::System,kT; ϵ=0, SumRule = "Classical",starting_offset = 0.2, maxiters=1_000,tol = 1e-10,method = "Nelder Mead", Nq = 20)
+
[…]
