+++ test/test_scga.jl
@@ -1,0 +1,213 @@
+@testitem "diamond_lattice" begin
+    # test against JuliaSCGA (S. Gao)
+    tol = 1e-8
+    dia = [0.7046602277469309, 0.8230846832863896, 0.23309034250417973, 0.40975668535137943, 0.8474163786642979, 0.8230846832694241, 0.723491683211756, 0.5939752161027589, 0.6506966347286152, 0.8012263819500781, 0.23309034265153963, 0.5939752161512792, 0.7779185770415442, 0.9619923476121188, 0.28363795492206234, 0.4097566857133061, 0.6506966347914551, 0.9619923474164132, 0.8576708385848646, 0.45534457001475764, 0.8474163779976567, 0.8012263817424181, 0.2836379554342603, 0.4553445702621035, 0.9352400940660723]
+    a = 8.5031 # (Å)
+    latvecs = lattice_vectors(a, a, a, 90, 90, 90)
+    cryst = Crystal(latvecs, [[0,0,0]], 227, setting="1")
+    dims = (1, 1, 1)
+    spininfos = [1 => Moment(; s=1, g=1)]
+    sys = System(cryst, spininfos, :dipole;dims, seed=0)
+    set_exchange!(sys, -1, Bond(1, 3, [0,0,0]))
+    set_exchange!(sys, 0.25, Bond(1, 2, [0,0,0]))
+    measure = ssf_perp(sys;)
+    scga = Sunny.SCGA(sys;measure)
+    kT = 15*meV_per_K
+    S=1
+    γ=S*(S+1)*length(cryst.positions)
+    grid = q_space_grid(cryst, [1, 0, 0], 0:0.9:4, [0, 1, 0], 0:0.9:4; orthogonalize=true)
+    res = Sunny.intensities_static(scga, grid; kT, SumRule="Classical",sublattice_resolved=false,Nq = 8)
+    # Gao had 2 basis sites
+    @test abs(sum(reshape(res.data,size(dia))/γ -dia/2))/length(grid.qs)< tol
+end
+
+@testitem "square_lattice" begin
+    # test against JuliaSCGA (S. Gao)
+    tol = 1e-5
+    sq=[0.3578208506624103, 0.3850668587212228, 0.4211633125595451, 0.3930558742921638, 0.3586715762816389, 0.3775147329089877, 0.4188431376589759, 0.4009835864744404, 0.36119385315852837, 0.3850668587212228, 0.4063051066815142, 0.4182331099724885, 
+    0.36751592492720886, 0.32829927921801794, 0.3490454303292574, 0.4062747596231383, 0.41626772462487194, 0.38789161791308147, 0.4211633125595451, 0.4182331099724885, 0.3710074695764304, 0.2924827325221696, 0.253584799014828, 0.2732638980307635, 
+    0.34406882609120737, 0.409719043169102, 0.4213848627804445, 0.3930558742921638, 0.36751592492720886, 0.2924827325221696, 0.21946942640699554, 0.18901809131848027, 0.2041521166894061, 0.26469888441402106, 0.3466451094806684, 0.3903985131607334, 
+    0.3586715762816389, 0.32829927921801794, 0.25358479901482806, 0.18901809131848027, 0.16311271576100106, 0.17594081183630647, 0.2284557904965832, 0.3060310966855855, 0.35526653426084637, 0.3775147329089877, 0.3490454303292574, 0.2732638980307635, 
[…]
