### 2025-04-18 bf7f788
kT to beta
