+++ src/SCGA/SCGA.jl
@@ -93,7 +93,7 @@ end
 # and "Newton's Method" which is a fast converging root finding algorithm.
 # SumRule specifices the sum rule that the Lagrange multipliers are chosen to
 # satisfy, either the Classical or Quantum sum rules.
-function find_lagrange_multiplier(scga::SCGA, kT)
+function find_lagrange_multiplier(scga::SCGA, β)
     starting_offset = 0.2
     maxiters = 500
     tol = 1e-10
@@ -133,15 +133,15 @@ function find_lagrange_multiplier(scga::SCGA, kT)
     end
 
     function f(λ)
-        sum_term = sum(1 ./ (λ .+ (1/(kT)).*eig_vals ))
+        sum_term = sum(1 ./ (λ .+ β * eig_vals ))
         return (1/(Na*length(q))) * sum_term
     end
     function J(λ)
-        sum_term = sum((1 ./ (λ .+ (1/(kT)).*eig_vals )).^2)
+        sum_term = sum((1 ./ (λ .+ β * eig_vals)).^2)
         return -(1/(Na*length(q))) * sum_term
     end
 
-    lower = -minimum(eig_vals)/kT
+    lower = - β * minimum(eig_vals)
     λn = starting_offset*0.1 + lower # Make more robust - regularized! Check optim.jl
 
     for n in 1:maxiters
@@ -158,13 +158,12 @@ end
 # Computes the static structure factor in the standard SCGA approach, with a
 # single Lagrange multiplier, over the qpts specified.
 
-function intensities_static_single(scga::SCGA, qpts; kT=0.0)
-    kT == 0.0 && error("kT must be non-zero")
+function intensities_static_single(scga::SCGA, qpts; β)
     qpts = convert(AbstractQPoints, qpts)
     (; sys, measure) = scga
     Na = natoms(sys.crystal)
     Nobs = num_observables(measure)
-    λ = find_lagrange_multiplier(scga, kT)
+    λ = find_lagrange_multiplier(scga, β)
     intensity = zeros(eltype(measure),length(qpts.qs))
     Ncorr = length(measure.corr_pairs)
     for (iq, q) in enumerate(qpts.qs)
@@ -180,8 +179,8 @@ function intensities_static_single(scga::SCGA, qpts; kT=0.0)
             end
         end
         J_mat = fourier_transform_interaction_matrix(sys; k=q_reshaped)
-        inverted_matrix = (inv(I(3*Na)*λ[1] + (1/kT)*J_mat)) # this is [(Iλ+J(q))^-1]^αβ_μν
-        # inverted_matrix = diagm(1 ./ (λ[1] .+ (1/(kT)).*eigvals(J_mat)))
+        inverted_matrix = (inv(I(3*Na)*λ[1] + β*J_mat)) # this is [(Iλ+J(q))^-1]^αβ_μν
+        # inverted_matrix = diagm(1 ./ (λ[1] .+ β .* eigvals(J_mat)))
         for i in 1:Na
             for j in 1:Na
                 intensitybuf += pref[1,i]*conj(pref[1,j])*inverted_matrix[1+3(i-1):3+3(i-1),1+3(j-1):3+3(j-1)]
@@ -203,14 +202,12 @@ end
 # Computes the static structure factor in the sublattice resolved SCGA method,
 # over the qpts specified.
 
-function intensities_static_sublattice(scga::SCGA, qpts; kT=0.0, λs_init)
-    iszero(kT) && error("kT must be non-zero")
-
+function intensities_static_sublattice(scga::SCGA, qpts; β, λs_init)
     qpts = convert(AbstractQPoints, qpts)
     (; sys, measure) = scga
     Na = natoms(sys.crystal)
[…]
