### 2025-04-18 bed43e4
Functions to look up quadratic interactions
Add get_exchange, get_quadratic_anisotropy, and related variants for
inhomogeneous systems.
