+++ ext/PlottingExt/ViewCrystal.jl
@@ -60,66 +60,6 @@ function propagate_reference_bond_for_cell(cryst, b_ref)
     return reduce(vcat, found)
 end
 
-function anisotropy_on_site(sys, i)
-    interactions = isnothing(sys) ? nothing : Sunny.interactions_homog(something(sys.origin, sys))
-    onsite = interactions[i].onsite
-    if onsite isa Sunny.HermitianC64
-        onsite = Sunny.StevensExpansion(Sunny.matrix_to_stevens_coefficients(onsite))
-    end
-    return onsite :: Sunny.StevensExpansion
-end
-
-# Get the quadratic anisotropy as a 3×3 exchange matrix for atom `i` in the
-# chemical cell.
-function quadratic_anisotropy(sys, i)
-    # Get certain Stevens expansion coefficients
-    (; c0, c2) = anisotropy_on_site(sys, i)
-
-    # Undo RCS renormalization for quadrupolar anisotropy for spin-s
-    if sys.mode == :dipole
-        s = (sys.Ns[i] - 1) / 2
-        c2 = c2 / Sunny.rcs_factors(s)[2] # Don't mutate c2 in-place!
-    end
-
-    # Stevens quadrupole operators expressed as 3×3 spin bilinears
-    quadrupole_basis = [
-        [1 0 0; 0 -1 0; 0 0 0],    # 𝒪₂₂  = SˣSˣ - SʸSʸ
-        [0 0 1; 0 0 0; 1 0 0] / 2, # 𝒪₂₁  = (SˣSᶻ + SᶻSˣ)/2
-        [-1 0 0; 0 -1 0; 0 0 2],   # 𝒪₂₀  = 2SᶻSᶻ - SˣSˣ - SʸSʸ
-        [0 0 0; 0 0 1; 0 1 0] / 2, # 𝒪₂₋₁ = (SʸSᶻ + SᶻSʸ)/2
-        [0 1 0; 1 0 0; 0 0 0],     # 𝒪₂₋₂ = SˣSʸ + SʸSˣ
-    ]
-
-    # The c0 coefficient incorporates a factor of S². For quantum spin
-    # operators, S² = s(s+1) I. For the large-s classical limit, S² = s² is a
-    # scalar.
-    S² = if sys.mode == :dipole_uncorrected
-        # Undoes extraction in `operator_to_stevens_coefficients`. Note that
-        # spin magnitude s² is set to κ², as originates from `onsite_coupling`
-        # for p::AbstractPolynomialLike.
-        sys.κs[i]^2
-    else
-        # Undoes extraction in `matrix_to_stevens_coefficients` where 𝒪₀₀ = I.
-        s = (sys.Ns[i]-1) / 2
-        s * (s+1)
-    end
-
-    return c2' * quadrupole_basis + only(c0) * I / S²
-end
-
-function coupling_on_bond(interactions, b)
-    isnothing(interactions) && return zero(Mat3)
-    pairs = interactions[b.i].pair
-    indices = findall(pc -> pc.bond == b, pairs)
-    return isempty(indices) ? nothing : pairs[only(indices)]
-end
-
-# Get the 3×3 exchange matrix for bond `b`
-function exchange_on_bond(interactions, b)
-    coupling = coupling_on_bond(interactions, b)
-    return isnothing(coupling) ? zero(Mat3) : coupling.bilin * Mat3(I)
-end
-
 # Get largest exchange interaction scale. For symmetric part, this is the
 # largest eigenvalue. For antisymmetric part, this is an empirical rescaling of
 # the norm of the DM vector. (Note that a typical system has small DM vector
@@ -268,7 +208,8 @@ function draw_bonds(; ax, obs, ionradius, exchange_mag, cryst, interactions, bon
[…]
