### 2025-04-18 973bfa6
Fix `view_crystal` on reshaped system
Requires transformation of atom index `i` on original crystal to a site index for a reshaped system.
