+++ ext/PlottingExt/ViewCrystal.jl
@@ -289,8 +289,9 @@ function label_atoms(cryst; ismagnetic, sys)
                 R_site = Sunny.rotation_between_sites(cryst, i, i_ref)
                 push!(ret, Sunny.allowed_g_tensor_string(cryst, i_ref; R_site, prefix="Aniso: ", digits=8, atol=1e-12))
             else
-                stvexp = Sunny.get_stevens_expansion(sys, i)
-                aniso = Sunny.unrenormalize_quadratic_anisotropy(stvexp, sys, (1, 1, 1, i))
+                site = Sunny.map_atom_to_other_system(cryst, i, sys)
+                stvexp = Sunny.get_stevens_expansion_at(sys, site)
+                aniso = Sunny.unrenormalize_quadratic_anisotropy(stvexp, sys, site)
                 basis_strs = Sunny.number_to_simple_string.(aniso; digits=3)
                 push!(ret, Sunny.formatted_matrix(basis_strs; prefix="Aniso: "))
                 if !iszero(stvexp.c4) || !iszero(stvexp.c6)
