### 2025-04-19 04fce6b
Functions to look up quadratic interactions from a system (#366)
Add `get_exchange`, `get_quadratic_anisotropy`, and variants for inhomogeneous systems.
