### 2025-04-19 fa965fc
Functions to look up quadratic interactions from a system (#366)
Add `get_exchange`, `get_quadratic_anisotropy`, and variants for inhomogeneous systems.
