### 2025-04-19 4d1adc0
Allow custom integrator in SampledCorrelations
