### 2025-04-19 e0244fb
Override normalization validation
