+++ src/System/Interactions.jl
@@ -271,8 +271,10 @@ end
 
 The total system energy. See also [`energy_per_site`](@ref).
 """
-function energy(sys::System{N}) where N
-    validate_normalization(sys)
+function energy(sys::System{N}; check_normalization=true) where N
+    if check_normalization 
+        validate_normalization(sys)
+    end
     E = 0.0
 
     # Zeeman coupling to external field
