### 2025-04-19 4750c88
Normalization override again
