### 2025-04-19 e6ea6c0
Abstract integrator type
