### 2025-04-20 eff5ed3
Allow custom integrator in SampledCorrelations (#371)
Co-authored-by: Kipton Barros <kbarros@gmail.com>
