### 2025-04-20 068be31
Allow custom integrator in SampledCorrelations (#371)
Co-authored-by: Kipton Barros <kbarros@gmail.com>
