+++ src/EntangledUnits/EntangledSampledCorrelations.jl
@@ -115,7 +115,7 @@ function SampledCorrelations(esys::EntangledSystem; measure, energies, dt, calcu
     crystal = esys.sys_origin.crystal
     origin_crystal = orig_crystal(esys.sys_origin)
     sc_new = SampledCorrelations(sc.data, sc.M, crystal, origin_crystal, sc.Δω, measure, observables, positions, source_idcs, measure.corr_pairs,
-                                 sc.measperiod, sc.dt, sc.nsamples, sc.samplebuf, sc.corrbuf, sc.space_fft!, sc.time_fft!, sc.corr_fft!, sc.corr_ifft!)
+                                 sc.integrator, sc.measperiod, sc.nsamples, sc.samplebuf, sc.corrbuf, sc.space_fft!, sc.time_fft!, sc.corr_fft!, sc.corr_ifft!)
 
     return EntangledSampledCorrelations(sc_new, esys)
 end
