### 2025-04-23 dae0059
Writing tweaks
