+++ examples/04_GSD_FeI2.jl
@@ -7,10 +7,10 @@
 # states](https://doi.org/10.1103/PhysRevB.106.054423).
 #
 # Compared to spin wave theory, classical spin dynamics in real-space is
-# typically much slower, and is limited in ``𝐪``-space resolution. The
-# approach, however, allows for thermal fluctuations, can be used to explore
-# [finite temperature phases](https://doi.org/10.1103/PhysRevB.109.014427), and
-# enables the study of [highly non-equilibrium
+# typically much slower and is limited in ``𝐪``-space resolution. The approach,
+# however, allows for thermal fluctuations. This allows to explore [finite
+# temperature phases](https://doi.org/10.1103/PhysRevB.109.014427) and enables
+# the study of [non-equilibrium
 # processes](https://doi.org/10.1103/PhysRevB.106.235154).
 #
 # The structure of this tutorial largely follows the [previous study of CoRh₂O₄
