### 2025-04-23 c3f8e9e
Powder averaging of static intensities
