### 2025-04-28 afe8506
Fix broken powder average
