### 2025-04-28 f77429a
Tweak colorrange setup
