### 2025-04-28 fa38ac9
Clean `print_site` on missing allowed anisotropy operators (#375)
