+++ examples/spinw_tutorials/SW10_Energy_cut.jl
@@ -5,7 +5,7 @@
 # Toth. It calculates the spin wave spectrum on a constant energy cut of the
 # frustrated square lattice.
 
-# Load Sunny and the GLMakie plotting package
+# Load Sunny and the GLMakie plotting package.
 
 using Sunny, GLMakie
 
