### 2025-04-28 ad40d1e
Miscellaneous cleanups (#373)
