+++ examples/spinw_tutorials/SW06_Complex_kagome_FM.jl
@@ -5,7 +5,7 @@
 # and Sandor Toth. It calculates the spin wave spectrum of the kagome lattice
 # with multiple competing interactions.
 
-# Load Sunny and the GLMakie plotting package
+# Load Sunny and the GLMakie plotting package.
 
 using Sunny, GLMakie
 
