+++ examples/spinw_tutorials/SW01_FM_Heseinberg_chain.jl
@@ -5,7 +5,7 @@
 # Nilsen and Sandor Toth. It calculates the spin wave spectrum of the
 # ferromagnetic Heisenberg nearest-neighbor spin chain.
 
-# Load Sunny and the GLMakie plotting package
+# Load Sunny and the GLMakie plotting package.
 
 using Sunny, GLMakie
 
