### 2025-04-29 c472e79
Adapt energy axes
