+++ ext/PlottingExt/PlotIntensities.jl
@@ -38,10 +38,11 @@ function colorrange_from_data(; data, saturation, sensitivity, allpositive)
 end
 
 """
-    plot_intensities!(panel, res::AbstractIntensities; opts...)
+    plot_intensities!(panel, res; opts...)
 
 Mutating variant of [`plot_intensities`](@ref) that allows drawing into a single
-panel of a Makie figure.
+panel of a Makie figure. Returns a `Makie.Axis` object to allow for further
+customization of labels, ticks, legends, etc.
 
 # Example
 
@@ -65,16 +66,16 @@ function Sunny.plot_intensities!(panel, res::Sunny.BandIntensities{Float64}; col
         fwhm = @something fwhm 0.02*(ylims[2]-ylims[1])
         σ = fwhm/2√(2log(2))
         kernel = Sunny.Broadening(x -> exp(-x^2/2σ^2)) # Gaussian without normalization
-        broadened = Sunny.broaden(res; energies, kernel)
+        (; data) = Sunny.broaden(res; energies, kernel)
 
-        colorrange_suggest = colorrange_from_data(; broadened.data, saturation, sensitivity, allpositive)
+        colorrange_suggest = colorrange_from_data(; data, saturation, sensitivity, allpositive)
         colormap = @something colormap (allpositive ? Makie.Reverse(:thermal) : :bwr)
         colorrange = @something colorrange colorrange_suggest
 
+        xlims = extrema(axes(data, 2))
         xticklabelrotation = maximum(length.(res.qpts.xticks[2])) > 3 ? π/6 : 0.0
         ax = Makie.Axis(panel; xlabel="Momentum (r.l.u.)", ylabel, res.qpts.xticks, xticklabelrotation, axisopts...)
-        Makie.ylims!(ax, ylims ./ unit_energy)
-        Makie.heatmap!(ax, axes(res.data, 2), collect(energies / unit_energy), broadened.data'; colorrange, colormap, lowclip=:white)
+        Makie.image!(ax, xlims, ylims ./ unit_energy, data'; colorrange, colormap, lowclip=:white)
         for i in axes(res.disp, 1)
             Makie.lines!(ax, res.disp[i,:] / unit_energy; color=:lightskyblue3)
         end
@@ -121,7 +122,9 @@ function suggest_labels_for_grid(grid::Sunny.QGrid{N}) where N
 end
 
 
-function Sunny.plot_intensities!(panel, res::Sunny.Intensities{Float64}; colormap=nothing, colorrange=nothing, saturation=0.9, allpositive=true, units=nothing, into=nothing, title="", axisopts=NamedTuple())
+function Sunny.plot_intensities!(panel, res::Sunny.Intensities{Float64}; colormap=nothing, colorrange=nothing,
+                                 saturation=0.9, allpositive=true, units=nothing, into=nothing, ylims=nothing,
+                                 title="", axisopts=NamedTuple())
     axisopts = (; title, axisopts...)
     (; crystal, qpts, data, energies) = res
 
@@ -131,18 +134,20 @@ function Sunny.plot_intensities!(panel, res::Sunny.Intensities{Float64}; colorma
 
     if qpts isa Sunny.QPath
         unit_energy, ylabel = get_unit_energy(units, into)
+        xlims = extrema(axes(data, 2))
+        ylims = @something ylims (extrema(energies) ./ unit_energy)
         xticklabelrotation = maximum(length.(qpts.xticks[2])) > 3 ? π/6 : 0.0
         ax = Makie.Axis(panel[1, 1]; xlabel="Momentum (r.l.u.)", ylabel, qpts.xticks, xticklabelrotation, axisopts...)
[…]
