### 2025-04-29 81c169f
Error on impossible quantum spin-1/2 couplings (#377)
Make `set_onsite_coupling!` and `set_pair_coupling!` throw an informative error for sites with quantum spin 1/2.
