### 2025-04-29 b16ae65
Powder averaging of static intensities (#374)
Allow to call `intensities_static` from within `powder_average`.

Also includes tweaks to `plot_intensities`. Documentation is streamlined and `ylims` is supported uniformly. Internally, calls to `Makie.heatmap!` have been replaced with `Makie.image!` since the pixel grid is uniform.

Various tutorials have been refreshed to make intensity plots more uniform in style and scale.
