+++ examples/spinw_tutorials/SW08_sqrt3_kagome_AFM.jl
@@ -62,7 +62,7 @@ fig
 # empirical `colorrange` that brings the lower-intensity features into focus.
 
 radii = range(0, 2.5, 200)
-energies = range(0, 3, 200)
+energies = range(0, 2.5, 200)
 kernel = gaussian(fwhm=0.05)
 res = powder_average(cryst, radii, 200) do qs
     intensities(swt, qs; energies, kernel)
