+++ ext/PlottingExt/PlotIntensities.jl
@@ -10,8 +10,10 @@ function get_unit_energy(units, into)
 end
 
 function colorrange_from_data(; data, saturation, sensitivity, allpositive)
-    cmax = Statistics.quantile(filter(!isnan, vec(maximum(data; dims=1))), saturation)
-    cmin = Statistics.quantile(filter(!isnan, vec(minimum(data; dims=1))), 1 - saturation)
+    max_data = ndims(data) == 1 ? data : vec(maximum(data; dims=1))
+    min_data = ndims(data) == 1 ? data : vec(minimum(data; dims=1))
+    cmax = Statistics.quantile(filter(!isnan, max_data), saturation)
+    cmin = Statistics.quantile(filter(!isnan, min_data), 1 - saturation)
 
     # The returned pair (lo, hi) should be strictly ordered, lo < hi, for use in
     # Makie.Colorbar
@@ -36,10 +38,11 @@ function colorrange_from_data(; data, saturation, sensitivity, allpositive)
 end
 
 """
-    plot_intensities!(panel, res::AbstractIntensities; opts...)
+    plot_intensities!(panel, res; opts...)
 
 Mutating variant of [`plot_intensities`](@ref) that allows drawing into a single
-panel of a Makie figure.
+panel of a Makie figure. Returns a `Makie.Axis` object to allow for further
+customization of labels, ticks, legends, etc.
 
 # Example
 
@@ -58,20 +61,21 @@ function Sunny.plot_intensities!(panel, res::Sunny.BandIntensities{Float64}; col
  
     if res.qpts isa Sunny.QPath 
         mindisp, maxdisp = extrema(res.disp)
-        ylims = @something ylims (min(0, mindisp), 1.2*maxdisp)
+        ylims = @something ylims (min(0, mindisp), 1.1*maxdisp)
         energies = range(ylims[1], ylims[2], 512)
         fwhm = @something fwhm 0.02*(ylims[2]-ylims[1])
         σ = fwhm/2√(2log(2))
         kernel = Sunny.Broadening(x -> exp(-x^2/2σ^2)) # Gaussian without normalization
-        broadened = Sunny.broaden(res; energies, kernel)
+        (; data) = Sunny.broaden(res; energies, kernel)
 
-        colorrange_suggest = colorrange_from_data(; broadened.data, saturation, sensitivity, allpositive)
+        colorrange_suggest = colorrange_from_data(; data, saturation, sensitivity, allpositive)
         colormap = @something colormap (allpositive ? Makie.Reverse(:thermal) : :bwr)
         colorrange = @something colorrange colorrange_suggest
 
+        xlims = extrema(axes(data, 2))
         xticklabelrotation = maximum(length.(res.qpts.xticks[2])) > 3 ? π/6 : 0.0
-        ax = Makie.Axis(panel; xlabel="Momentum (r.l.u.)", ylabel, limits=(nothing, ylims ./ unit_energy), res.qpts.xticks, xticklabelrotation, axisopts...)
-        Makie.heatmap!(ax, axes(res.data, 2), collect(energies / unit_energy), broadened.data'; colorrange, colormap, lowclip=:white)
+        ax = Makie.Axis(panel; xlabel="Momentum (r.l.u.)", ylabel, res.qpts.xticks, xticklabelrotation, axisopts...)
+        Makie.image!(ax, xlims, ylims ./ unit_energy, data'; colorrange, colormap, lowclip=:white)
         for i in axes(res.disp, 1)
             Makie.lines!(ax, res.disp[i,:] / unit_energy; color=:lightskyblue3)
         end
@@ -118,28 +122,32 @@ function suggest_labels_for_grid(grid::Sunny.QGrid{N}) where N
 end
 
 
[…]
