+++ dev/assets/scripts/SW18_Distorted_kagome.jl
@@ -48,7 +48,7 @@ res = intensities_bands(swt, path)
 plot_intensities(res; units)
 
 radii = range(0, 2, 100) # (1/Å)
-energies = range(0, 6, 200)
+energies = range(0, 5, 200)
 kernel = gaussian(fwhm=0.05)
 res = powder_average(cryst, radii, 400) do qs
     intensities(swt, qs; energies, kernel)
