+++ previews/PR355/assets/scripts/SW02_AFM_Heisenberg_chain.jl
@@ -23,5 +23,5 @@ plot_intensities(res; units)
 isapprox(res.disp[1, :], res.disp[2, :])
 
 xs = [q[1] for q in path.qs]
-ys = log10.(res.data[1, :] + res.data[2, :])
-lines(xs, ys; axis=(; xlabel="[H, 0, 0]", ylabel="Log intensity (dimensionless)"))
+ys = res.data[1, :] + res.data[2, :]
+lines(xs, ys; axis=(; xlabel="[H, 0, 0]", ylabel="Intensity", yscale=log10))
