### 2025-05-06 d8f54eb
Unify solvers
