### 2025-05-06 0c789d8
Avoid memory allocation for speedup
