### 2025-05-06 0e001b1
Allow integer kT
