### 2025-05-06 0b14f57
More speedup
