### 2025-05-06 317cdc2
Rename calculator to just `SCGA`
