### 2025-05-06 86bff41
Support general observables
