+++ src/Spiral/LuttingerTisza.jl
@@ -1,47 +1,71 @@
-# Returns the minimum eigenvalue of the Luttinger-Tisza exchange matrix. It is a
-# lower-bound on the exchange energy for magnetic order that has propagation
-# wavevector k between Bravais lattice cells. This analysis ignores local
-# normalization constraints for the spins within the cell.
-function luttinger_tisza_exchange(sys::System; k, ϵ=0)
+function fourier_exchange_matrix(sys::System; q)
     @assert sys.mode in (:dipole, :dipole_uncorrected) "SU(N) mode not supported"
     @assert sys.dims == (1, 1, 1) "System must have only a single cell"
+    q_reshaped = to_reshaped_rlu(sys, q)
 
     Na = natoms(sys.crystal)
-    J_k = zeros(ComplexF64, 3, Na, 3, Na)
+    J_q = zeros(ComplexF64, 3, Na, 3, Na)
 
-    for i in 1:natoms(sys.crystal)
+    for i in 1:Na
         for coupling in sys.interactions_union[i].pair
-            (; isculled, bond, bilin) = coupling
+            (; isculled, bond, bilin, biquad) = coupling
             isculled && break
+            iszero(biquad) || error("Biquadratic interactions not supported")
 
             (; j, n) = bond
-            J = exp(2π * im * dot(k, n)) * Mat3(bilin*I)
-            J_k[:, i, :, j] += J / 2
-            J_k[:, j, :, i] += J' / 2
+            J = exp(2π * im * dot(q_reshaped, n)) * Mat3(bilin*I)
+            J_q[:, i, :, j] += J
+            J_q[:, j, :, i] += J'
         end
     end
 
     if !isnothing(sys.ewald)
-        A = Sunny.precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), k) * sys.ewald.μ0_μB²
-        A = reshape(A, Na, Na)
+        A_q = Sunny.precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), q_reshaped) * sys.ewald.μ0_μB²
+        A_q = reshape(A_q, Na, Na)
         for i in 1:Na, j in 1:Na
-            J_k[:, i, :, j] += sys.gs[i]' * A[i, j] * sys.gs[j] / 2
+            J_q[:, i, :, j] += sys.gs[i]' * A_q[i, j] * sys.gs[j]
         end
     end
 
-    J_k = reshape(J_k, 3*Na, 3*Na)
-    @assert diffnorm2(J_k, J_k') < 1e-15
-    J_k = hermitianpart(J_k)
+    for i in 1:Na
+        onsite_coupling = sys.interactions_union[i].onsite
+        (; c2, c4, c6) = onsite_coupling
+        iszero(c4) && iszero(c6) || error("Single-ion anisotropy beyond quadratic order not supported")
+        anisotropy = SA[c2[1]-c2[3]        c2[5] 0.5c2[2];
+                              c2[5] -c2[1]-c2[3] 0.5c2[4];
+                           0.5c2[2]     0.5c2[4]   2c2[3]]
+        J_q[:, i, :, i] += 2 * anisotropy
+    end
+
+    J_q = reshape(J_q, 3Na, 3Na)
+    @assert diffnorm2(J_q, J_q') < 1e-15
+    return hermitianpart(J_q)
+end
+
+
+# Returns the Luttinger-Tisza predicted exchange energy associated with the
+# propagation wavevector k between chemical cells. The LT analysis minimizes
+# energy E = (1/2) Sₖ† Jₖ Sₖ, where Sₖ is some length-3Nₐ vector and Jₖ is the
+# 3Nₐ×3Nₐ exchange matrix. Given a minimum energy eigenpair (λₖ, Sₖ) the
[…]
