### 2025-05-06 5a53954
Newton method working
