### 2025-05-06 548b254
Improve accuracy with pivoted Cholesky
