+++ src/SCGA/NewtonBacktracking.jl
@@ -1,4 +1,4 @@
-function newton_with_backtracking(fgh!, x0; g_abstol=NaN, f_reltol=1e-12, maxiters=20, armijo_c=1e-4, armijo_backoff=0.5, armijo_α_min=1e-6, show_trace=false)
+function newton_with_backtracking(fgh!, x0; f_reltol=NaN, x_reltol=NaN, g_abstol=NaN, maxiters=20, armijo_c=1e-2, armijo_backoff=0.1, armijo_α_min=1e-4, show_trace=false)
     # Make a copy of the initial guess.
     x = copy(x0)
 
@@ -10,11 +10,13 @@ function newton_with_backtracking(fgh!, x0; g_abstol=NaN, f_reltol=1e-12, maxite
 
     # Evaluate objective function f, gradient, and Hessian.
     f = fgh!(0.0, g, H, x)
+    norm(g) < g_abstol && return x
 
-    function has_converged(f, candidate_f, g)
-        return (norm(g) < g_abstol) || isapprox(f, candidate_f; rtol=f_reltol)
+    function has_converged(x, candidate_x, f, candidate_f, g)
+        return (!isnan(x_reltol) && isapprox(x, candidate_x; rtol=x_reltol)) ||
+               (!isnan(f_reltol) && isapprox(f, candidate_f; rtol=f_reltol)) ||
+               (!isnan(g_abstol) && norm(g) < g_abstol)
     end
-    has_converged(NaN, NaN, g) && return x
 
     for k in 1:maxiters
         # Newton direction p = H \ g. If H is not guaranteed positive definite,
@@ -34,7 +36,7 @@ function newton_with_backtracking(fgh!, x0; g_abstol=NaN, f_reltol=1e-12, maxite
         # Backtracking line search until Armijo condition is satisfied:
         # f(candidate_x) ≤ f(x) - c * α * dot(g, p)
         while candidate_f > f - armijo_c * α * g_dot_p
-            has_converged(f, candidate_f, g) && return x
+            has_converged(x, candidate_x, f, candidate_f, g) && return candidate_x
             α < armijo_α_min && error("Failed to satisfy Armijo condition. Consider reducing armijo_α_min=$armijo_α_min or a tolerance parameter.")
 
             α *= armijo_backoff
@@ -45,7 +47,7 @@ function newton_with_backtracking(fgh!, x0; g_abstol=NaN, f_reltol=1e-12, maxite
         if show_trace
             println("Iter $k: f(x)=$candidate_f, x=$candidate_x, α=$α, |g|=$(norm(g))")
         end
-        has_converged(f, candidate_f, g) && return x
+        has_converged(x, candidate_x, f, candidate_f, g) && return candidate_x
 
         # Accept candidate updates. Note that g and H will also be reused.
         x .= candidate_x
