+++ src/SCGA/StaticCorrelationsSCGA.jl
@@ -88,7 +88,6 @@ function find_lagrange_multiplier(sys, Js, β)
     for n in 1:maxiters
         λ = λn + (1/J(λn))*(s²-f(λn))
         if abs(λ-λn) < tol
-            println("Newton's method converged to within tolerance, $tol, after $n steps.")
             return λ
         else
             λn = λ
@@ -97,7 +96,7 @@ function find_lagrange_multiplier(sys, Js, β)
 end
 
 
-function find_lagrange_multiplier_opt_sublattice(sys, Js, β; rtol=1e-10)
+function find_lagrange_multiplier_opt_sublattice(sys, Js, β)
     Na = natoms(sys.crystal)
     s² = vec(sys.κs .^ 2)
 
@@ -110,38 +109,41 @@ function find_lagrange_multiplier_opt_sublattice(sys, Js, β; rtol=1e-10)
             hbuffer .= 0
         end
 
-        Λ = diagm(repeat(λs, inner=3))
+        Λ = Diagonal(repeat(λs, inner=3))
+        A = zeros(ComplexF64, 3Na, 3Na)
+        A⁻¹ = zeros(ComplexF64, 3, Na, 3, Na)
 
+        # Determine the Lagrange multipliers λ by maximizing the "grand" free
+        # energy G(λ) = log det A / 2β - ∑ᵢ λᵢ s²ᵢ, where A = J + Λ. Implement
+        # this numerically as minimization of the objective function f = -G.
         for J in Js
-            A = β * (J + Λ)
-            T = eigen(A)
-            eig_vals = T.values
-            U = T.vectors
-            A⁻¹ = U * Diagonal(inv.(eig_vals)) * U'
-            A⁻¹ = reshape(A⁻¹, 3, Na, 3, Na)
-
-            # Determine Lagrange multipliers λ by minimizing f, the negative of
-            # the "grand" free energy G(λ), involving Legendre transform into
-            # the λ variables. Physical eigenvalues of each (J + Λ) matrix must
-            # be positive, otherwise we apply an infinite penalty to the
-            # objective function.
-            if minimum(eig_vals) < 0
-                fbuffer = Inf
-            else
-                fbuffer += λs' * s² / 2 - (1/2β) * sum(log.(eig_vals))
+            # Cholesky decomposition fails if the matrix A is not positive
+            # definite. This implies unphysical λ values, which we penalize by
+            # making the objective function infinite.
+            @. A = J + Λ
+            A_chol = cholesky!(A, RowMaximum(); check=false)
+            if !issuccess(A_chol)
+                gbuffer .= NaN
+                hbuffer .= NaN
+                return Inf
             end
+
+            ldiv!(reshape(A⁻¹, 3Na, 3Na), A_chol, I(3Na))
+
+            # The finite-valued objective function f
+            fbuffer += (λs' * s²) / 2 - logdet(A_chol) / 2β
 
             # Gradient of f
             if !isnothing(gbuffer)
                 for i in 1:Na
-                    gbuffer[i] += s²[i] / 2 - real(tr(view(A⁻¹, :, i, :, i))) / 2
+                    gbuffer[i] += s²[i] / 2 - real(tr(view(A⁻¹, :, i, :, i))) / 2β
                 end
             end
 
             # Hessian of f
             if !isnothing(hbuffer)
                 for i in 1:Na, j in 1:Na
-                    hbuffer[i, j] += + (β/2) * norm(view(A⁻¹, :, i, :, j))^2
+                    hbuffer[i, j] += + norm(view(A⁻¹, :, i, :, j))^2 / 2β
                 end
             end
[…]
