### 2025-05-06 92f9d80
Custom Newton with backtracking
