+++ src/Sunny.jl
@@ -116,6 +116,7 @@ include("SampledCorrelations/DataRetrieval.jl")
 export SampledCorrelations, SampledCorrelationsStatic, add_sample!, clone_correlations,
     merge_correlations
 
+include("SCGA/NewtonBacktracking.jl")
 include("SCGA/StaticCorrelationsSCGA.jl")
 export StaticCorrelationsSCGA
 
