+++ previews/PR355/assets/scripts/SW19_Different_Ions.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.6"
+@assert pkgversion(Sunny) >= v"0.7.7"
 
 units = Units(:meV, :angstrom)
 a = 3.0
