+++ previews/PR355/assets/scripts/SW11_La2CuO4.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.6"
+@assert pkgversion(Sunny) >= v"0.7.7"
 
 units = Units(:meV, :angstrom)
 latvecs = lattice_vectors(3.85, 3.85, 12.25, 90, 90, 90)
