+++ previews/PR355/assets/scripts/SW14_YVO3.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.6"
+@assert pkgversion(Sunny) >= v"0.7.7"
 
 units = Units(:meV, :angstrom)
 a = 5.2821 / sqrt(2)
