+++ previews/PR355/assets/scripts/SW10_Energy_cut.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.6"
+@assert pkgversion(Sunny) >= v"0.7.7"
 
 units = Units(:meV, :angstrom)
 latvecs = lattice_vectors(1.0, 1.0, 3.0, 90, 90, 90)
