+++ previews/PR355/assets/scripts/SW08_sqrt3_kagome_AFM.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.6"
+@assert pkgversion(Sunny) >= v"0.7.7"
 
 units = Units(:meV, :angstrom)
 latvecs = lattice_vectors(6, 6, 40, 90, 90, 120)
