+++ previews/PR355/assets/scripts/SW18_Distorted_kagome.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.6"
+@assert pkgversion(Sunny) >= v"0.7.7"
 
 units = Units(:meV, :angstrom)
 latvecs = lattice_vectors(10.2, 5.94, 7.81, 90, 117.7, 90)
