### 2025-05-06 1ad1f5b
Revise notation
