### 2025-05-09 d2cb75c
Convert `Makie.image` to `Makie.heatmap`
This allows to use `Makie.Resampler` as a workaround to https://github.com/MakieOrg/Makie.jl/issues/4950.
