+++ ext/PlottingExt/PlotIntensities.jl
@@ -61,9 +61,13 @@ function Sunny.plot_intensities!(panel, res::Sunny.BandIntensities{Float64}; col
  
     if res.qpts isa Sunny.QPath 
         mindisp, maxdisp = extrema(res.disp)
-        ylims = @something ylims (min(0, mindisp), 1.1*maxdisp)
-        energies = range(ylims[1], ylims[2], 512)
-        fwhm = @something fwhm 0.02*(ylims[2]-ylims[1])
+        ebounds = if isnothing(ylims)
+            (min(0, mindisp), 1.1*maxdisp)
+        else
+            ylims .* unit_energy
+        end
+        energies = range(ebounds[1], ebounds[2], 512)
+        fwhm = @something fwhm 0.02*(ebounds[2]-ebounds[1])
         σ = fwhm/2√(2log(2))
         kernel = Sunny.Broadening(x -> exp(-x^2/2σ^2)) # Gaussian without normalization
         (; data) = Sunny.broaden(res; energies, kernel)
@@ -72,12 +76,11 @@ function Sunny.plot_intensities!(panel, res::Sunny.BandIntensities{Float64}; col
         colormap = @something colormap (allpositive ? Makie.Reverse(:thermal) : :bwr)
         colorrange = @something colorrange colorrange_suggest
 
-        xlims = extrema(axes(data, 2))
         xticklabelrotation = maximum(length.(res.qpts.xticks[2])) > 3 ? π/6 : 0.0
         ax = Makie.Axis(panel; xlabel="Momentum (r.l.u.)", ylabel, res.qpts.xticks, xticklabelrotation, axisopts...)
-        Makie.image!(ax, xlims, ylims ./ unit_energy, data'; colorrange, colormap, lowclip=:white)
+        Makie.heatmap!(ax, axes(data, 2), collect(energies/unit_energy), data'; colorrange, colormap, lowclip=:white)
         for i in axes(res.disp, 1)
-            Makie.lines!(ax, res.disp[i,:] / unit_energy; color=:lightskyblue3)
+            Makie.lines!(ax, res.disp[i,:]/unit_energy; color=:lightskyblue3)
         end
         return ax
     else
@@ -134,20 +137,19 @@ function Sunny.plot_intensities!(panel, res::Sunny.Intensities{Float64}; colorma
 
     if qpts isa Sunny.QPath
         unit_energy, ylabel = get_unit_energy(units, into)
-        xlims = extrema(axes(data, 2))
-        ylims = @something ylims (extrema(energies) ./ unit_energy)
         xticklabelrotation = maximum(length.(qpts.xticks[2])) > 3 ? π/6 : 0.0
         ax = Makie.Axis(panel[1, 1]; xlabel="Momentum (r.l.u.)", ylabel, qpts.xticks, xticklabelrotation, axisopts...)
-        hm = Makie.image!(ax, xlims, ylims, data'; colormap, colorrange)
+        isnothing(ylims) || Makie.ylims!(ax, ylims)
+        hm = Makie.heatmap!(ax, axes(data, 2), collect(energies/unit_energy), data'; colormap, colorrange)
         Makie.Colorbar(panel[1, 2], hm)
         return ax
     elseif qpts isa Sunny.QGrid{2}
         if isone(length(energies))
             aspect = grid_aspect_ratio(crystal, qpts)
             xlabel, ylabel = suggest_labels_for_grid(qpts)
+            (xs, ys) = map(range, qpts.coefs_lo, qpts.coefs_hi, size(qpts.qs))
             ax = Makie.Axis(panel[1, 1]; xlabel, ylabel, aspect, axisopts...)
[…]
