### 2025-05-12 07399e6
Eliminate unnecessary `Ns` from `EntangledSpinWaveTheory`
