### 2025-05-13 a889556
Self-consistent Gaussian Approximation (SCGA) (#355)
Adds `SCGA` and `enable_spin_rescaling_for_static_sum_rule`.
