+++ src/SCGA/SCGA.jl
@@ -1,0 +1,229 @@
+"""
+    SCGA(sys::System; measure, kT, dq)
+
+Constructs an object to calculate [`intensities_static`](@ref) within the
+self-consistent Gaussian approximation (SCGA). This theory assumes a classical
+Boltzmann distribution with temperature `kT`. It is expected to be meaningful
+above the ordering temperature, where fluctuations are approximately Gaussian.
+
+Only `:dipole` and `:dipole_uncorrected` system modes are supported.
+
+The theory of SCGA approximates local spin magnitude constraints with a _weaker_
+global constraint condition. For each spin sublattice, the global spin sum rule
+can be expressed as an integral over the unit cube ``𝐪 ∈ [0,1]^3`` for
+wavevectors ``𝐪`` in reciprocal lattice units (RLU). Each such integral will be
+approximated as a discrete sum over a regular grid of `floor(1/dq)^3`
+wavevectors for the provided `dq` value.
+
+If the conventional crystal cell admits a smaller primitive cell, then the SCGA
+calculations can be accelerated. Construct a smaller system with
+[`reshape_supercell`](@ref) and [`primitive_cell`](@ref). In this case, the
+discretized ``𝐪``-point grid runs over the full Brillouin zone associated with
+the primitive cell of the crystal.
+"""
+struct SCGA
+    sys :: System
+    measure :: MeasureSpec
+    β :: Float64
+    λs :: Vector{Float64}
+
+    function SCGA(sys::System; measure::Union{Nothing, MeasureSpec}, kT::Real, dq::Float64)
+        sys.dims == (1, 1, 1) || error("System dims must be (1, 1, 1).")
+
+        measure = @something measure empty_measurespec(sys)
+        if size(eachsite(sys)) != size(measure.observables)[2:5]
+            error("Size mismatch. Check that measure is built using consistent system.")
+        end
+
+        if !(sys.mode in (:dipole, :dipole_uncorrected))
+            error("SCGA requires :dipole or :dipole_uncorrected mode.")
+        end
+
+        kT > 0 || error("Temperature kT must be positive")
+        β = 1 / kT
+
+        0 < dq < 1 || error("Select q-space resolution 0 < dq < 1.")
+        qs = make_q_grid(sys, dq)
+        Js = [fourier_exchange_matrix(sys; q) for q in qs]
+
+        # Initial guess for the Lagrange multipliers in the physically allowed
+        # space (all shifted J matrices positive definite). If `eigmin` becomes
+        # a bottleneck, we can try weaker bounds using Gershgorin's circle
+        # theorem: https://en.wikipedia.org/wiki/Gershgorin_circle_theorem.
+        λ_init = -minimum(eigmin(J) for J in Js) + 1/β
+
+        λs = if allequal(sys.crystal.classes)
+            find_lagrange_multiplier_single(sys, Js, β, λ_init)
+        else
+            find_lagrange_multiplier_multi(sys, Js, β, λ_init)
+        end
+
+        return new(sys, measure, β, λs)
+    end
+end
+
+function make_q_grid(sys, dq)
+    wraps = [false, false, false]
+    for i in 1:natoms(sys.crystal)
+        for coupling in sys.interactions_union[i].pair
+            (; isculled, bond) = coupling
+            isculled && break
+            @. wraps = wraps || !iszero(bond.n)
+        end
+    end
+
[…]
