### 2025-05-13 42627ef
Self-consistent Gaussian Approximation (SCGA) (#355)
Adds `SCGA` and `enable_spin_rescaling_for_static_sum_rule`.
