### 2025-05-15 3118907
Fix `plot_intensities` for data on very large grids (#379)
The 2D array data for `Makie.heatmap` will be wrapped in a `Makie.Resampler` if either of its dimensions exceed a threshold. This works around https://github.com/MakieOrg/Makie.jl/issues/4950.
