+++ ext/PlottingExt/PlotIntensities.jl
@@ -37,6 +37,14 @@ function colorrange_from_data(; data, saturation, sensitivity, allpositive)
     end
 end
 
+# Makie.heatmap can fail on large grids due to an OpenGL error. A workaround is
+# to wrap the data in a Makie.Resampler if any of the array dimensions is
+# "large" https://github.com/MakieOrg/Makie.jl/issues/4950. The empirical cutoff
+# 2000 should (hopefully) be conservative for most systems.
+function maybe_resample(data)
+    all(<(2000), size(data)) ? data : Makie.Resampler(data)
+end
+
 """
     plot_intensities!(panel, res; opts...)
 
@@ -61,9 +69,10 @@ function Sunny.plot_intensities!(panel, res::Sunny.BandIntensities{Float64}; col
  
     if res.qpts isa Sunny.QPath 
         mindisp, maxdisp = extrema(res.disp)
-        ylims = @something ylims (min(0, mindisp), 1.1*maxdisp)
-        energies = range(ylims[1], ylims[2], 512)
-        fwhm = @something fwhm 0.02*(ylims[2]-ylims[1])
+        ylims = @something ylims (min(0, mindisp), 1.1*maxdisp) ./ unit_energy
+        ebounds = ylims .* unit_energy
+        energies = range(ebounds[1], ebounds[2], 512)
+        fwhm = @something fwhm 0.02*(ebounds[2]-ebounds[1])
         σ = fwhm/2√(2log(2))
         kernel = Sunny.Broadening(x -> exp(-x^2/2σ^2)) # Gaussian without normalization
         (; data) = Sunny.broaden(res; energies, kernel)
@@ -72,12 +81,11 @@ function Sunny.plot_intensities!(panel, res::Sunny.BandIntensities{Float64}; col
         colormap = @something colormap (allpositive ? Makie.Reverse(:thermal) : :bwr)
         colorrange = @something colorrange colorrange_suggest
 
-        xlims = extrema(axes(data, 2))
         xticklabelrotation = maximum(length.(res.qpts.xticks[2])) > 3 ? π/6 : 0.0
-        ax = Makie.Axis(panel; xlabel="Momentum (r.l.u.)", ylabel, res.qpts.xticks, xticklabelrotation, axis...)
-        Makie.image!(ax, xlims, ylims ./ unit_energy, data'; colorrange, colormap, lowclip=:white)
+        ax = Makie.Axis(panel; xlabel="Momentum (r.l.u.)", ylabel, res.qpts.xticks, xticklabelrotation, limits=(nothing, ylims), axis...)
+        Makie.heatmap!(ax, (1, size(data, 2)), ylims, maybe_resample(collect(data')); colorrange, colormap, lowclip=:white)
         for i in axes(res.disp, 1)
-            Makie.lines!(ax, res.disp[i,:] / unit_energy; color=:lightskyblue3)
+            Makie.lines!(ax, res.disp[i,:]/unit_energy; color=:lightskyblue3)
         end
         return ax
     else
@@ -134,20 +142,18 @@ function Sunny.plot_intensities!(panel, res::Sunny.Intensities{Float64}; colorma
 
     if qpts isa Sunny.QPath
         unit_energy, ylabel = get_unit_energy(units, into)
-        xlims = extrema(axes(data, 2))
-        ylims = @something ylims (extrema(energies) ./ unit_energy)
         xticklabelrotation = maximum(length.(qpts.xticks[2])) > 3 ? π/6 : 0.0
[…]
