### 2025-05-16 55d914f
Document demagnetization
