### 2025-05-19 a366132
Fix spiral calculations on reshaped systems
