### 2025-05-20 cd2e597
Document demagnetization
