+++ src/System/Interactions.jl
@@ -62,7 +62,7 @@ end
 
 
 """
-    enable_dipole_dipole!(sys::System, μ0_μB²)
+    enable_dipole_dipole!(sys::System, μ0_μB²; demag)
 
 Enables long-range interactions between magnetic dipole moments,
 
@@ -71,33 +71,78 @@ Enables long-range interactions between magnetic dipole moments,
 ```
 
 where the sum is over all pairs of sites (singly counted), including periodic
-images, regularized using the Ewald summation convention. Each magnetic moment
-is ``μ = -g μ_B 𝐒``, where ``𝐒`` is the spin angular momentum dipole. The
-parameter `μ0_μB²` specifies the physical constant ``μ_0 μ_B^2``, which has
-dimensions of length³-energy. Obtain this constant for a given system of
-[`Units`](@ref) via its `vacuum_permeability` property.
+images. Each magnetic moment is ``μ_i = -g μ_B 𝐒_i``, where ``𝐒_i`` is the
+spin angular momentum dipole. The parameter `μ0_μB²` specifies the physical
+constant ``μ_0 μ_B^2``, which has dimensions of length³-energy. Obtain this
+constant for a given system of [`Units`](@ref) via its `vacuum_permeability`
+property.
+
+Geometry of the macroscopic sample enters through the demagnetization tensor
+``ℕ``, denoted `demag`. Special cases are:
+
+  * `demag = 1/3` for a spherical geometry in vacuum.
+  * `demag = Diagonal(0, 0, 1)`` for a slab-like geometry perpendicular to the
+    ``ẑ``-axis.
+  * `demag = Diagonal(1/2, 1/2, 0)` for a rod-like geometry aligned with the
+    ``ẑ``-axis.
+
+Selecting `demag = 0` will neglect demagnetization. This is generally not
+desirable, but can be justified in certain special cases such as a rod-like
+geometry that is aligned with the magnetization dipole.
 
 # Example
 
 ```julia
 units = Units(:meV, :angstrom)
-enable_dipole_dipole!(sys, units.vacuum_permeability)
+enable_dipole_dipole!(sys, units.vacuum_permeability; demag=1/3)
 ```
 
+!!! tip "Origin of demagnetization"  
+
+  Formal summation over the infinitely many dipole-dipole pair interactions
+  becomes mathematically ambiguous when the sample has a net magnetic dipole,
+  ```math
+      𝐌 = ∑_i μ_i.
+  ```
+
+  The traditional Ewald method resolves this ambiguity by neglecting surface
+  effects that would lead to demagnetization. For physical correctness, however,
+  the Ewald energy should be augmented with a surface contribution to the total
+  energy,
+  ```math
+      E_s = μ_0 𝐌⋅ℕ 𝐌 / 2V.
+  ```
+
+  Geometry enters through the demagnetization tensor ``ℕ``, which can be written
+  as an integral over the sample volume,
+  ```math
+      ℕ = - (1 / 4π) ∫ d𝐱 ∇ ∇ (1 / |𝐱|).
+  ```
+
+  This integral is valid in vacuum, such that ``ℕ`` has trace 1. If the sample
+  is embedded in another material with relative permeability ``μ' > 1`` it will
+  typically lead to a reduction in ``ℕ``. For example, a spherical inclusion has
+  ``ℕ = 1/(2μ'+1)``, which coincides with the vacuum case when ``μ' = 1``. In
+  the opposite limit ``μ' → ∞`` (sometimes called "tin foil boundary
+  conditions") the demagnetization tensor ``ℕ`` vanishes.
+
 !!! tip "Efficiency considerations"  
-    Dipole-dipole interactions are very efficient in the context of spin
-    dynamics simulation, e.g. [`Langevin`](@ref). Sunny applies the fast Fourier
[…]
