+++ test/test_ewald.jl
@@ -14,18 +14,18 @@
     end
 
     # Long-range energy of single dipole in cubic box with PBC
-    latvecs = lattice_vectors(1,1,1,90,90,90)
-    positions = [[0,0,0]]
+    latvecs = lattice_vectors(1, 1, 1, 90, 90, 90)
+    positions = [[0, 0, 0]]
     cryst = Crystal(latvecs, positions)
     moments = [1 => Moment(s=1, g=1)]
     sys = System(cryst, moments, :dipole)
-    enable_dipole_dipole!(sys, 1.0)
+    enable_dipole_dipole!(sys, 1.0; demag=1/3)
     @test ewalder_energy(sys) ≈ -1/6
     @test isapprox(energy(sys), -1/6; atol=1e-13)
 
     # Same thing, with multiple unit cells
     sys = System(cryst, moments, :dipole; dims=(2, 3, 4))
-    enable_dipole_dipole!(sys, 1.0)
+    enable_dipole_dipole!(sys, 1.0; demag=1/3)
     @test isapprox(energy_per_site(sys), -1/6; atol=1e-13)
 
     # Create a random box
@@ -39,7 +39,7 @@
         3 => Moment(s=2, g=rand(3,3)),
     ]
     sys = System(cryst, moments, :dipole)
-    enable_dipole_dipole!(sys, 1.0)
+    enable_dipole_dipole!(sys, 1.0; demag=1/3)
     randomize_spins!(sys)
 
     # Energy per site is independent of resizing
