+++ test/test_optimization.jl
@@ -73,7 +73,7 @@ end
         sys = System(cryst, [1 => Moment(; s, g=2)], mode; dims, seed)
         set_exchange!(sys, -1, Bond(1,1,[1,0,0]))
         set_onsite_coupling!(sys, S -> -S[3]^2, 1)
-        enable_dipole_dipole!(sys, 1.0)
+        enable_dipole_dipole!(sys, 1.0; demag=1/3)
         sys
     end
 
