### 2025-05-20 2d73088
Implement demagnetization
