+++ test/test_lswt.jl
@@ -296,7 +296,7 @@ end
 
     for mode in (:dipole, :SUN)
         sys = System(cryst, [1 => Moment(s=1, g=1)], mode)
-        enable_dipole_dipole!(sys, 1.0; demag=1/3)
+        enable_dipole_dipole!(sys, 1.0; demag=0)
 
         polarize_spins!(sys, (0,0,1))
         @test energy_per_site(sys) ≈ -0.1913132980155851
@@ -313,7 +313,7 @@ end
         units = Units(:meV, :angstrom)
         cryst = Sunny.bcc_crystal()
         sys = System(cryst, [1 => Moment(s=1, g=2)], :dipole, seed=2)
-        enable_dipole_dipole!(sys, units.vacuum_permeability; demag=1/3)
+        enable_dipole_dipole!(sys, units.vacuum_permeability; demag=0)
         polarize_spins!(sys, (1,2,3)) # arbitrary direction
 
         R = hcat([1,1,-1], [-1,1,1], [1,-1,1]) / 2

[…]
