+++ examples/07_Dipole_Dipole.jl
@@ -32,7 +32,7 @@ set_exchange!(sys, J1, Bond(1, 2, [0,0,0]))
 # using Ewald summation.
 
 sys_lr = clone_system(sys)
-enable_dipole_dipole!(sys_lr, units.vacuum_permeability; demag=1/3)
+enable_dipole_dipole!(sys_lr, units.vacuum_permeability)
 
 # Create a copy of the system and add long-range dipole-dipole interactions up
 # to a 5 Å cutoff distance.
@@ -75,8 +75,7 @@ res3 = intensities_bands(swt, path);
 # Create a panel corresponding to Fig. 2 of [Del Maestro and
 # Gingras](https://arxiv.org/abs/cond-mat/0403494). Dashed lines show the effect
 # of truncating dipole-dipole interactions at 5 Å. The Del Maestro and Gingras
-# paper underreported the energy scale by a factor of two, and requires slight
-# corrections to its third dispersion band.
+# paper underreported the energy scale by a factor of two.
 
 fig = Figure(size=(768, 300))
 plot_intensities!(fig[1, 1], res1; units, ylims=(0, 4), title="Without long-range dipole")
