+++ test/test_units.jl
@@ -5,7 +5,7 @@
         randomize_spins!(sys)
         set_exchange!(sys, 2 * units.K, Bond(1, 2, [0, 0, 0]))
         set_field!(sys, [0, 0, 1] * units.T)
-        enable_dipole_dipole!(sys, units.vacuum_permeability; demag=1/3)
+        enable_dipole_dipole!(sys, units.vacuum_permeability)
         E = energy(sys) / units.meV
         ∇E = Sunny.energy_grad_dipoles(sys) / units.THz
         return (E, ∇E)
