### 2025-05-23 #8138 @phibeck
Hi @jabirali thank you for the update regarding your timeline. While we aim for reviews to be completed in a time frame of 4-6 weeks, the average number of days in review is still around 3-4 months per submission (see https://www.theoj.org/joss-analytics/joss-submission-analytics.html#Pre-reviewreview_time_over_time). Given that it might take 2-4 weeks to recruit another reviewer and factoring in a possible vacation break during summer, the effective speed-up of recruiting another reviewer might end up marginal. So unless there's a strong incentive to have the publication out by then I would be fine with this timeline and keep you as a reviewer (unless you prefer to be replaced). Since JOSS reviews are incremental, there's always a chance you find time in between to cross off another point of your checklist.

@ddahlbom does this sound reasonable to you or do you prefer we find another reviewer in the meantime?
