### 2025-05-26 53d1c38
Revise tutorial 08
Fix analytical solutions and clarify presentation.
