### 2025-05-29 957a954
Replace shadowed variable with assertion check
Signed-off-by: Steven Hahn <hahnse@ornl.gov>
