### 2025-06-07 c2f8bd5
Switch to Ising convention
