+++ examples/08_Momentum_Conventions.jl
@@ -1,8 +1,8 @@
 # # 8. Momentum transfer conventions
 #
-# Sunny defines the dynamical spin structure factor following the sign
-# conventions in [Squire](https://doi.org/10.1017/CBO9781139107808) and
-# [Boothroyd](https://groups.physics.ox.ac.uk/Boothroyd/PNS/) textbooks,
+# Sunny defines the dynamical spin structure factor following conventions such
+# as in [Squire](https://doi.org/10.1017/CBO9781139107808) and
+# [Boothroyd](https://groups.physics.ox.ac.uk/Boothroyd/PNS/).
 #
 # ```math
 # \mathcal{S}^{α, β}(𝐪, ω) ≡ \frac{1}{2π} \int_{-∞}^{∞} e^{-iωt} ⟨\hat{M}^{†α}_𝐪(0) \hat{M}^β_𝐪(t)⟩ dt.
@@ -22,11 +22,10 @@
 #
 # With appropriate contraction of spin components, ``\mathcal{S}^{α, β}(𝐪, ω)``
 # can be directly related to the neutron scattering cross-section, where ``𝐪``
-# and ``ω`` denote momentum and energy transfer _to_ the sample. Special care is
-# required for models that lack inversion symmetry, in which case ``± 𝐪``
-# become inequivalent. We illustrate such a case using a 1D chain with
-# Dzyaloshinskii–Moriya coupling between neighboring sites. With an external
-# field, the inequivalence of ``\mathcal{S}(±𝐪,ω)`` becomes apparent.
+# and ``ω`` denote momentum and energy transfer _to_ the sample. For models that
+# lack inversion symmetry, the intensities at ``± 𝐪`` may become inequivalent.
+# We illustrate such a case using a 1D chain with competing Ising and
+# Dzyaloshinskii–Moriya couplings between neighboring sites.
 
 using Sunny, GLMakie
 
@@ -40,44 +39,38 @@ latvecs = lattice_vectors(2, 2, 1, 90, 90, 90)
 cryst = Crystal(latvecs, [[0, 0, 0]], "P1")
 
 # Consider a 1D chain oriented along the third lattice vector, such that site
-# ``j`` has position ``𝐫_j = j 𝐚_3``. The Hamiltonian includes DM coupling
-# between nearest neighbors and Zeeman coupling,
+# ``j`` has position ``𝐫_j = j 𝐚_3``. The Hamiltonian includes DM and
+# Ising-like couplings between nearest neighbors on the chain,
 #
 # ```math
-# ℋ = 𝐃 ⋅ ∑_j 𝐒_j × 𝐒_{j+1} - 𝐁 ⋅ ∑_j μ_j.
+# ℋ = 𝐃 ⋅ ∑_j 𝐒_j × 𝐒_{j+1} - J ∑_j ∑_j S^z_j S^z_{j+1}.
 # ```
-# Select DM vector ``𝐃 = D ê`` and external field ``𝐁 = B ê`` in an
-# arbitrary direction ``ê``. This field couples to each dimensionless
-# [`magnetic_moment`](@ref) ``μ_j = - g 𝐒_j``. The convention of Sunny is to
-# absorb the ``μ_B`` factor into ``𝐁``, giving it units of energy. Fix ``g =
-# -2`` for simplicity; other values could be viewed as a rescaling of ``𝐁``.
+# Select the DM vector ``𝐃 = D ẑ`.
 
-s = 2
+s = 3/2
 sys = System(cryst, [1 => Moment(; s, g=-2)], :dipole)
-B = 1.0
-D = 0.1
-e = [1, 0, 0]
[…]
