### 2025-06-07 7cd080d
Check reality of symbolic spin polynomial coefficients
Any imaginary part would correspond to a non-Hermitian operator, which is unphsyical. The code below will now report a similar "not Hermitian" error independent of mode:

```
set_onsite_coupling!(sys, S -> im * S[3]^2, 1)
```

If `S` is a finite-dimensional matrix representation, then one can use `hermitianpart`. If `S` is an infinite dimensional symbolic representation,, then it is currently necessary to use `real` instead, which maps the spin polynomial coefficients.. See also https://github.com/JuliaAlgebra/DynamicPolynomials.jl/issues/
[…]
