### 2025-06-08 c90e0a9
Unify notation for iterating over flattened "interactions" array
