+++ src/System/System.jl
@@ -73,6 +73,7 @@ function System(crystal::Crystal, moments::Vector{Pair{Int, Moment}}, mode::Symb
     gs = reshape(gs, 1, 1, 1, :)
 
     interactions = empty_interactions(mode, na, N)
+    param_map = Dict{String, ParamGroup}()
     ewald = nothing
 
     extfield = zeros(Vec3, 1, 1, 1, na)
@@ -83,7 +84,7 @@ function System(crystal::Crystal, moments::Vector{Pair{Int, Moment}}, mode::Symb
 
     rng = isnothing(seed) ? Random.Xoshiro(rand(UInt64, 4)...) : Random.Xoshiro(seed)
 
-    ret = System(nothing, mode, crystal, (1, 1, 1), Ns, κs, gs, interactions, ewald,
+    ret = System(nothing, mode, crystal, (1, 1, 1), Ns, κs, gs, interactions, param_map, ewald,
                  extfield, dipoles, coherents, dipole_buffers, coherent_buffers, rng)
     polarize_spins!(ret, (0,0,1))
     return dims == (1, 1, 1) ? ret : repeat_periodically(ret, dims)
@@ -141,7 +142,8 @@ Creates a full clone of the system, such that mutable updates to one copy will
 not affect the other, and thread safety is guaranteed.
 """
 function clone_system(sys::System{N}) where N
-    (; origin, mode, crystal, dims, Ns, gs, κs, extfield, interactions_union, ewald, dipoles, coherents, rng) = sys
+    (; origin, mode, crystal, dims, Ns, gs, κs, extfield, interactions_union,
+        param_map, ewald, dipoles, coherents, rng) = sys
 
     origin_clone = isnothing(origin) ? nothing : clone_system(origin)
     ewald_clone = nothing # TODO: use clone_ewald(ewald)
@@ -149,14 +151,16 @@ function clone_system(sys::System{N}) where N
     # Dynamically dispatch to the correct `map` function for either homogeneous
     # (Vector) or inhomogeneous interactions (4D Array)
     interactions_clone = map(clone_interactions, interactions_union)
-    
+    param_map_clone = Dict{String, ParamGroup}((k, copy(v)) for (k, v) in param_map)
+
     # Empty buffers are required for thread safety.
     empty_dipole_buffers = Array{Vec3, 4}[]
     empty_coherent_buffers = Array{CVec{N}, 4}[]
 
     ret = System(origin_clone, mode, crystal, dims, Ns, copy(κs), copy(gs),
-                 interactions_clone, ewald_clone, copy(extfield), copy(dipoles), copy(coherents),
-                 empty_dipole_buffers, empty_coherent_buffers, copy(rng))
+                 interactions_clone, param_map_clone, ewald_clone, copy(extfield),
+                 copy(dipoles), copy(coherents), empty_dipole_buffers,
+                 empty_coherent_buffers, copy(rng))
 
     if !isnothing(ewald)
         # At the moment, clone_ewald is unavailable, so instead rebuild the
