### 2025-06-10 65280bd
Fix spiral calculations on reshaped systems (#382)
The ordering wavevector k must always be interpreted in RLU for the original cell. This can be relevant for calculations performed on a primitive cell.
