+++ src/Spiral/SpinWaveTheorySpiral.jl
@@ -3,11 +3,11 @@
 
 Analogous to [`SpinWaveTheory`](@ref), but interprets the provided system as
 having a generalized spiral order. This order is described by a single
-propagation wavevector `k`, which may be incommensurate. The `axis` vector
-defines the polarization plane via its surface normal. Typically the spin
-configuration in `sys` and the propagation wavevector `k` will be optimized
-using [`minimize_spiral_energy!`](@ref). In contrast, `axis` will typically be
-determined from symmetry considerations.
+propagation wavevector `k` (in reciprocal lattice units, RLU) an `axis` vector
+that is normal to the polarization plane (in global Cartesian coordinates).
+Typically the spin configuration in `sys` and the propagation wavevector `k`
+will be optimized using [`minimize_spiral_energy!`](@ref). In contrast, `axis`
+will typically be determined from symmetry considerations.
 
 The resulting object can be used to calculate the spin wave
 [`dispersion`](@ref), or the structure factor via [`intensities_bands`](@ref)
@@ -19,8 +19,9 @@ implemented in the [SpinW code](https://spinw.org/).
 """
 struct SpinWaveTheorySpiral <: AbstractSpinWaveTheory
     swt :: SpinWaveTheory
-    k :: Vec3
-    axis :: Vec3
+    k :: Vec3              # Propagation wavevector
+    k_case :: Int          # Label (1, 2, 3) if (kᵢ ∈ ℕ, 2kᵢ ∈ ℕ, otherwise)
+    axis :: Vec3           # Axis for polarization plane
     buffers :: Vector{Array{CMat3, 2}}
 
     function SpinWaveTheorySpiral(sys::System; k::AbstractVector, axis::AbstractVector, measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
@@ -28,9 +29,10 @@ struct SpinWaveTheorySpiral <: AbstractSpinWaveTheory
             error("SpinWaveTheorySpiral requires :dipole or :dipole_uncorrected mode.")
         end
 
+        k_case = spiral_propagation_case(to_reshaped_rlu(sys, k))
         L = length(eachsite(sys))
         buffers = [zeros(CMat3, L, L) for _ in 1:6]
-        return new(SpinWaveTheory(sys; measure, regularization), k, normalize(axis), buffers)
+        return new(SpinWaveTheory(sys; measure, regularization), k, k_case, normalize(axis), buffers)
     end
 end
 
@@ -49,7 +51,7 @@ end
 # Niraula and Xiaojian Bai developed and implemented the approprate
 # generalization in https://github.com/SunnySuite/Sunny.jl/pull/304.
 
-function fourier_bilinear_interaction!(J_k, swt::SpinWaveTheory, q)
+function fourier_bilinear_interaction!(J_k, swt::SpinWaveTheory, q_reshaped)
     (; sys, data) = swt
     (; local_rotations) = data
     (; gs) = sys
@@ -65,14 +67,14 @@ function fourier_bilinear_interaction!(J_k, swt::SpinWaveTheory, q)
 
             (; j, n) = bond
             J_lab = Rs[i] * Mat3(bilin*I) * Rs[j]' # Undo transformation in `swt_data`
-            J = exp(-2π * im * dot(q, n)) * J_lab
+            J = exp(-2π * im * dot(q_reshaped, n)) * J_lab
[…]
