+++ src/System/Interactions.jl
@@ -62,7 +62,7 @@ end
 
 
 """
-    enable_dipole_dipole!(sys::System, μ0_μB²)
+    enable_dipole_dipole!(sys::System, μ0_μB²; demag=1/3)
 
 Enables long-range interactions between magnetic dipole moments,
 
@@ -71,11 +71,24 @@ Enables long-range interactions between magnetic dipole moments,
 ```
 
 where the sum is over all pairs of sites (singly counted), including periodic
-images, regularized using the Ewald summation convention. Each magnetic moment
-is ``μ = -g μ_B 𝐒``, where ``𝐒`` is the spin angular momentum dipole. The
-parameter `μ0_μB²` specifies the physical constant ``μ_0 μ_B^2``, which has
-dimensions of length³-energy. Obtain this constant for a given system of
-[`Units`](@ref) via its `vacuum_permeability` property.
+images. Each magnetic moment is ``μ_i = -g μ_B 𝐒_i``, where ``𝐒_i`` is the
+spin angular momentum dipole. The parameter `μ0_μB²` specifies the physical
+constant ``μ_0 μ_B^2``, which has dimensions of length³-energy. Obtain this
+constant for a given system of [`Units`](@ref) via its `vacuum_permeability`
+property.
+
+Geometry of the macroscopic sample enters through the demagnetization factor
+tensor ``ℕ``, denoted `demag`. In a vacuum background, `tr(demag) == 1`. Special
+cases are:
+
+  * `demag = 1/3` for isotropic demagnetization. This is the default and is
+    valid for sphere and cube sample geometries.
+  * `demag = Diagonal([0, 0, 1])`` for a sheet-like geometry with surface normal
+    in ``ẑ``.
+  * `demag = Diagonal([1/2, 1/2, 0])` for a needle-like geometry aligned with
+    `ẑ``.
+
+Set `demag = 0` to neglect demagnetization effects.
 
 # Example
 
@@ -84,25 +97,51 @@ units = Units(:meV, :angstrom)
 enable_dipole_dipole!(sys, units.vacuum_permeability)
 ```
 
+!!! tip "Origin of demagnetization"  
+
+    Formal summation over the infinitely many dipole-dipole pair interactions
+    becomes mathematically ambiguous when the sample has a net magnetic dipole,
+    ```math
+        𝐌 = ∑_i μ_i.
+    ```
+
+    The traditional Ewald method resolves this ambiguity by neglecting surface
+    effects that would lead to demagnetization. For physical correctness, however,
+    the Ewald energy should be augmented with a surface contribution to the total
+    energy,
+    ```math
+        E_s = μ_0 𝐌⋅ℕ 𝐌 / 2V,
+    ```
+
+    where ``ℕ`` is the demagnetization factor tensor. Assuming vacuum background,
+    it can be expressed as an integral over the sample volume,
+    ```math
+        ℕ = - (1 / 4π) ∫ d𝐱 ∇ ∇ (1 / |𝐱|).
+    ```
+
+    As constructed, ``ℕ`` has trace 1. If the sample is embedded in another
+    material with relative permeability ``μ' > 1``, then ``ℕ`` may be reduced. For
+    example, a spherical inclusion has ``ℕ = 1/(2μ'+1)`` in the general case.
+
 !!! tip "Efficiency considerations"  
-    Dipole-dipole interactions are very efficient in the context of spin
-    dynamics simulation, e.g. [`Langevin`](@ref). Sunny applies the fast Fourier
-    transform (FFT) to spins on each Bravais sublattice, such that the
-    computational cost to integrate one time-step scales like ``M^2 N \\ln N``,
-    where ``N`` is the number of cells in the system and ``M`` is the number of
[…]
