### 2025-06-10 aa4ff23
Rename `set_spin_rescaling_for_static_sum_rule!`
