+++ docs/src/library.md
@@ -36,7 +36,7 @@ dmvec
 domain_average
 eachsite
 enable_dipole_dipole!
-enable_spin_rescaling_for_static_sum_rule!
+set_spin_rescaling_for_static_sum_rule!
 energy
 energy_per_site
 excitations
