### 2025-06-11 65730a6
Rename `set_spin_rescaling_for_static_sum_rule!` (#389)
