+++ src/Sunny.jl
@@ -59,9 +59,9 @@ include("System/PairExchange.jl")
 include("System/OnsiteCoupling.jl")
 include("System/Ewald.jl")
 include("System/Interactions.jl")
-export Moment, System, Site, clone_system, eachsite, position_to_site, global_position, magnetic_moment,
-    set_coherent!, set_dipole!, polarize_spins!, randomize_spins!, set_spin_rescaling!,
-    enable_spin_rescaling_for_static_sum_rule!, energy, energy_per_site,
+export Moment, System, Site, clone_system, eachsite, position_to_site, global_position,
+    magnetic_moment, set_coherent!, set_dipole!, polarize_spins!, randomize_spins!,
+    set_spin_rescaling!, set_spin_rescaling_for_static_sum_rule!, energy, energy_per_site,
     spin_label, set_onsite_coupling!, set_pair_coupling!, set_exchange!, dmvec, enable_dipole_dipole!,
     set_field!, to_inhomogeneous, set_field_at!, set_vacancy_at!, set_onsite_coupling_at!,
     set_exchange_at!, set_pair_coupling_at!, symmetry_equivalent_bonds, remove_periodicity!,
