### 2025-06-11 7dd815c
Rename `set_spin_rescaling_for_static_sum_rule!` (#389)
