+++ docs/src/versions.md
@@ -3,9 +3,8 @@
 ## v0.7.7
 (In development)
 
-* Add [`enable_spin_rescaling_for_static_sum_rule!`](@ref) which sets the
-  classical dipole magnitude to ``\sqrt{s (s + 1)}`` for each quantum spin-``s``
-  moment.
+* Add [`set_spin_rescaling_for_static_sum_rule!`](@ref) which sets the classical
+  dipole magnitude to ``\sqrt{s (s + 1)}`` for each quantum spin-``s`` moment.
 * Add module [`SCGA`](@ref) for calculating [`intensities_static`](@ref) within
   the self-consistent Gaussian approximation ([PR
   #355](https://github.com/SunnySuite/Sunny.jl/pull/355)).
