### 2025-06-11 05133a2
Use `int` singular to denote Interactions on one site
