### 2025-06-12 9086b48
CompatHelper: bump compat for WGLMakie in [weakdeps] to 0.12, (keep existing compat)
