### 2025-06-12 47d90f3
CompatHelper: bump compat for GLMakie in [weakdeps] to 0.12, (keep existing compat)
