+++ Project.toml
@@ -41,7 +41,7 @@ CrystalInfoFramework = "0.8, 1"
 DynamicPolynomials = "0.6.0"
 ElasticArrays = "1.2.12"
 FFTW = "1.4.5"
-GLMakie = "0.9, 0.10, 0.11"
+GLMakie = "0.9, 0.10, 0.11, 0.12"
 HCubature = "1.7.0"
 JLD2 = "0.5.4"
 LinearAlgebra = "1.10"
