### 2025-06-13 baed906
Updates for Makie v0.23
The new `Makie.arrows3d` interface requires significant changes in
Sunny, but overall the code becomes cleaner. After these changes, it is
not possible to maintain backwards compatibility with prior Makie
versions.
