+++ ext/PlottingExt/ViewCrystal.jl
@@ -103,7 +103,7 @@ function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
     ### Ellipsoids for symmetric exchanges
 
     # Dimensionless scalings and rotations associated with principle axes
-    decomps = exchange_decomposition.(scaled_exchanges)            
+    decomps = exchange_decomposition.(scaled_exchanges)
     scalings = map(x -> x[1], decomps)
     rotation = map(x -> x[2], decomps)
 
@@ -148,26 +148,52 @@ function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
              dim == 3 ? 1 : c]
         end
         markersize2 = [ms .* rs for (ms, rs) in zip(markersize, rescalings)]
-    
+
         o = Makie.meshscatter!(ax, pts; color, markersize=markersize2, rotation, marker=cylinders[dim], inspectable=false)
-        Makie.connect!(o.visible, obs)            
+        Makie.connect!(o.visible, obs)
     end
 
-    ### Cones for DM vectors. Because they tend to be weaker in magnitude,
-    ### we apply some heuristic amplification to the arrow size.
+    ### Cones for DM vectors
 
+    # DM vectors normalized by the overall exchange scale
     dmvecs = Sunny.extract_dmvec.(scaled_exchanges)
-    dirs = @. Makie.Vec3f(normalize(dmvecs))
+
+    # Filter nonzero DM vectors and retain associated exchange scalings
+    nonzeros_indices = findall(norm(dmvec) > 0 for dmvec in dmvecs)
+    dm_pts = pts[nonzeros_indices]
+    dmvecs = dmvecs[nonzeros_indices]
+    scalings_nonzero = scalings[nonzeros_indices]
+
     # The largest possible ellipsoid occurs in the case of `scalings ==
     # [1,1,1]`, yielding a sphere with size `ionradius`.
-    ellipsoid_radii = @. ionradius * norm(scalings) / √3
-    arrowsize = @. 2ionradius * cbrt(norm(dmvecs)) # size of arrow head
-    dm_pts = @. pts + 1.1ellipsoid_radii * dirs
-    o = Makie.arrows!(ax, dm_pts, dirs; lengthscale=0, arrowsize, diffuse=1.15, color=:magenta, specular=0.0, inspectable=false) 
+    ellipsoid_radii = @. ionradius * norm(scalings_nonzero) / √3
+
+    # Size of cone scales like cube root of DM vector magnitude
+    markersize = @. 2ionradius * cbrt(norm(dmvecs))
+
+    # Offset cone according to size of exchange ellipsoid
+    dm_pts = @. Makie.Point3f(dm_pts + 1.1ellipsoid_radii * normalize(dmvecs))
+
+    # Draw cones
+    marker = Makie.GeometryBasics.Cone(Makie.Point3f(0, 0, 0), Makie.Point3f(0, 0, 1), 0.5)
+    o = Makie.meshscatter!(ax, dm_pts; marker, markersize, rotation=dmvecs, color=:magenta, specular=0.0, diffuse=1.15, inspectable=false)
     Makie.connect!(o.visible, obs)
 end
 
-function draw_bonds(; ax, obs, ionradius, exchange_mag, cryst, interactions, bonds, refbonds, color)
+function draw_bonds(; ax, obs, ionradius, exchange_mag, cryst, interactions, bonds, refbond, color)
+    # The bond is directed if it allows for a DM exchange. This will be
+    # visualized with an arrow.
+    isdirected = begin
+        basis = Sunny.basis_for_symmetry_allowed_couplings(cryst, refbond)
+        any(J -> J ≈ -J', basis)
+    end
+
+    if isempty(bonds)
+        bonds = [refbond]
+        show_ghosts = false
+    else
[…]
