### 2025-06-15 b037dcb
Use `int` singular to denote Interactions on one site
