### 2025-06-16 24f93dc
Replace shadowed variable with assertion check (#387)
Also unifies notation for iterating over flattened interactions lists.

---------

Signed-off-by: Steven Hahn <hahnse@ornl.gov>
Co-authored-by: Kipton Barros <kbarros@gmail.com>
