### 2025-06-16 7cbd976
Demagnetization tensor in Ewald summation
Extend `enable_dipole_dipole!` to accept a demagnetization factor tensor
`demag`. The new default is isotropic demagnetization, `demag = 1/3`,
appropriate for a spherical sample in vacuum. Set `demag = 0` to disable
demagnetization.
