+++ src/Reshaping.jl
@@ -106,7 +106,7 @@ function reshape_supercell_aux(sys::System{N}, new_cryst::Crystal, new_dims::NTu
     # Restore dipole-dipole interactions if present. This involves pre-computing
     # an interaction matrix that depends on `new_dims`.
     if !isnothing(sys.ewald)
-        enable_dipole_dipole!(new_sys, sys.ewald.μ0_μB²)
+        enable_dipole_dipole!(new_sys, sys.ewald.μ0_μB²; sys.ewald.demag)
     end
 
     return new_sys
