### 2025-06-16 62d890d
Tweak docstring
