### 2025-06-17 7b38b1d
Require latest Optim and increase robustness
