+++ src/System/Interactions.jl
@@ -77,8 +77,8 @@ constant ``μ_0 μ_B^2``, which has dimensions of length³-energy. Obtain this
 constant for a given system of [`Units`](@ref) via its `vacuum_permeability`
 property.
 
-Geometry of the macroscopic sample enters through the demagnetization factor
-tensor ``ℕ``, denoted `demag`. Special cases are:
+Geometry of the macroscopic sample enters through the demagnetization factor or
+tensor `demag`. Special cases are:
 
   * `demag = 1/3` for isotropic demagnetization. This is the default and is
     valid for sphere and cube sample geometries.
@@ -102,41 +102,43 @@ See also [`modify_exchange_with_truncated_dipole_dipole!`](@ref).
 !!! tip "Demagnetization details"  
 
     Formal summation over the infinitely many dipole-dipole pair interactions
-    becomes mathematically ambiguous when the sample has a nonzero net magnetic
-    moment, ``𝐌 = ∑_i μ_i``. The traditional Ewald method resolves this ambiguity
-    by neglecting surface effects that would lead to demagnetization. For physical
-    correctness, however, the Ewald energy must be augmented with a surface energy
-    correction,
+    becomes mathematically ambiguous when the macroscopic sample has a nonzero net
+    magnetic moment, ``𝐌 = ∑_i μ_i``. The traditional Ewald method resolves this
+    ambiguity by neglecting surface effects that would lead to demagnetization. For
+    physical correctness, however, the Ewald energy must be augmented with a surface
+    energy term,
     ```math
-        E_s = \\frac{μ_0}{2V} 𝐌⋅ℕ 𝐌,
+        E_\\mathrm{surf} = \\frac{μ_0}{2V} 𝐌⋅\\mathcal{N} 𝐌,
     ```
-    where ``ℕ`` is the demagnetization factor tensor. Assuming vacuum background, it
-    can be expressed as an integral over the sample volume ``V``,
+    where ``\\mathcal{N}`` is the demagnetization tensor (`demag`). Assuming vacuum
+    background, it can be expressed as an integral over the sample volume ``V``,
     ```math
-        ℕ = - \\frac{1}{4π} ∫_V d𝐱 ∇ ∇ |𝐱|^{-1}.
+        \\mathcal{N}^{αβ} = - \\frac{1}{4π} ∫_V d𝐱 ∇^α ∇^β |𝐱|^{-1}.
     ```
-    Note that ``ℕ`` has trace 1 because ``∇^2|𝐱|^{-1} = -4πδ(𝐱)``.
+    Note that ``\\mathcal{N}`` has trace 1 because ``∇^2|𝐱|^{-1} = -4πδ(𝐱)``
+    when the integration domain contains the origin.
 
     This surface correction to the Ewald energy originally appeared in S. de Leeuw,
     J. Perram, and E. Smith, Proc. R. Soc. London A **373**, 27 (1980); **373**, 57
     (1980); **388**, 177 (1983). For a pedagogical review, see [V. Ballenegger, J.
     Chem. Phys. **140**, 161102 (2014)](https://doi.org/10.1063/1.4872019).
 
-    If the sample is embedded in another material, the surface correction ``E_s``
-    remains valid, but ``ℕ`` should be calculated differently. For example, a
-    spherical inclusion generally has ``ℕ = 1/(2μ'+1) ≤ 1/3`` where ``μ' ≥ 1``
-    denotes the relative permeability of the background medium.
+    If the sample is embedded in another material, the surface correction
+    ``E_\\mathrm{surf}`` still applies, but ``\\mathcal{N}`` should be calculated
+    differently. For example, a spherical inclusion generally has ``\\mathcal{N} =
[…]
