+++ examples/08_Momentum_Conventions.jl
@@ -5,7 +5,7 @@
 # [Boothroyd](https://groups.physics.ox.ac.uk/Boothroyd/PNS/),
 #
 # ```math
-# \mathcal{S}^{α, β}(𝐪, ω) ≡ \frac{1}{2π} \int_{-∞}^{∞} e^{-iωt} ⟨\hat{M}^{†α}_𝐪(0) \hat{M}^β_𝐪(t)⟩ dt.
+# \mathcal{S}^{αβ}(𝐪, ω) ≡ \frac{1}{2π} \int_{-∞}^{∞} e^{-iωt} ⟨\hat{M}^{†α}_𝐪(0) \hat{M}^β_𝐪(t)⟩ dt.
 # ```
 # 
 # The magnetic moment in momentum space ``\hat{𝐌}_𝐪`` is obtained from the
@@ -20,30 +20,32 @@
 # by the number of chemical cells in the sample. For full details, see the
 # documentation page [Structure Factor Conventions](@ref).
 #
-# With appropriate contraction of spin components, ``\mathcal{S}^{α, β}(𝐪, ω)``
-# can be directly related to the neutron scattering cross-section, where ``𝐪``
-# and ``ω`` denote momentum and energy transfer _to_ the sample. For models that
-# lack inversion symmetry, the intensities at ``±𝐪`` may be inequivalent. We
-# illustrate such a case using a 1D chain with competing Ising and
+# With appropriate contraction of spin components, ``\mathcal{S}^{αβ}(𝐪, ω)``
+# can be directly related to the neutron scattering cross-section with ``𝐪``
+# and ``ω`` denoting momentum and energy transfer **to** the sample. For models
+# that lack inversion symmetry, the intensities at ``±𝐪`` may be inequivalent.
+# We illustrate such a case using a 1D chain with competing Ising and
 # Dzyaloshinskii–Moriya couplings between neighboring sites.
 #
-# Be aware that other codes (e.g. [SpinW](https://spinw.org/)) may employ an
-# alternative sign convention that effectively reverses the direction of
-# momentum transfer, ``𝐪 → -𝐪``.
+# Be aware that other codes, e.g. [SpinW](https://spinw.org/), may employ a
+# different Fourier transform convention that effectively reverses the direction
+# of momentum transfer, ``𝐪 → -𝐪``.
+
+# ### Build a 1D model without reflection symmetry
 
 using Sunny, GLMakie
 
-# Selecting the P1 spacegroup will effectively disable all symmetry analysis.
-# This can be a convenient way to avoid symmetry-imposed constraints on the
-# couplings. With this choice, all bonds within the unit cell are treated as
-# symmetry-inequivalent and must be specified independently.
+# The model will live on a 1D chain. Select the spacegroup P1 to effectively
+# disable all symmetry analysis. This eliminates all symmetry-imposed
+# constraints on the couplings. Furthermore, because all bonds become
+# symmetry-inequivalent, each coupling must be specified independently (there is
+# no automatic propagation to "equivalent" bonds).
 
-latvecs = lattice_vectors(2, 2, 1, 90, 90, 90)
+latvecs = lattice_vectors(1, 1, 1, 90, 90, 90)
 cryst = Crystal(latvecs, [[0, 0, 0]], "P1")
 
-# Consider a 1D chain oriented along the lattice vector ``𝐚_3`` such that site
-# ``j`` has position ``𝐫_j = j 𝐚_3``. The Hamiltonian includes DM and
-# Ising-like couplings between nearest neighbors on the chain,
[…]
