### 2025-06-20 3971b28
Dipole-dipole interactions incorporate demagnetization factor tensor (#385)
Extend `enable_dipole_dipole!` to accept a demagnetization factor tensor
`demag`. The new default is isotropic demagnetization, `demag = 1/3`,
appropriate for a spherical sample in vacuum. Set `demag = 0` to disable
demagnetization.
