+++ src/System/Interactions.jl
@@ -62,7 +62,7 @@ end
 
 
 """
-    enable_dipole_dipole!(sys::System, μ0_μB²)
+    enable_dipole_dipole!(sys::System, μ0_μB²; demag=1/3)
 
 Enables long-range interactions between magnetic dipole moments,
 
@@ -71,11 +71,24 @@ Enables long-range interactions between magnetic dipole moments,
 ```
 
 where the sum is over all pairs of sites (singly counted), including periodic
-images, regularized using the Ewald summation convention. Each magnetic moment
-is ``μ = -g μ_B 𝐒``, where ``𝐒`` is the spin angular momentum dipole. The
-parameter `μ0_μB²` specifies the physical constant ``μ_0 μ_B^2``, which has
-dimensions of length³-energy. Obtain this constant for a given system of
-[`Units`](@ref) via its `vacuum_permeability` property.
+images. Each magnetic moment is ``μ_i = -g μ_B 𝐒_i``, where ``𝐒_i`` is the
+spin angular momentum dipole. The parameter `μ0_μB²` specifies the physical
+constant ``μ_0 μ_B^2``, which has dimensions of length³-energy. Obtain this
+constant for a given system of [`Units`](@ref) via its `vacuum_permeability`
+property.
+
+Geometry of the macroscopic sample enters through the demagnetization factor or
+tensor `demag`. Special cases are:
+
+  * `demag = 1/3` for isotropic demagnetization. This is the default and is
+    valid for sphere and cube sample geometries.
+  * `demag = Diagonal([0, 0, 1])` for a sheet-like geometry with surface normal
+    in ``ẑ``.
+  * `demag = Diagonal([1/2, 1/2, 0])` for a needle-like geometry aligned with
+    ``ẑ``.
+
+In a vacuum background, the demagnetization tensor should have trace 1. Set
+`demag = 0` to artificially neglect demagnetization effects.
 
 # Example
 
@@ -84,25 +97,55 @@ units = Units(:meV, :angstrom)
 enable_dipole_dipole!(sys, units.vacuum_permeability)
 ```
 
-!!! tip "Efficiency considerations"  
-    Dipole-dipole interactions are very efficient in the context of spin
-    dynamics simulation, e.g. [`Langevin`](@ref). Sunny applies the fast Fourier
-    transform (FFT) to spins on each Bravais sublattice, such that the
-    computational cost to integrate one time-step scales like ``M^2 N \\ln N``,
-    where ``N`` is the number of cells in the system and ``M`` is the number of
-    Bravais sublattices per cell. Conversely, dipole-dipole interactions are
-    highly _inefficient_ in the context of a [`LocalSampler`](@ref). Each Monte
-    Carlo update of a single spin currently requires scanning over all other
-    spins in the system.
-
 See also [`modify_exchange_with_truncated_dipole_dipole!`](@ref).
+
+!!! tip "Demagnetization details"  
+
+    Formal summation over the infinitely many dipole-dipole pair interactions
+    becomes mathematically ambiguous when the macroscopic sample has a nonzero net
+    magnetic moment, ``𝐌 = ∑_i μ_i``. The traditional Ewald method resolves this
+    ambiguity by neglecting surface effects that would lead to demagnetization. For
+    physical correctness, however, the Ewald energy must be augmented with a surface
+    energy term,
+    ```math
+        E_\\mathrm{surf} = \\frac{μ_0}{2V} 𝐌⋅\\mathcal{N} 𝐌,
+    ```
+    where ``\\mathcal{N}`` is the demagnetization tensor (`demag`). Assuming vacuum
[…]
