+++ examples/08_Momentum_Conventions.jl
@@ -27,8 +27,9 @@
 # illustrate such a case using a 1D chain with competing Ising and
 # Dzyaloshinskii–Moriya couplings between neighboring sites.
 #
-# This example serves as a point of comparison between codes. For example,
-# [SpinW](https://spinw.org/) employs the opposite sign convention for ``𝐪``.
+# This model may serve as a point of comparison between codes. For example,
+# [SpinW](https://spinw.org/) employs the opposite sign convention for momentum
+# transfer, which effectively relabels ``𝐪 → -𝐪``.
 
 # ### 1D model lacking reflection symmetry
 
