### 2025-06-22 8c7e628
CompatHelper: bump compat for WGLMakie in [weakdeps] to 0.13, (keep existing compat)
