+++ Project.toml
@@ -56,6 +56,6 @@ SpecialFunctions = "2.1.0"
 Spglib = "0.9.4"
 StaticArrays = "1.9.1"
 Statistics = "1.10"
-WGLMakie = "0.9, 0.10, 0.11"
+WGLMakie = "0.9, 0.10, 0.11, 0.13"
 WriteVTK = "1.18.0"
 julia = "1.10"
