### 2025-06-22 f98a350
CompatHelper: bump compat for Makie in [weakdeps] to 0.24, (keep existing compat)
