+++ Project.toml
@@ -45,7 +45,7 @@ GLMakie = "0.9, 0.10, 0.11"
 HCubature = "1.7.0"
 JLD2 = "0.5.4"
 LinearAlgebra = "1.10"
-Makie = "0.20, 0.21.1, 0.22"
+Makie = "0.20, 0.21.1, 0.22, 0.24"
 OffsetArrays = "1.10.8"
 Optim = "1.7.6"
 PrecompileTools = "1.2.0"
