+++ examples/08_Momentum_Conventions.jl
@@ -24,10 +24,9 @@
 #
 # For models that lack inversion symmetry, the intensities at ``±𝐪`` may be
 # inequivalent. A simple example is the 1D chain with competing Ising and
-# Dzyaloshinskii–Moriya couplings between neighboring sites. This model may
-# serve as a point of comparison between Sunny and other codes. For example,
-# [SpinW](https://spinw.org/) employs the opposite sign convention for momentum
-# transfer ``𝐪``.
+# Dzyaloshinskii–Moriya couplings. This model may serve as a point of comparison
+# between Sunny and other codes. For example, [SpinW](https://spinw.org/)
+# employs the opposite sign convention for momentum transfer ``𝐪``.
 
 # ### 1D model lacking reflection symmetry
 
