+++ Project.toml
@@ -41,11 +41,11 @@ CrystalInfoFramework = "0.8, 1"
 DynamicPolynomials = "0.6.0"
 ElasticArrays = "1.2.12"
 FFTW = "1.4.5"
-GLMakie = "0.12"
+GLMakie = "0.13"
 HCubature = "1.7.0"
 JLD2 = "0.5.4"
 LinearAlgebra = "1.10"
-Makie = "0.23"
+Makie = "0.24"
 OffsetArrays = "1.10.8"
 Optim = "1.13.0"
 PrecompileTools = "1.2.0"
@@ -56,6 +56,6 @@ SpecialFunctions = "2.1.0"
 Spglib = "0.9.4"
 StaticArrays = "1.9.1"
 Statistics = "1.10"
-WGLMakie = "0.12"
+WGLMakie = "0.13"
 WriteVTK = "1.18.0"
 julia = "1.10"
