+++ ext/PlottingExt/ViewCrystal.jl
@@ -98,7 +98,7 @@ function exchange_decomposition(J)
     return (vals, q)
 end
 
-function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
+function draw_exchange_geometries(; ax, visible, ionradius, pts, scaled_exchanges)
 
     ### Ellipsoids for symmetric exchanges
 
@@ -126,8 +126,7 @@ function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
         d = c+(1-c)*abs(y) # c ≤ d ≤ 1
         y > 0 ? Makie.RGBf(c, c, d) : Makie.RGBf(d, c, c)
     end
-    o = Makie.meshscatter!(ax, pts; color, markersize, rotation, specular=0, diffuse=1.5, inspectable=false)
-    Makie.connect!(o.visible, obs)
+    Makie.meshscatter!(ax, pts; color, markersize, rotation, specular=0, diffuse=1.5, visible, inspectable=false)
 
     # Draw dots using cylinders
     cylinders = map(eachcol(Sunny.Mat3(I))) do x
@@ -149,8 +148,7 @@ function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
         end
         markersize2 = [ms .* rs for (ms, rs) in zip(markersize, rescalings)]
 
-        o = Makie.meshscatter!(ax, pts; color, markersize=markersize2, rotation, marker=cylinders[dim], inspectable=false)
-        Makie.connect!(o.visible, obs)
+        Makie.meshscatter!(ax, pts; color, markersize=markersize2, rotation, marker=cylinders[dim], visible, inspectable=false)
     end
 
     ### Cones for DM vectors
@@ -176,11 +174,11 @@ function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
 
     # Draw cones
     marker = Makie.GeometryBasics.Cone(Makie.Point3f(0, 0, 0), Makie.Point3f(0, 0, 1), 0.5)
-    o = Makie.meshscatter!(ax, dm_pts; marker, markersize, rotation=dmvecs, color=:magenta, specular=0.0, diffuse=1.15, inspectable=false)
-    Makie.connect!(o.visible, obs)
+    Makie.meshscatter!(ax, dm_pts; marker, markersize, rotation=dmvecs, color=:magenta,
+                       specular=0.0, diffuse=1.15, visible, inspectable=false)
 end
 
-function draw_bonds(; ax, obs, ionradius, exchange_mag, cryst, interactions, bonds, refbond, color)
+function draw_bonds(; ax, visible, ionradius, exchange_mag, cryst, interactions, bonds, refbond, color)
     # The bond is directed if it allows for a DM exchange. This will be
     # visualized with an arrow.
     isdirected = begin
@@ -207,14 +205,15 @@ function draw_bonds(; ax, obs, ionradius, exchange_mag, cryst, interactions, bon
         ghosts = findall(b -> !iszero(b.n), bonds)
 
         # Concatenate ghosts to the end of arrays
-        bonds = vcat(bonds, bonds[ghosts])
-        color = vcat(color, color[ghosts])
+        append!(bonds, bonds[ghosts])
+        append!(color, color[ghosts])
 
         # Ghost bonds are offset by -n multiples of lattice vectors
-        segments = vcat(segments, map(ghosts) do i
+        ghost_segments = map(ghosts) do i
             offset = - cryst.latvecs * bonds[i].n
-            segments[i] .+ Ref(offset)
-        end)
+            Makie.Point3f.(segments[i] .+ Ref(offset))
+        end
+        append!(segments, ghost_segments)
     end
 
     # String for each bond b′. Like print_bond(b′), but shorter.
[…]
