### 2025-06-23 88b88d4
Make energy minimization more reliable (#397)
Use CG instead of L-BFGS everywhere. This is possible because of the more robust Hager-Zhang line search introduced in Optim v1.13.

Define CG tolerance parameters such that convergence is reached only when updates to the dimensionless spin variables go below a threshold `x_abstol`, which is of order `1e-14`.
