### 2025-06-23 0eaaad7
Revise tutorial on momentum transfer conversions (#386)
Tutorial 08 now includes:
* A brief definition of the structure factor conventions employed by Sunny.
* The statement that Sunny and SpinW use opposite conventions for q.
* A model Hamiltonian that does not include an external field.
* Correction to the analytical expression of the dispersion.
