+++ examples/08_Momentum_Conventions.jl
@@ -1,92 +1,117 @@
 # # 8. Momentum transfer conventions
 #
-# This example illustrates Sunny's conventions for dynamical structure factor
-# intensities, ``\mathcal{S}(𝐪,ω)``, as documented in the page [Structure
-# Factor Conventions](@ref). The variables ``𝐪`` and ``ω`` describe momentum
-# and energy transfer _to_ the sample.
+# Sunny defines the dynamical structure factor following conventions as in
+# [Squire](https://doi.org/10.1017/CBO9781139107808) and
+# [Boothroyd](https://groups.physics.ox.ac.uk/Boothroyd/PNS/),
 #
-# The structure factor intensities at ``± 𝐪`` may be inequivalent if the model
-# lacks inversion symmetry. Consider, for example, a 1D chain that includes only
-# Dzyaloshinskii–Moriya interactions between neighboring sites. Coupling to an
-# external field then breaks time-reversal symmetry, giving rise to an
-# inequivalence of intensities ``\mathcal{S}(±𝐪,ω)``
+# ```math
+# \mathcal{S}^{αβ}(𝐪, ω) ≡ \frac{1}{2π} \int_{-∞}^{∞} e^{-iωt} ⟨\hat{M}^{†α}_𝐪(0) \hat{M}^β_𝐪(t)⟩ dt.
+# ```
+# 
+# The momentum-space dipole operator ``\hat{𝐌}_𝐪`` is obtained from the
+# real-space density ``\hat{𝐌}(𝐫)`` using the Fourier transform convention,
+#
+# ```math
+# \hat{𝐌}_𝐪 ≡ \int_V e^{+ i 𝐪⋅𝐫} \hat{𝐌}(𝐫) d𝐫.
+# ```
+#
+# With appropriate contraction of spin components, ``\mathcal{S}^{αβ}(𝐪, ω)``
+# directly relates to the neutron scattering cross-section. Here, ``𝐪`` and
+# ``ω`` represent momentum and energy transfer _to_ the sample. Sunny will
+# report the structure factor as an intensive quantity by dividing by the number
+# of chemical cells in the macroscopic sample. Full details are given in the
+# documentation page [Structure Factor Conventions](@ref). 
+#
+# If the spin Hamiltonian lacks inversion symmetry, intensities at ``±𝐪`` may
+# be inequivalent. A simple example is the 1D chain with competing Ising and
+# Dzyaloshinskii–Moriya couplings. Sunny calculations on this model can be
+# compared to those of other codes. For example, [SpinW](https://spinw.org/)
+# employs the opposite sign convention for the momentum transfer ``𝐪``.
+
+# ### 1D model lacking reflection symmetry
 
 using Sunny, GLMakie
 
-# Selecting the P1 spacegroup will effectively disable all symmetry analysis.
-# This can be a convenient way to avoid symmetry-imposed constraints on the
-# couplings. A disadvantage is that all bonds are treated as
-# symmetry-inequivalent, such that each coupling within the unit cell must be
-# specified independently.
+# Create a [`Crystal`](@ref) with spacegroup P1 to effectively disable all
+# symmetry analysis. This avoids any symmetry-imposed constraints on the
+# couplings. Because all bonds are treated as symmetry-inequivalent, each
+# coupling must be specified independently; there is no automatic propagation to
+# "equivalent" bonds.
 
-latvecs = lattice_vectors(2, 2, 1, 90, 90, 90)
[…]
