### 2025-06-24 f21adf0
Avoid storing duplicate logo files
Sunny logos are stored in the root `assets` directory. These will be copied into `docs/build/assets` at the beginning of each docs build.
