### 2025-06-24 6d5324d
Avoid storing duplicate logo files (#401)
Sunny logos are stored in the root `assets` directory. These will be copied into `docs/build/assets` at the beginning of each docs build.
