+++ docs/src/versions.md
@@ -1,7 +1,7 @@
 # Version History
 
 ## v0.7.7
-(In development)
+(Jun 25, 2025)
 
 * Add [`set_spin_rescaling_for_static_sum_rule!`](@ref) which sets the classical
   dipole magnitude to ``\sqrt{s (s + 1)}`` for each quantum spin-``s`` moment.
