### 2025-06-30 SunnySuite/Sunny.jl#403 @Chronum94
JOSS review comments/questions/etc
These are a few things for the JOSS review. Not all of these need to be implemented, but an explanation would be sufficient some (most?) cases. Some things (such as picking a subcrystal in the FeI3 tutorial) have been fixed already between me running the examples the first time weeks ago, and running them again now.

1. Example 2: The initial spin directions (first image), the suggested Langevin dt in the post-equilibration suggestion is slightly different. I get 0.4193 instead of 0.4223. This of course is not a huge deal but it would be nice to know if this is just due to different hardware, or if this can be made reproducible using explicit specification of random seeds.
2. Example 3: In the step with
```julia
sys = resize_supercell(sys, (4, 4, 4))
randomize_spins!(sys)
minimize_energy!(sys)
```
the minimized energy comes out to 63 for me.
```julia
julia> sys = resize_supercell(sys, (4, 4, 4))
System [SU(3)]
Supercell (4×4×4)×1
Energy per site -353/250

julia> randomize_spins!(sys)
julia> minimize_energy!(sys)
63
```
I assume the tutorial fell into a local minimum, but it would be nice to have a sentence like 'you may get something different since this is just a local minimum'.

3. Same goes for dynamical quenching tutorial (tutorial 4). I get slightly different images as I run the tutorial. This time it's a bit more significant than 3 sig-figs in:
![Image](https://github.com/user-attachments/assets/31333de5-0b23-44ed-8593-21c46f362f9c)

I imagine much of this can be fixed by fixing the random seed, but it would be good to have it.

I'll keep adding more things, but hopefully this starts out the discussion and apologies for the delay.
