### 2025-06-30 454e5fb
Prepare 0.7.8 release
