### 2025-06-30 990e0c2
Arrow shaft represents spin length
