### 2025-07-01 5cd6768
Tweak cone scaling
