### 2025-07-01 0cfb6e8
Arrow shaft now centered on atom
