+++ ext/PlottingExt/ViewCrystal.jl
@@ -98,12 +98,12 @@ function exchange_decomposition(J)
     return (vals, q)
 end
 
-function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
+function draw_exchange_geometries(; ax, visible, ionradius, pts, scaled_exchanges)
 
     ### Ellipsoids for symmetric exchanges
 
     # Dimensionless scalings and rotations associated with principle axes
-    decomps = exchange_decomposition.(scaled_exchanges)            
+    decomps = exchange_decomposition.(scaled_exchanges)
     scalings = map(x -> x[1], decomps)
     rotation = map(x -> x[2], decomps)
 
@@ -126,8 +126,7 @@ function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
         d = c+(1-c)*abs(y) # c ≤ d ≤ 1
         y > 0 ? Makie.RGBf(c, c, d) : Makie.RGBf(d, c, c)
     end
-    o = Makie.meshscatter!(ax, pts; color, markersize, rotation, specular=0, diffuse=1.5, inspectable=false)
-    Makie.connect!(o.visible, obs)
+    Makie.meshscatter!(ax, pts; color, markersize, rotation, specular=0, diffuse=1.5, visible, inspectable=false)
 
     # Draw dots using cylinders
     cylinders = map(eachcol(Sunny.Mat3(I))) do x
@@ -148,26 +147,51 @@ function draw_exchange_geometries(; ax, obs, ionradius, pts, scaled_exchanges)
              dim == 3 ? 1 : c]
         end
         markersize2 = [ms .* rs for (ms, rs) in zip(markersize, rescalings)]
-    
-        o = Makie.meshscatter!(ax, pts; color, markersize=markersize2, rotation, marker=cylinders[dim], inspectable=false)
-        Makie.connect!(o.visible, obs)            
+
+        Makie.meshscatter!(ax, pts; color, markersize=markersize2, rotation, marker=cylinders[dim], visible, inspectable=false)
     end
 
-    ### Cones for DM vectors. Because they tend to be weaker in magnitude,
-    ### we apply some heuristic amplification to the arrow size.
+    ### Cones for DM vectors
 
+    # DM vectors normalized by the overall exchange scale
     dmvecs = Sunny.extract_dmvec.(scaled_exchanges)
-    dirs = @. Makie.Vec3f(normalize(dmvecs))
+
+    # Filter nonzero DM vectors and retain associated exchange scalings
+    nonzeros_indices = findall(norm(dmvec) > 0 for dmvec in dmvecs)
+    dm_pts = pts[nonzeros_indices]
+    dmvecs = dmvecs[nonzeros_indices]
+    scalings_nonzero = scalings[nonzeros_indices]
+
     # The largest possible ellipsoid occurs in the case of `scalings ==
     # [1,1,1]`, yielding a sphere with size `ionradius`.
-    ellipsoid_radii = @. ionradius * norm(scalings) / √3
-    arrowsize = @. 2ionradius * cbrt(norm(dmvecs)) # size of arrow head
-    dm_pts = @. pts + 1.1ellipsoid_radii * dirs
-    o = Makie.arrows!(ax, dm_pts, dirs; lengthscale=0, arrowsize, diffuse=1.15, color=:magenta, specular=0.0, inspectable=false) 
-    Makie.connect!(o.visible, obs)
+    ellipsoid_radii = @. ionradius * norm(scalings_nonzero) / √3
+
+    # Size of cone scales like cube root of DM vector magnitude
+    markersize = @. 2ionradius * cbrt(norm(dmvecs))
+
[…]
