+++ docs/src/versions.md
@@ -1,5 +1,11 @@
 # Version History
 
+## v0.7.8
+(Jul 1, 2025)
+
+* Compatibility with Makie v0.24 ([PR
+  #393](https://github.com/SunnySuite/Sunny.jl/pull/393)).
+
 ## v0.7.7
 (Jun 25, 2025)
 
