### 2025-07-01 SunnySuite/Sunny.jl#403 @ddahlbom
Hi @Chronum94. Thanks for the comments, and sorry for not responding more promptly.

There are two possible sources for the numerical discrepancies you observe. One, as you suggest, is RNG seed inconsistency and the resulting identification of different local minima. The other is updates to the `Optim.jl` package and its deployment. A [recent Sunny update](https://github.com/SunnySuite/Sunny.jl/pull/397) should make convergence more consistent (to a local minimum).

Regarding the seed, in an early iteration of the docs we explicitly fixed the RNG seed inside the tutorial scripts. Sunny carries an RNG in each system, so generally this should be done by the `seed` keyword when creating a `System`. Using this approach, we found that some users would copy and paste the code from the example and be confused why the results did not change with each run (when they would execute the block in entirety). So we eliminated the fixed seed from the visible code, essentially due to user feedback. This also has the nice side effect of keeping the tutorial code examples cleaner.

That said, the process used by the doc build process is entirely repeatable. The global RNG seed is injected into the process here: https://github.com/SunnySuite/Sunny.jl/blob/main/docs/make.jl#L42-L46. This is hidden when a user just looks at the rendered docs or downloads them with the link at the top of the example page.

If you want to exactly reproduce the examples locally, you will need to call `import Random; Random.seed!(0)` at the top of each script and execute sequentially from that point on. This needs to be done with latest version of Sunny (0.7.7) and Optim (1.13.0), which were used to build the latest stable docs.

To address your specific points.

1) This is likely due to some combination of the two factors listed above.
2) This is again the same phenomenon. Note that `minimize_energy!` returns the number of iterations performed (not the energy itself).
3) This is essentially a MC calculation; the Langevin dynamics samples from the classical Boltzmann distribution at finite temperature. So significant discrepancies are expected when the number of samples is small (as it must be for autobuilt docs).

We're a bit reluctant to make too many disclaimers throughout the docs, as we've found users generally expect some stochasticity, and setting a seed is explicitly discussed in tutorials 3 and 7. If you think there is a particular place that it would be good to add further discussion, please let us know.
