### 2025-07-01 25a5da8
Slight tweak to rescaling heuristics
Improves a bit the visualization in tutorial "SW19_Different_Ions"
