### 2025-07-02 ba5dcaf
Rescaling of precision parameter for Spglib
