### 2025-07-03 665af29
Replace opnorm with minimum(svdvals(...)))
