+++ src/Symmetry/Crystal.jl
@@ -124,14 +124,13 @@ function Crystal(latvecs, positions, symbol::Union{Int, String}; types::Union{No
 end
 
 
-# Sunny crystal specification is agnostic to length units, and the Sunny symprec
-# parameter is dimensionless. Spglib, however, expects lattice vector magnitudes
-# to be order one (this will be the case for atomic crystals specified in Å).
-# The Spglib wrappers below do the following: (1) Identify a natural length
-# scale as the smallest singular value of the lattice vectors as a matrix, (2)
-# Non-dimensionalize the lattice vectors using this length, (3) Perform Spglib
-# symmetry analysis, (4) Re-introduce length dimensions in the appropriate
-# return values.
+# Sunny crystal specification is agnostic to length units. Spglib, however,
+# expects lattice vector magnitudes to be order one. The wrappers below do the
+# following: (1) Identify a natural length scale as the smallest singular value
+# of the lattice vectors as a matrix, (2) Non-dimensionalize the lattice vectors
+# using this length, (3) Perform Spglib symmetry analysis, (4) Re-introduce
+# length dimensions in the appropriate return values. See discussion and
+# limitations in PR #405.
 function dimensionless_cell(cell)
     (; lattice, positions, atoms, magmoms) = cell
     a0 = minimum(svdvals(Mat3(lattice)))
