+++ docs/src/versions.md
@@ -3,8 +3,8 @@
 ## v0.7.9
 (In development)
 
-* Fix tolerance of Spglib symmetry inference ([Issue
-  #404](https://github.com/SunnySuite/Sunny.jl/issues/404)).
+* Fix tolerance of Spglib symmetry inference ([PR
+  #405](https://github.com/SunnySuite/Sunny.jl/pull/405)).
 
 ## v0.7.8
 (Jul 1, 2025)
