### 2025-07-04 966b91b
Rescaling of precision parameter for Spglib (#405)
