### 2025-07-04 fb91ebf
Rescaling of precision parameter for Spglib (#405)
