### 2025-07-04 11d59cd
Entangled units fix (#406)
* Fix SampledCorrelationsStatic for entangled units

* Coverage for SampledCorrelationsStatic in tests
