### 2025-07-04 6bd7960
Entangled units fix (#406)
* Fix SampledCorrelationsStatic for entangled units

* Coverage for SampledCorrelationsStatic in tests
