+++ dev/assets/scripts/08_Momentum_Conventions.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.8"
+@assert pkgversion(Sunny) >= v"0.7.9"
 
 latvecs = lattice_vectors(1, 1, 1, 90, 90, 90)
 cryst = Crystal(latvecs, [[0, 0, 0]], "P1")
