+++ dev/assets/scripts/03_LSWT_SU3_FeI2.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.8"
+@assert pkgversion(Sunny) >= v"0.7.9"
 
 units = Units(:meV, :angstrom)
 a = b = 4.05012  # Lattice constants for triangular lattice (Å)
