+++ dev/assets/scripts/SW05_Simple_kagome_FM.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.8"
+@assert pkgversion(Sunny) >= v"0.7.9"
 
 units = Units(:meV, :angstrom)
 latvecs = lattice_vectors(6, 6, 5, 90, 90, 120)
