+++ dev/assets/scripts/06_CP2_Skyrmions.jl
@@ -1,5 +1,5 @@
 using Sunny, GLMakie
-@assert pkgversion(Sunny) >= v"0.7.8"
+@assert pkgversion(Sunny) >= v"0.7.9"
 
 lat_vecs = lattice_vectors(1, 1, 10, 90, 90, 120)
 positions = [[0, 0, 0]]
