### 2025-07-04 SunnySuite/Sunny.jl#403 @Chronum94
Hi @kbarros @ddahlbom

I've had a look at both the above comments and they seem reasonable to me (most of them appear to be differences in design choices that one expects, and I defer to the authors/members of the community in terms of expected API behaviour).

I'm on vacation for this and the upcoming week, so chances are I will not be the most available, but I'll post more comments/questions here when I go over more examples etc (which may happen during the vacation since I do like reviewing code, but I wouldn't rely on it).

Thanks for the responses so far either way!
