### 2025-07-14 #8138 @ddahlbom
For future reference, I'll link the issue @Chronum94 opened: https://github.com/SunnySuite/Sunny.jl/issues/403
