### 2025-07-14 d4c8b46
Port 2-magnon scattering script
