### 2025-07-15 fb59011
Implemented free two magnon intensities for the current version of Sunny
