+++ src/SpinWaveTheory/LSWTCorrections.jl
@@ -121,30 +121,18 @@ end
 
 
 """
-    dssf_tc(swt::SpinWaveTheory, q, ωs, η::Float64; opts...)
+    intensities_free_two_magnon(swt::SpinWaveTheory, q, energies, η::Float64; opts...)
 
-Given a [`SpinWaveTheory`](@ref) object, computes the dynamical spin structure
-factor,
-```math
-    𝒮^{αβ}(𝐪, ω) = 1/(2πN)∫dt ∑_𝐫 \\exp[i(ωt - 𝐪⋅𝐫)] ⟨S^α(𝐫, t)S^β(0, 0)⟩,
-```
-
-from the two-particle continuum contribution,
-
-```math
-    𝒮^{αβ}(𝐪, ω) =1/N ∑_n ∑_m ∑_{𝐤}  |A_n^{αβ}_{nm}(𝐤, 𝐪)|^2 δ[ω-ω_{n}(𝐪+𝐤)-ω_{m}(-𝐤)].
-```
-
-``q`` must be a 3-vector in reciprocal lattice units (RLU), i.e., in the basis of
-reciprocal lattice vectors.
+Calculates dynamical pair correlation intensities for a provided ``𝐪``-point in
+reciprocal space.
 """
-function dssf_tc(swt::SpinWaveTheory, q, ωs, η::Float64; opts...)
+function intensities_free_two_magnon(swt::SpinWaveTheory, q, energies, η::Float64; opts...)
     kernel = lorentzian(; fwhm=2η)
 
     (; sys, data, measure) = swt
     cryst = orig_crystal(sys)
-    q_reshaped = to_reshaped_rlu(swt.sys, q)
     q_global = cryst.recipvecs * q
+    q_reshaped = to_reshaped_rlu(sys, q)
 
     # Number of atoms in magnetic cell
     Nm = length(sys.dipoles)
@@ -156,7 +144,7 @@ function dssf_tc(swt::SpinWaveTheory, q, ωs, η::Float64; opts...)
     L = nbands(swt)
 
     # Get some integer numbers for later use
-    numωs = length(ωs)
+    num_energies = length(energies)
     num_obs = num_observables(measure)
     num_corrs = num_correlations(measure)
 
@@ -168,16 +156,20 @@ function dssf_tc(swt::SpinWaveTheory, q, ωs, η::Float64; opts...)
     H2 = zeros(ComplexF64, 2L, 2L)
     V2 = zeros(ComplexF64, 2L, 2L)
 
-    Avec_pref = zeros(ComplexF64, Nm)
+    Avec_pref = zeros(ComplexF64, num_obs, Nm)
     Avec = zeros(ComplexF64, num_obs, L, L)
-    corrbuf = zeros(ComplexF64, num_corrs*numωs)
+    corrbuf = zeros(ComplexF64, num_corrs, num_energies)
+    resbuf = zeros(num_energies)
 
-    for i = 1:Nm
-        @assert Nm == natoms(sys.crystal)
-        Avec_pref[i] = exp(-2π*im * dot(q_reshaped, sys.crystal.positions[i]))
+
+    for i = 1:Nm, μ in 1:num_obs
+        r_global = global_position(sys, (1,1,1,i))
+        ff = get_swt_formfactor(measure, μ, i)
+        Avec_pref[μ, i] = exp(-1im * dot(q_global, r_global))
+        Avec_pref[μ, i] *= compute_form_factor(ff, norm2(q_global))
     end
 
-    Sqω = hcubature((0,0,0), (1,1,1); opts...) do k_reshaped
+    ints = hcubature((0,0,0), (1,1,1); opts...) do k_reshaped
[…]
