### 2025-07-18 #8138 @Chronum94
> Hi [@Chronum94](https://github.com/Chronum94) - thanks for opening the issue. It would be great if either you or [@ddahlbom](https://github.com/ddahlbom) could post an update here on the status of this if there's any progress. Also please make sure to link this review issue when you create new issues on the repo so that it's easier for everyone to follow - thank you!
>
@phibeck - Hi, I've gone through most of the tutorials and examples. I'm going through more of them to check functionality/edge cases. A bit of this also comes down to making sure I'm not being misled by expert blindspots etc. I should have another big blob of updates this weekend (or worst case, early next week).

Real life has a nasty habit of getting in the way, but I also understand wanting a completed review. Please let me know if there's a strong time-requirement for this (for whatever reasons), and I'll try to see what I can do. But otherwise, this weekend/early next week should have my last bits of review.
