### 2025-07-22 e1f586f
Remove unneeded assertion in spiral energy calculation
Two matrices may be approximately equal at some relative tolerance, but this does not guarantee the same for individual matrix elements.
