### 2025-07-23 34d9050
Transparency in `plot_intensities` for bands data
This transparency achieves consistency with other Makie lines plots. Additionally, it significantly reduces the file size for PDF exports using the CairoMakie backend.
