### 2025-07-24 6ba6d1f
Improve color scheme
