+++ docs/src/versions.md
@@ -5,6 +5,8 @@
 
 * Fix tolerance of Spglib symmetry inference ([PR
   #405](https://github.com/SunnySuite/Sunny.jl/pull/405)).
+* Modify color gradient in [`plot_intensities`](@ref) when plotting discrete
+  bands ([PR #411](https://github.com/SunnySuite/Sunny.jl/pull/411)).
 
 ## v0.7.8
 (Jul 1, 2025)
