+++ ext/PlottingExt/PlottingUtils.jl
@@ -28,19 +28,19 @@ const seaborn_muted = [
 ]
 
 # Numbers adapted from :thermal colormap defined in matplotlib/cmocean
-const reverse_thermal_fade = let
+reverse_thermal_fade = let
     data = [
         (Makie.RGBAf(1.00, 1.00, 0.60, 0.00), 0.00),
-        (Makie.RGBAf(0.97, 0.95, 0.40, 0.10), 0.01),
-        (Makie.RGBAf(0.97, 0.95, 0.40, 0.20), 0.03),
-        (Makie.RGBAf(0.94, 0.90, 0.31, 0.40), 0.05),
-        (Makie.RGBAf(0.98, 0.80, 0.23, 0.50), 0.11),
-        (Makie.RGBAf(0.86, 0.60, 0.39, 0.70), 0.22),
-        (Makie.RGBAf(0.67, 0.40, 0.51, 0.90), 0.35),
-        (Makie.RGBAf(0.51, 0.31, 0.55, 1.00), 0.45),
-        (Makie.RGBAf(0.37, 0.24, 0.60, 1.00), 0.55),
-        (Makie.RGBAf(0.18, 0.20, 0.60, 1.00), 0.70),
-        (Makie.RGBAf(0.04, 0.18, 0.36, 1.00), 0.85),
+        (Makie.RGBAf(0.97, 0.95, 0.40, 0.30), 0.01),
+        (Makie.RGBAf(0.97, 0.95, 0.40, 0.50), 0.05),
+        (Makie.RGBAf(0.94, 0.90, 0.31, 0.70), 0.10),
+        (Makie.RGBAf(0.98, 0.80, 0.23, 0.90), 0.20),
+        (Makie.RGBAf(0.86, 0.60, 0.39, 0.95), 0.30),
+        (Makie.RGBAf(0.67, 0.40, 0.51, 1.00), 0.40),
+        (Makie.RGBAf(0.51, 0.31, 0.55, 1.00), 0.52),
+        (Makie.RGBAf(0.37, 0.24, 0.60, 1.00), 0.65),
+        (Makie.RGBAf(0.18, 0.20, 0.60, 1.00), 0.77),
+        (Makie.RGBAf(0.04, 0.18, 0.36, 1.00), 0.89),
         (Makie.RGBAf(0.01, 0.13, 0.20, 1.00), 1.00),
     ]
     Makie.cgrad(getindex.(data, 1), getindex.(data, 2))
