### 2025-07-25 e881abd
Faded colormap for case allpositive=false
