### 2025-07-26 45575d0
Colormap for `allpositive=false` has 1% range of full transparency
