+++ ext/PlottingExt/PlottingUtils.jl
@@ -50,7 +50,8 @@ blue_white_red_fade = let
     data = [
         (Makie.RGBAf(0.00, 0.00, 1.00, 1.00), 0.00),
         (Makie.RGBAf(0.80, 0.80, 1.00, 1.00), 0.45),
-        (Makie.RGBAf(1.00, 1.00, 1.00, 0.00), 0.50),
+        (Makie.RGBAf(1.00, 1.00, 1.00, 0.00), 0.495),
+        (Makie.RGBAf(1.00, 1.00, 1.00, 0.00), 0.505),
         (Makie.RGBAf(1.00, 0.80, 0.80, 1.00), 0.55),
         (Makie.RGBAf(1.00, 0.00, 0.00, 1.00), 1.00),
     ]
