+++ src/Symmetry/Parsing.jl
@@ -1,6 +1,6 @@
-# Functions for parsing Crystals / SymOps from text / .cif files
+# Reading CIF and mCIF files
 
-function parse_cif_float_with_err(str::String)
+function parse_cif_float_with_err(str::String; maybe_fractional)
     parts = split(str, '(')
     number_str = parts[1]
     number = parse(Float64, number_str)
@@ -9,8 +9,20 @@ function parse_cif_float_with_err(str::String)
     if length(parts) == 1
         # E.g., 0.2 has an assumed error bar of ±0.05
         err = 10.0^(-sigfigs) / 2
+
+        # Components of each Wyckoff position in an ITA standard cell may also
+        # be an exact multiple of {1/2, 1/3, 1/4, 1/6, 1/8}. For example, the
+        # position 0.125 = 1/8 is likely to be exact. Such strings should not
+        # contributed to the estimated error.
+        if maybe_fractional
+            smallest_remainder = minimum(abs(rem(number * c, 1, RoundNearest)) / c for c in (2, 3, 4, 6, 8))
+            if smallest_remainder < 1e-12
+                err = 0.0
+            end
+        end
     else
         # E.g., 0.2(3) has an error bar of ±0.3
+        @assert parts[2][end] == ')'
         err_str = parts[2][begin:end-1]
         err = 10.0^(-sigfigs) * parse(Int, err_str)
     end
@@ -18,12 +30,11 @@ function parse_cif_float_with_err(str::String)
     return (number, err)
 end
 
-
-
 function parse_cif_float(str::String)
-    return parse_cif_float_with_err(str)[1]
+    return parse_cif_float_with_err(str; maybe_fractional=false)[1]
 end
 
+#=
 # Components x of each Wyckoff position will either be:
 # 1. An exact multiple of {1/2, 1/3, 1/4, 1/6, 1/8}, or
 # 2. An arbitrary real number  
@@ -41,7 +52,7 @@ function guess_precision(x)
 
     min(tol1, tol2)
 end
-
+=#
 
 function parse_number_or_fraction(s)
     # Parse a number or fraction
@@ -102,29 +113,49 @@ end
 
 # Reads the crystal from a `.cif` file located at the path `filename`. See
 # extended doc string in Crystal.jl.
-function Crystal(filename::AbstractString; keep_supercell=false, symprec=1e-4, override_symmetry=nothing)
+function Crystal(filename::AbstractString; keep_supercell=false, symprec=nothing, override_symmetry=nothing)
     if !isnothing(override_symmetry)
         @warn "Ignoring deprecated `override_symmetry` option"
     end
 
     cif = CIF.Cif(filename)
-    # For now, assumes there is only one data collection per .cif
+    # Accessor `oneof` assumes collections in CIF are unique
     cif = cif[first(keys(cif))]
     oneof(fields...) = findfirstval(in(keys(cif)), fields)
 
-    a = parse_cif_float(cif["_cell_length_a"][1])
-    b = parse_cif_float(cif["_cell_length_b"][1])
-    c = parse_cif_float(cif["_cell_length_c"][1])
+    # If user-provided symprec0 is nothing, we will need to infer a symprec from
+    # strings for lattice constants and coordinate positions.
+    symprec0 = symprec
+    symprec = 0
+
+    (a, err_a) = parse_cif_float_with_err(cif["_cell_length_a"][1]; maybe_fractional=false)
+    (b, err_b) = parse_cif_float_with_err(cif["_cell_length_b"][1]; maybe_fractional=false)
+    (c, err_c) = parse_cif_float_with_err(cif["_cell_length_c"][1]; maybe_fractional=false)
[…]
