+++ src/Symmetry/Parsing.jl
@@ -34,25 +34,6 @@ function parse_cif_float(str::String)
     return parse_cif_float_with_err(str; maybe_fractional=false)[1]
 end
 
-#=
-# Components x of each Wyckoff position will either be:
-# 1. An exact multiple of {1/2, 1/3, 1/4, 1/6, 1/8}, or
-# 2. An arbitrary real number  
-# In practice, the CIF will provide x as a decimal expansion. Work backwards to
-# make a conservative guess about the precision of x. For example, 0.125 = 1/8
-# could be exact, but 0.126 is likely ±0.0005
-function guess_precision(x)
-    # See if x closely matches any simple fraction
-    tol1 = minimum(abs(rem(x * c, 1, RoundNearest)) / c for c in (2, 3, 4, 6, 8))
-
-    # Count digits in a decimal expansion
-    x_decimal = round(x, digits=12)
-    fraction_digits = length(split(repr(x_decimal), ".")[2])
-    tol2 = 10.0^(-fraction_digits) / 2
-
-    min(tol1, tol2)
-end
-=#
 
 function parse_number_or_fraction(s)
     # Parse a number or fraction
