### 2025-08-05 f854c46
Tweak versions note
