+++ docs/src/versions.md
@@ -8,7 +8,7 @@
   always downfold to a standard chemical cell. As before, use
   [`set_dipoles_from_mcif!`](@ref) to set dipoles on a reshaped [`System`](@ref)
   ([PR #413](https://github.com/SunnySuite/Sunny.jl/pull/413)). 
-* Fix tolerance of Spglib symmetry inference ([PR
+* Scale `symprec` prior to symmetry inference as expected by Spglib ([PR
   #405](https://github.com/SunnySuite/Sunny.jl/pull/405)).
 
 ## v0.7.8
