+++ src/Symmetry/WyckoffData.jl
@@ -31,7 +31,7 @@ end
 # that F = F′ and c = c′ mod 1.
 function is_periodic_copy(w1::WyckoffExpr, w2::WyckoffExpr)
     atol = 1e-12 # FP precision because all coefficients are perfect fractions
-    return isapprox(w1.F, w2.F; atol) && all_integer(w2.c - w1.c; symprec=atol)
+    return isapprox(w1.F, w2.F; atol) && is_periodic_copy(w1.c, w2.c; symprec=atol)
 end
 
 function crystallographic_orbit(symops::Vector{SymOp}, w::WyckoffExpr)
