### 2025-08-05 685d329
Clean up checks for position overlap
