+++ src/Symmetry/Crystal.jl
@@ -211,9 +211,10 @@ function sort_sites!(cryst::Crystal)
             end
         end
 
-        # Should never hit this case because either `validate_positions` or
-        # `crystallographic_orbits_distinct` should already be called.
-        error("Symmetry-equivalent sites $r1 and $r2")
+        # Should never get here because `validate_positions` or
+        # `validate_orbits` should already have passed. But report something
+        # meaningful just in case.
+        error("Symmetry-equivalent positions $r1 and $r2")
     end
     p = sort(eachindex(cryst.positions), lt=less_than)
     permute_sites!(cryst, p)
@@ -315,18 +316,9 @@ function crystal_from_inferred_symmetry(latvecs::Mat3, positions::Vector{Vec3}, 
     # Print a warning if non-conventional lattice vectors are detected.
     try cell_type(latvecs) catch e @warn e.msg end
 
-    for i in 1:length(positions)
-        for j in i+1:length(positions)
-            ri = positions[i]
-            rj = positions[j]
-            if is_periodic_copy(ri, rj; symprec)
-                error("Positions $ri and $rj are equivalent.")
-            end
-        end
-    end
-
     recipvecs = 2π*Mat3(inv(latvecs)')
     positions = wrap_to_unit_cell.(positions; symprec)
+    validate_positions(positions; symprec)
 
     cell = Spglib.Cell(latvecs, positions, types)
     d = spg_get_dataset_scaled(cell, symprec)
@@ -354,7 +346,7 @@ function crystal_from_inferred_symmetry(latvecs::Mat3, positions::Vector{Vec3}, 
     @assert unique(classes) == 1:maximum(classes)
 
     ret = Crystal(nothing, latvecs, recipvecs, positions, types, classes, sg, symprec)
-    validate(ret)
+    validate_crystal(ret)
     for i in 1:natoms(ret)
         w = get_wyckoff(ret, i)
         @assert w.letter == d.wyckoffs[i]
@@ -388,7 +380,7 @@ function is_spacegroup_type_consistent(sgt, latvecs)
 end
 
 
-function crystallographic_orbit(symops::Vector{SymOp}, position::Vec3; symprec)
+function crystallographic_orbit(position::Vec3; symops::Vector{SymOp}, symprec)
     orbit = Vec3[]
     for s = symops
         x = wrap_to_unit_cell(transform(s, position); symprec)
@@ -399,19 +391,38 @@ function crystallographic_orbit(symops::Vector{SymOp}, position::Vec3; symprec)
     return orbit
 end
 
-function crystallographic_orbits_distinct(symops::Vector{SymOp}, positions::Vector{Vec3}; symprec, multiplicities=nothing, wyckoffs=nothing)
-    orbits = crystallographic_orbit.(Ref(symops), positions; symprec)
+function validate_positions(positions; symprec)
+    for i in eachindex(positions), j in i+1:length(positions)
+        ri, rj = positions[[i, j]]
+        ri_str, rj_str = fractional_vec3_to_string.((ri, rj))
+        symprec_str = number_to_simple_string(symprec; digits=2)
+        if is_periodic_copy(ri, rj; symprec)
+            error("Overlapping positions $ri_str and $rj_str at $symprec_str")
+        elseif is_periodic_copy(ri, rj; symprec=10symprec)
+            @warn "Nearly overlapping positions $ri_str and $rj_str at $symprec_str"
+        end
+    end
+end
 
+function validate_orbits(positions, orbits; symprec, multiplicities=nothing, wyckoffs=nothing)
+    @assert size(positions) == size(orbits)
     # Check that orbits are distinct
-    for (i, ri) in enumerate(positions), j in i+1:length(orbits)
[…]
