### 2025-08-05 a45d560
More tests. Also detect near-overlaps
