+++ test/test_cif.jl
@@ -14,6 +14,18 @@ end
     @test Sunny.get_wyckoff(cryst, 7).letter == 'c'
 end
 
+@testitem "FeI2_orth" begin
+    filename = joinpath(@__DIR__, "cifs", "FeI2_orth.cif")
+
+    msg = "This CIF uses a non-standard spacegroup setting, making symmetry \
+        analysis unreliable! Use `standardize(cryst)` to obtain the \
+        standard chemical cell. Then use `reshape_supercell(sys, shape)` \
+        for calculations on an arbitrarily shaped system."
+    cryst = @test_logs (:warn, msg) Crystal(filename)
+    @test Sunny.get_wyckoff(cryst, 1).letter == 'a'
+    @test Sunny.get_wyckoff(cryst, 3).letter == 'd'
+end
+
 @testitem "Alpha quartz" begin
     msg = "This CIF uses a non-standard spacegroup setting, making symmetry \
            analysis unreliable! Use `standardize(cryst)` to obtain the \
