### 2025-08-06 43f3731
Tweak docstring
