### 2025-08-06 bcb05b8
Tweak doc string
