### 2025-08-07 a80abd6
Improvements to CIF and mCIF loading (#413)
Dimensionless precision `symprec` is now inferred. Loading an mCIF will automatically convert to a standard chemical cell compatible with `set_dipoles_from_mcif!`. Various other internal robustness improvements.
