+++ src/Symmetry/SpacegroupData.jl
@@ -240,10 +240,10 @@ end
 
 # Given a spacegroup number and a table of symops, try to infer the affine map
 # that transforms to the ITA standard setting.
-function hall_number_from_symops(sgnum, symops)
+function hall_number_from_symops(sgnum, symops; atol)
     sgts = filter(all_spacegroup_types_for_symbol(sgnum)) do sgt
         Rs, Ts = Spglib.get_symmetry_from_database(sgt.hall_number)
-        SymOp.(Rs, Ts) ≈ symops
+        isapprox(SymOp.(Rs, Ts), symops; atol)
     end
 
     if isempty(sgts)
@@ -322,6 +322,13 @@ const standard_primitive_basis = Dict(
     'F' => SA[0 1/2 1/2; 1/2 0 1/2; 1/2 1/2 0],
     'R' => SA[2/3 -1/3 -1/3; 1/3 1/3 -2/3; 1/3 1/3 1/3],
 )
+
+# Let `setting` denote the transformation from a custom setting to the standard
+# one: xₛ = transform(setting, x). Given a `symop` that acts in the standard
+# setting, return the transformed symop that acts in the custom setting.
+function map_symop_to_setting(symop; setting)
+    return inv(setting) * symop * setting
+end
 
 struct Spacegroup
     symops    :: Vector{SymOp} # Symmetry operations
