+++ test/test_cif.jl
@@ -1,30 +1,57 @@
 @testitem "ErOBr" begin
     filename = joinpath(@__DIR__, "cifs", "ErOBr.cif")
-    cryst = Crystal(filename; symprec=1e-3)
+    cryst = Crystal(filename)
     @test Sunny.get_wyckoff(cryst, 1).letter == 'c'
     @test Sunny.get_wyckoff(cryst, 3).letter == 'a'
     @test Sunny.get_wyckoff(cryst, 5).letter == 'c'
 end
 
+@testitem "UPt3" begin
+    filename = joinpath(@__DIR__, "cifs", "UPt3.cif")
+    cryst = Crystal(filename)
+    @test cryst.sg.setting == one(Sunny.SymOp)
+    @test Sunny.get_wyckoff(cryst, 1).letter == 'h'
+    @test Sunny.get_wyckoff(cryst, 7).letter == 'c'
+end
+
+@testitem "FeI2_orth" begin
+    filename = joinpath(@__DIR__, "cifs", "FeI2_orth.cif")
+    msg = "Inconsistent symmetry operations! This may occur with an incomplete CIF, \
+           a non-standard setting, or failed inference. Try overriding `symprec` \
+           (inferred 5e-06)."
+    cryst = @test_logs (:warn, msg) Crystal(filename)
+    @test Sunny.get_wyckoff(cryst, 1).letter == 'a'
+    @test Sunny.get_wyckoff(cryst, 3).letter == 'd'
+end
+
+@testitem "Alpha quartz" begin
+    filename = joinpath(@__DIR__, "cifs", "alpha_quartz.cif")
+    msg = "Cell appears non-standard. Consider `standardize(cryst)` and then \
+           `reshape_supercell(sys, shape)` for calculations on an arbitrarily shaped \
+           system."
+    cryst = @test_logs (:info, msg) Crystal(filename)
+    @test cryst.sg.number == 154
+    @test Sunny.get_wyckoff(cryst, 1).letter == 'a'
+    @test Sunny.get_wyckoff(cryst, 4).letter == 'c'
+end
+
 @testitem "mCIF ZnFe2O4" begin
     filename = joinpath(@__DIR__, "cifs", "ZnFe2O4_jana.cif")
 
-    msg = """Loading the magnetic cell as chemical cell for TESTING PURPOSES only.
-             Set the option `override_symmetry=true` to infer the standard chemical
-             cell and its spacegroup symmetries."""
-    cryst1 = @test_logs (:warn, msg) Crystal(filename; symprec=1e-2)
-    cryst1 = subcrystal(cryst1, "Fe_1", "Fe_4")
-    sys1 = System(cryst1, [1 => Moment(s=3/2, g=-2), 17 => Moment(s=3/2, g=-2)], :dipole)
+    msg = "Use `keep_supercell=true` for testing purposes only! Inferred symmetries are unreliable."
+    cryst1 = @test_logs (:warn, msg) Crystal(filename; keep_supercell=true)
+    cryst1 = subcrystal(cryst1, "Fe")
+    sys1 = System(cryst1, [1 => Moment(s=3/2, g=-2)], :dipole)
     set_dipoles_from_mcif!(sys1, filename)
-    
-    cryst2 = Crystal(filename; override_symmetry=true, symprec=1e-2)
+
+    cryst2 = Crystal(filename)
     cryst2 = subcrystal(cryst2, "Fe")
     sys2 = System(cryst2, [1 => Moment(s=3/2, g=-2)], :dipole)
     msg = "Use `resize_supercell(sys, (1, 1, 2))` to get compatible system"    
[…]
