+++ src/Symmetry/Crystal.jl
@@ -3,16 +3,17 @@
 An object describing a crystallographic unit cell and its space group symmetry.
 Constructors are as follows:
 
+    Crystal(filename; symprec=nothing)
 
-    Crystal(filename; override_symmetry=false, symprec=nothing)
+Reads the crystal from a CIF or mCIF at `filename`. Lattice vectors will be in
+units of angstrom. In most cases, `symprec` can be omitted. If provided, it will
+specify the precision of the dimensionless site position data (commonly between
+1e-2 and 1e-5).
 
-Reads the crystal from a `.cif` file located at the path `filename`. If
-`override_symmetry=true`, the spacegroup will be inferred based on atom
-positions and the returned unit cell may be reduced in size. For an mCIF file,
-the return value is the magnetic supercell, unless `override_symmetry=true`. If
-a dimensionless precision for spacegroup symmetries cannot be inferred from the
-CIF file, it must be specified with `symprec`. The `latvecs` field of the
-returned `Crystal` will be in units of angstrom.
+If reading an mCIF, the magnetic cell will be converted to an inferred
+conventional chemical cell and returned. Use this crystal to build an
+appropriately shaped [`System`](@ref) and then load the magnetic order with
+[`set_dipoles_from_mcif!`](@ref).
 
     Crystal(latvecs, positions; types=nothing, symprec=1e-5)
 
@@ -205,14 +206,17 @@ function sort_sites!(cryst::Crystal)
         end
         ri = cryst.positions[i]
         rj = cryst.positions[j]
+
         for k = 3:-1:1
-            if !isapprox(ri[k], rj[k], atol=10cryst.symprec)
+            if !is_integer(ri[k]-rj[k]; atol=cryst.symprec)
                 return ri[k] < rj[k]
             end
         end
-        str1, str2 = fractional_vec3_to_string.((ri, rj))
-        error("""Detected two very close atoms ($str1 and $str2).
-                 If positions inferred from spacegroup, try increasing `symprec` parameter to `Crystal`.""")
+
+        # Should never get here because `validate_positions` or
+        # `validate_orbits` should already have passed. But report something
+        # meaningful just in case.
+        error("Symmetry-equivalent positions $r1 and $r2")
     end
     p = sort(eachindex(cryst.positions), lt=less_than)
     permute_sites!(cryst, p)
@@ -276,19 +280,21 @@ function standardize(cryst::Crystal; idealize=true)
     lattice = Mat3(lattice)
 
     if !idealize
-        # Empirically, these lattice vectors from spglib are accurate to about 6
-        # digits. However, spglib produces much higher accuracy with the
-        # `idealize=true` option. Rotate the higher precision lattice vectors so
-        # that they give the best match to the ones for the non-idealized cell.
+        # Even if we're not idealizing the site positions, it is still important
+        # to tune the lattice vectors so that the lattice system is exactly
+        # consistent with the spacegroup setting (e.g. 90° angles are expected
+        # for cubic and tetragonal spacegroups). This adjustment is needed to
+        # ensure that `latvecs * symop.R * inv(latvecs)` is exactly orthogonal
+        # for each symmetry operation of the spacegroup.
         std_lattice = Mat3(spg_standardize_cell_scaled(cell, symprec; no_idealize=false).lattice)
         R = closest_unitary(lattice / std_lattice)
-        isapprox(R*std_lattice, lattice; rtol=1e-5) || error("Failed to standardize the cell")
[…]

[…]
