### 2025-08-07 #8138 @phibeck
Hi @Chronum94 – thanks, I completely understand. As far as I know there’s no fixed deadline for this review, but of course we’d love to wrap it up as soon as possible. As you know, the JOSS review is more incremental than traditional manuscript reviews for other journals, so it’s mostly important to keep the conversation moving as there can be a lot more back and forth between reviewers and authors, which can take time. That being said, there’s certainly no rush before @jabirali gets to it in September as we need two completed reviews.
