+++ src/System/PairExchange.jl
@@ -34,7 +34,6 @@ function Base.iszero(c::PairCoupling)
 end
 
 function Base.:+(c1::PairCoupling, c2::PairCoupling)
-    @assert c1.isculled == c2.isculled
     @assert c1.bond == c2.bond
 
     scalar = c1.scalar + c2.scalar
@@ -55,14 +54,18 @@ function Base.:+(c1::PairCoupling, c2::PairCoupling)
     PairCoupling(c1.bond, scalar, bilin, biquad, general)
 end
 
+function Base.:*(pc::PairCoupling, c::Real)
+    return PairCoupling(pc.bond, pc.scalar * c, pc.bilin * c, pc.biquad * c, pc.general * c)
+end
+
 # Internal function only
-function replace_coupling!(list, coupling::PairCoupling; accum=false)
+function replace_coupling!(pairs, coupling::PairCoupling; accum=false)
     (; bond) = coupling
 
     # Find and remove existing couplings for this bond
-    idxs = findall(c -> c.bond == bond, list)
-    existing = list[idxs]
-    deleteat!(list, idxs)
+    idxs = findall(pc -> pc.bond == bond, pairs)
+    existing = pairs[idxs]
+    deleteat!(pairs, idxs)
 
     # If the new coupling is exactly zero, and we're not accumulating, then
     # return early
@@ -76,12 +79,11 @@ function replace_coupling!(list, coupling::PairCoupling; accum=false)
     # Add to the list and sort by isculled. Sorting after each insertion will
     # introduce quadratic scaling in length of `couplings`. If this becomes
     # slow, we could swap two PairCouplings instead of performing a full sort.
-    push!(list, coupling)
-    sort!(list, by=c->c.isculled)
+    push!(pairs, coupling)
+    sort!(pairs, by=pc->pc.isculled)
 
     return
 end
-
 
 # If A ≈ α B, then return the scalar α. Otherwise, return A.
 function proportionality_factor(A, B; atol=1e-12)
@@ -163,6 +165,13 @@ function Base.:+(op1::TensorDecomposition, op2::TensorDecomposition)
     =#
 end
 
+function Base.:*(op::TensorDecomposition, c::Real)
+    data = map(op.data) do (A, B)
+        (A*c, B)
+    end
+    return TensorDecomposition(op.gen1, op.gen2, data)
+end
+
 function Base.isapprox(op1::TensorDecomposition, op2::TensorDecomposition; kwargs...)
     isempty(op1.data) == isempty(op2.data) && return true
     op1′ = sum(kron(A, B) for (A, B) in op1.data)
@@ -205,11 +214,12 @@ function check_allowable_dipole_coupling(tensordec, mode)
     end
 end
 
-function set_pair_coupling_aux!(sys::System, scalar::Float64, bilin::Union{Float64, Mat3}, biquad::Union{Float64, Mat5}, tensordec::TensorDecomposition, bond::Bond)
+function set_pair_coupling_aux!(sys::System, scalar::Float64, bilin::Union{Float64, Mat3}, biquad::Union{Float64, Mat5},
+                                tensordec::TensorDecomposition, bond::Bond, param)
     # If `sys` has been reshaped, then operate first on `sys.origin`, which
     # contains full symmetry information.
     if !isnothing(sys.origin)
-        set_pair_coupling_aux!(sys.origin, scalar, bilin, biquad, tensordec, bond)
+        set_pair_coupling_aux!(sys.origin, scalar, bilin, biquad, tensordec, bond, param)
         transfer_interactions!(sys, sys.origin)
         return
     end
[…]
