+++ src/System/Interactions.jl
@@ -17,18 +17,13 @@ function Base.copy(param::ModelParam)
     return ModelParam(param.label, param.scale; param.onsites, param.pairs)
 end
 
-#=
-function param_has_atom(i)
-    param -> any(i == j for (j, _) in param.onsites)
-end
-=#
 
 # Find an existing param satisfying `matches` or, if missing, create a new one
 # with a unique label.
 function get_default_param(sys::System, matches::Function)
     label = nothing
 
-    is_unlabeled(param) = startswith(string(param.label), "Param_")
+    is_unlabeled(param) = startswith(string(param.label), "Unnamed")
 
     # Search for an existing param that `matches` and is unlabeled.
     inds = findall(sys.params) do param
@@ -39,26 +34,31 @@ function get_default_param(sys::System, matches::Function)
         # There can be at most one. Use it.
         label = sys.params[only(inds)].label
     else
-        # Otherwise, create a unique "Param_" label.
+        # Otherwise, create a unique "Unnamed" label.
         cnt = count(is_unlabeled, sys.params)
-        label = Symbol("Param_$(cnt+1)")
+        label = Symbol("Unnamed$(cnt+1)")
     end
 
     return (label, 1.0)
 end
 
 
-function repopulate_pair_couplings!(sys::System)
+function repopulate_couplings_from_params!(sys::System)
     @assert is_homogeneous(sys)
     ints = interactions_homog(sys)
 
-    # Clear existing pair couplings
-    for (; pair) in ints
-        empty!(pair)
+    # Clear current interactions
+    for i in eachindex(ints)
+        ints[i].onsite = ints[i].onsite * 0.0
+        empty!(ints[i].pair)
     end
 
     # Accumulate from params
     for param in sys.params
+        for (i, oc) in param.onsites
+            ints[i].onsite += oc * param.scale
+        end
+
         for pc in param.pairs
             b = pc.bond
             scaled_pc = pc * param.scale
@@ -78,29 +78,8 @@ function repopulate_pair_couplings!(sys::System)
     for (; pair) in ints
         sort!(pair, by = pc -> pc.isculled)
     end
-
-    return
 end
 
-function repopulate_onsite_couplings!(sys::System{N}) where N
-    @assert is_homogeneous(sys)
-    ints = interactions_homog(sys)
-
-    # Clear current onsite couplings
-    for i in eachindex(ints)
-        ints[i].onsite = empty_anisotropy(sys.mode, N)
-    end
-
-    # Accumulate from params
-    for param in sys.params
-        s = param.scale
-        for (i, oc) in param.onsites
-            ints[i].onsite += oc * s
-        end
-    end
-
-    return
-end
 
 function empty_interactions(mode::Symbol, Na::Int, N::Int)
     # Cannot use `fill` because the PairCoupling arrays must be
