+++ src/Reshaping.jl
@@ -30,80 +30,72 @@ function reshape_supercell(sys::System, shape)
     return reshape_supercell_aux(sys, new_cryst, new_dims)
 end
 
-#=
-# Transfer interactions from `src` to reshaped `sys`.
-#
-# Frequently `src` will refer to `sys.origin`, associated with the original
-# crystal, which will make symmetry analysis available. In this case, the
-# process to set a new interaction is to first modify `sys.origin`, and then to
-# call this function.
-function transfer_interactions_from_origin!(sys::System)
-    @assert is_homogeneous(sys)
-    (; origin) = sys
-
-    new_ints = interactions_homog(sys)
-
-    for new_i in 1:natoms(sys.crystal)
-        i = map_atom_to_other_crystal(sys.crystal, new_i, origin.crystal)
-        src_int = origin.interactions_union[i]
-
-        # Copy onsite couplings
-        new_ints[new_i].onsite = src_int.onsite
-
-        # Copy pair couplings
-        new_pc = PairCoupling[]
-        for pc in src_int.pair
-            new_bond = map_bond_to_other_crystal(origin.crystal, pc.bond, sys.crystal, new_i)
-            push!(new_pc, PairCoupling(new_bond, pc.scalar, pc.bilin, pc.biquad, pc.general))
-        end
-        new_pc = sort!(new_pc, by=c->c.isculled)
-        new_ints[new_i].pair = new_pc
-    end
-end
-=#
 
 # Transfer interactions from `sys.origin` to reshaped `sys`.
 function transfer_params_from_origin!(sys::System)
     @assert is_homogeneous(sys)
     (; origin) = sys
 
-    # Map atom (for origin crystal) to vector of atoms (for new crystal)
-    old_to_new = Dict(i => Int[] for i in 1:natoms(origin.crystal))
-    for i in 1:natoms(sys.crystal)
-        i_old = map_atom_to_other_crystal(sys.crystal, i, origin.crystal)
-        push!(old_to_new[i_old], i)
+    # Map atom in origin crystal to vector of atoms in new crystal
+    origin_to_new = Dict(i => Int[] for i in 1:natoms(origin.crystal))
+    for new_i in 1:natoms(sys.crystal)
+        i = map_atom_to_other_crystal(sys.crystal, new_i, origin.crystal)
+        # Append `new_i` to the vector at `origin_to_new[i]`
+        push!(origin_to_new[i], new_i)
     end
 
     empty!(sys.params)
 
     for param in origin.params
-        onsites = empty(param.onsites)
-        for (i_old, oc) in param.onsites
-            for i in old_to_new[i_old]
-                push!(onsites, (i, oc))
+        new_onsites = empty(param.onsites)
+        for (i, oc) in param.onsites
+            for new_i in origin_to_new[i]
+                push!(new_onsites, (new_i, oc))
             end
         end
 
-        pairs = PairCoupling[]
+        new_pairs = PairCoupling[]
         for pc in param.pairs
             i_old = pc.bond.i
-            for i in old_to_new[i_old]
-                bond = map_bond_to_other_crystal(origin.crystal, pc.bond, sys.crystal, i)
-                push!(pairs, PairCoupling(bond, pc.scalar, pc.bilin, pc.biquad, pc.general))
+            for new_i in origin_to_new[i_old]
+                new_bond = map_bond_to_other_crystal(origin.crystal, pc.bond, sys.crystal, new_i)
+                push!(new_pairs, PairCoupling(new_bond, pc.scalar, pc.bilin, pc.biquad, pc.general))
             end
         end
 
-        push!(sys.params, ModelParam(param.label, param.scale; onsites, pairs))
[…]
