### 2025-08-21 5196354
Rename scale to val in ModelParam
