### 2025-08-23 950fd90
LSWT doesn't actually need to sort sites prior to reshaping
