### 2025-08-24 3925ff2
Idealization in all cases. Improve position display
