### 2025-08-26 ac8a552
Always use orbits in WyckoffExpr for high precision
