### 2025-09-02 SunnySuite/Sunny.jl#424 @jabirali
CONTRIBUTING.md
This comment is part of the [JOSS review process](https://github.com/openjournals/joss-reviews/issues/8138) for your Sunny.jl manuscript.

One of the points that we as referees should check for, is "Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support". I think it's quite clear how to report issues (Github issues) and seek support (there's a link to your Slack channel), but perhaps not how to *contribute* to the software. One common way to do that, is to simply add a short `CONTRIBUTING.md` file to your repository as documented [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors). Perhaps that might be useful?

This is not required for publication, since it is clear that one might open a Pull Request in the standard way in this repository to offer a contribution, but I think it might still be useful for potential contributors to know a bit about what you would require to accept such a pull request with regards to code style or autoformatting, required testing, required documentation, etc.
