### 2025-09-02 #8138 @jabirali
@phibeck @ddahlbom

I have lately finally had time to finish my report. I believe it's an impressive and well-documented software project that deserves to be published, and that it "checks all the boxes" that are required for publication in JOSS. The tutorial is quite good, and everything that I tested seems to work as described on my computer.

I have opened two minor issues (linked above), related to the test suite and contribution statements; these are not required to be resolved to continue to manuscript acceptance, but are potential improvements that @ddahlbom and the Sunny.jl team can consider to further improve their project. Apologies again for the delay before I could produce a proper report.
