### 2025-09-05 #8138 @phibeck
Thank you @jabirali for completing your checklist and for confirming that the package is ready for publication! 🎉 Thanks for linking the minor issues that, as I understand, are recommendations to the authors for the future, but do not seem big enough to hold up publication.

@ddahlbom I think we're close to wrapping this up based on the completed checklists, but let me check in with @Chronum94 to get an update regarding the status.

@Chronum94 I wanted to follow up on the status of your review. I noticed that all points on the checklist have been marked as complete, but from your last comments here and in the issue, it sounded like you were planning to revisit the review before finalizing it. Could you let me know how things stand? Thanks!
