+++ Project.toml
@@ -43,7 +43,7 @@ ElasticArrays = "1.2.12"
 FFTW = "1.4.5"
 GLMakie = "0.13"
 HCubature = "1.7.0"
-JLD2 = "0.5.4"
+JLD2 = "0.6.0"
 LinearAlgebra = "1.10"
 Makie = "0.24"
 OffsetArrays = "1.10.8"
