### 2025-09-10 9cdfbc3
`load_nxs` should always parse covectors from axis labels (#420)
Also bump JLD2 compat to v0.6.0.
