+++ test/test_nxs.jl
@@ -1,13 +1,6 @@
 @testitem "load_nxs" begin
-    import CodecZlib, IOCapture
-
     filename = joinpath(@__DIR__, "nxs", "1Dcut_sym_0p00T_K.nxs")
-
-    # Discard warnings
-    capt = IOCapture.capture() do
-        load_nxs(filename)
-    end
-    params, signal = capt.value
+    params, signal = load_nxs(filename)
 
     @test isapprox(params.binstart, [-0.05, 0.617, -0.25, -0.1]; atol=1e-7)
     @test isapprox(params.binend, [0.0, 0.667, 0.0, 0.6]; atol=1e-7)
