### 2025-09-10 8621c4b
Always idealize latvecs
