### 2025-09-14 11f80a6
Reduce use of symprec
