### 2025-09-15 73b4678
Improve standardize and conventionalize_setting
