+++ src/Symmetry/Crystal.jl
@@ -224,95 +224,44 @@ function sort_sites!(cryst::Crystal)
 end
 
 
-# Attempt to find a permutation matrix P (with sign flips) such that latvecs*P
-# has a more standard form.
-function permute_to_standardize_lattice_vectors(latvecs)
-    P = Mat3(I) # Iteratively build permutation matrix
-
-    # Clip small matrix elements to zero
-    latvecs = [abs(x) < 1e-12 ? 0 : x for x in latvecs]
-
-    # Conventionally, a1 should be aligned with x direction
-    a1, a2, a3 = eachcol(latvecs)
-    if iszero(a2[2]) && iszero(a2[3]) && norm(a2) ≈ norm(a1)
-        P = P * SA[0 1 0; 1 0 0; 0 0 1] # Permute (a1, a2)
-    elseif iszero(a3[2]) && iszero(a3[3]) && norm(a3) ≈ norm(a1)
-        P = P * SA[0 0 1; 0 1 0; 1 0 0] # Permute (a1, a3)
-    end
-
-    # Conventionally, a2 should be in xy plane
-    _, a2, a3 = eachcol(latvecs * P)
-    if iszero(a3[3]) && norm(a3) ≈ norm(a2)
-        P = P * SA[1 0 0; 0 0 1; 0 1 0] # Permute (a2, a3)
-    end
-
-    # Flip columns so that the diagonal elements are positive
-    signs = [a < 0 ? -1 : 1 for a in diag(latvecs*P)]
-    P = P * Diagonal(signs)
-
-    # To preserve volume-orientation, we may need to flip one lattice
-    # vector. Pick a2 arbitrarily.
-    @assert det(P) ≈ 1 || det(P) ≈ -1
-    if det(P) ≈ -1
-        P = P * Diagonal(SA[1,-1,1])
-    end
-
-    @assert det(P) ≈ 1
-    @assert P' ≈ inv(P)
-    return P
-end
-
-
 """
     standardize(cryst::Crystal)
 
 Return the symmetry-inferred, standardized crystal unit cell under ITA
 conventions.
 """
-function standardize(cryst::Crystal; idealize=true)
-    !isnothing(cryst.root) && error("Call this function on root crystal instead")
+function standardize(cryst::Crystal)
+    isnothing(cryst.root) || error("Call this function on root crystal instead")
 
-    (; symprec) = cryst
-    cell = Spglib.Cell(cryst.latvecs, cryst.positions, cryst.types)
-    (; lattice, positions, atoms) = spg_standardize_cell_scaled(cell, symprec; no_idealize=!idealize)
-    positions = Vec3.(positions)
-    lattice = Mat3(lattice)
+    # Lattice vectors of ITA standard cell
+    latvecs = lattice_vectors(lattice_params(cryst.latvecs / cryst.sg.setting.R)...)
 
-    # TODO: Reimplement. Use case is when loading mCIFs where the positions in
-    # Cartesian coordinates must be preserved.
-    if !idealize
-        # Even if we're not idealizing the site positions, it is still important
-        # to tune the lattice vectors so that the lattice system is exactly
-        # consistent with the spacegroup setting (e.g. 90° angles are expected
-        # for cubic and tetragonal spacegroups). This adjustment is needed to
-        # ensure that `latvecs * symop.R * inv(latvecs)` is exactly orthogonal
-        # for each symmetry operation of the spacegroup.
-        std_lattice = Mat3(spg_standardize_cell_scaled(cell, symprec; no_idealize=false).lattice)
-        R = closest_unitary(lattice / std_lattice)
-        isapprox(R*std_lattice, lattice; rtol=cryst.symprec) || error("Lattice vectors inconsistent at symprec=$symprec")
[…]
