### 2025-09-16 ba37bfd
Fix reshape_crystal implementation
