+++ src/Symmetry/Crystal.jl
@@ -193,6 +193,7 @@ function permute_sites!(cryst::Crystal, p)
     cryst.positions .= cryst.positions[p]
     cryst.classes .= cryst.classes[p]
     cryst.types .= cryst.types[p]
+    return cryst
 end
 
 # Sort the sites according to class and fractional coordinates. Any changes here
@@ -220,7 +221,7 @@ function sort_sites!(cryst::Crystal)
         @assert false "Symmetry-equivalent positions $(pos_to_string(ri)) and $(pos_to_string(rj))"
     end
     p = sort(eachindex(cryst.positions), lt=less_than)
-    permute_sites!(cryst, p)
+    return permute_sites!(cryst, p)
 end
 
 
@@ -411,8 +412,9 @@ function validate_crystal(cryst::Crystal)
 end
 
 
+# repeat_multiple(["a", "b"], [2, 3]) == ["a", "a", "b", "b", "b"]
 function repeat_multiple(vals, lens)
-    reduce(vcat, map(vals, lens) do val, len
+    return reduce(vcat, map(vals, lens) do val, len
         fill(val, len)
     end)
 end
@@ -483,9 +485,9 @@ end
 
 
 function check_shape_commensurate(cryst, shape)
-    prim_cell = @something primitive_cell(cryst) Mat3(I)
+    prim_cell = primitive_cell(cryst)
     shape_in_prim = prim_cell \ shape
-    if !all_integer(shape_in_prim; atol=cryst.symprec)
+    if !all_integer(shape_in_prim; atol=1e-12)
         if prim_cell ≈ I
             error("Elements of `shape` must be integer. Received $shape.")
         else
@@ -494,49 +496,60 @@ function check_shape_commensurate(cryst, shape)
     end
 end
 
-function reshape_crystal_aux(cryst::Crystal, new_latvecs)
-    # This matrix defines a mapping from fractional coordinates `x` in `cryst`
-    # to fractional coordinates `y` in the new unit cell.
-    B = new_latvecs \ cryst.latvecs
 
-    # Symmetry precision needs to be rescaled for the new unit cell. Use cube
-    # root of volume ratio.
-    volume_ratio = abs(det(B))
-    new_symprec = cryst.symprec * cbrt(volume_ratio)
+# Indices of atoms in the primitive cell
+function primitive_atoms(cryst::Crystal)
+    P = primitive_cell(cryst)
+    P ≈ I && return collect(1:natoms(cryst))
+    invP = inv(P)
 
-    # Our goal is to loop over `cryst` cells (n1, n2, n3) and make sure we
-    # completely cover `new_latvecs` cell, {0 ≤ y₁ ≤ 1, 0 ≤ y₂ ≤ 1, 0 ≤ y₃ ≤ 1}.
-    # Due to the linear relationship `y = Bx`, the required range of (n1, n2,
-    # n3) is associated with inv(B). In particular, the cells (n1, n2, n3) must
-    # cover all corners of the parallelpiped `inv(B) * [{0,1}, {0,1}, {0,1}]`.
-    #
-    # The bounds `sum.(eachrow(...))` on `nmax` below is intended as a
-    # conservative, heuristic guess. Any mistake here will likely be detected as
-    # as mismatch between the ratio of atoms and the ratio of cell volumes.
-    nmax = round.(Int, sum.(eachrow(abs.(inv(B))))) .+ 1
-
-    new_positions = Vec3[]
-    new_types     = String[]
-    new_classes   = Int[]
-
-    for i in 1:natoms(cryst)
-        for n1 in -nmax[1]:nmax[1], n2 in -nmax[2]:nmax[2], n3 in -nmax[3]:nmax[3]
[…]
