+++ Project.toml
@@ -12,6 +12,7 @@ FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 HCubature = "19dc6840-f33b-545b-b366-655c7e3ffd49"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+MatInt = "f23b31af-f0ea-4208-b2e0-bbfc29c446c9"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
@@ -46,6 +47,7 @@ HCubature = "1.7.0"
 JLD2 = "0.5.4"
 LinearAlgebra = "1.10"
 Makie = "0.24"
+MatInt = "0.1.2"
 OffsetArrays = "1.10.8"
 Optim = "1.13.0"
 PrecompileTools = "1.2.0"
