### 2025-12-01 #8138 @ddahlbom
Hi @phibeck. Thanks very much for pushing this forward.

We just released v0.8.0, which addresses the issues brought up by @Chronum94 in [Issue 403](https://github.com/SunnySuite/Sunny.jl/issues/403).

The DOI is the following: [https://doi.org/10.5281/zenodo.11363938](https://doi.org/10.5281/zenodo.11363938)
