### 2025-12-04 #8138 @ddahlbom
> -     Double check authors and affiliations (including ORCIDs)

Checked. Markdown of JOSS manuscript has been updated and Zenodo archive coordinated.

> -     Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Release of v0.8.0 addresses reviewer concerns: https://github.com/SunnySuite/Sunny.jl/releases/tag/v0.8.0

> -     Archive the release on Zenodo/figshare/etc and post the DOI here.

This release is archived here: https://zenodo.org/records/17772178

> -     Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

ORCIDs match. Archive author list includes two additional authors (Mose Giordano and Sagar Ramchandani), as discussed above.

> -     Make sure that the license listed for the archive is the same as the software license.

The Zenodo archive now lists MIT license, identical to the github repo.

I believe that completes the checks on our end. Please let us know if you need anything else.
