### 2025-12-12 #8138 @ddahlbom
Many thanks, @phibeck. I have corrected all the issues you mentioned (see below). I also updated two of the references, replacing arxiv citations with citations to the now published articles.

> Re: the starred footnote, JOSS acknowledges the statement. Please remove the statement for the final version of record.

I will strip the footnote after posting this message. I'm required by Oak Ridge National Laboratory to have the statement in the "accepted" version of the manuscript.

> list of affiliations: please add the country (written out fully) for each affiliation

Done.

> line 10: please correct "Univeriy"

Thank you. Done.

> line 30: missing "as" after "such"

Added.

> line 58: add @ in front of weber

Fixed.

> lines 63-67: please either repeat subject "it" after each semicolon or use commas instead

Semicolons removed.

> line 73: replace semicolon by comma

Done.

> lines 95f: put Figure references in parenthesis

Done.

> line 102: is the hyphen in spin-s correct or should the variable s declare the spin?

This is the notation we typically use, as in spin-1, spin-3/2. Here we just intend spin-s, where s is any half integer. However, it's not being used as an adjective here, so I suppose it would be proper to remove the hyphen. This has been done. (Sorry if I am not understanding your point.)

> line 141: please spell out acronym ORNL explicitly

Done

> lines 189, 25, 263: please use $S=\frac{1}{2}$

Done

> change the title of the archived version to match that of the paper

Done.
