### 2025-12-15 #8138 @phibeck
Hi @ddahlbom thank you for taking care of this! Could you please add the last two DOIs of the changed references - then we're ready to send this off! Thanks
