### 2025-12-15 #8138 @ddahlbom
@phibeck I think that takes care of the issue. Please let me know if I've overlooked anything else.

In case that's all that's necessary, please let me thank you for guiding this through! We really appreciate your efforts. Thank you also to @Chronum94 and @jabirali for volunteering to review. The feedback was very useful.
