### 2025-12-15 #8138 @phibeck
Sorry @ddahlbom I overlooked this: could you please fix the DOI of the `zhang:2024` reference, thanks.
