### 2025-12-19 #8138 @kyleniemeyer
@ddahlbom I think this is because you used an `authors` field rather than `author`, which is what BibTex is expecting. Can you change all those to using `author` instead?
